Bivariate angular dependence function for extrapolation based on rays

The scale parameter \(g(w)\) in the Ledford and Tawn approach is estimated empirically for \(x\) large as $$\frac{\Pr(X_P>xw, Y_P>x(1-w))}{\Pr(X_P>x, Y_P>x)}$$ where the sample (\(X_P, Y_P\)) are observations on a common unit Pareto scale. The coefficient \(\eta\) is estimated using maximum likelihood as the shape parameter of a generalized Pareto distribution on \(\min(X_P, Y_P)\).

Usage

angextrapo(dat, qu = 0.95, w = seq(0.05, 0.95, length = 20))

Arguments

dat

an \(n\) by \(2\) matrix of multivariate observations

qu

quantile level on uniform scale at which to threshold data. Default to 0.95

w

vector of unique angles between 0 and 1 at which to evaluate scale empirically.

Value

a list with elements

w: angles between zero and one

g: scale function at a given value of w

eta: Ledford and Tawn tail dependence coefficient

References

Ledford, A.W. and J. A. Tawn (1996), Statistics for near independence in multivariate extreme values. Biometrika, 83(1), 169–187.

Examples

angextrapo(rmev(n = 1000, model = 'log', d = 2, param = 0.5))
#> $w
#>  [1] 0.05000000 0.09736842 0.14473684 0.19210526 0.23947368 0.28684211
#>  [7] 0.33421053 0.38157895 0.42894737 0.47631579 0.52368421 0.57105263
#> [13] 0.61842105 0.66578947 0.71315789 0.76052632 0.80789474 0.85526316
#> [19] 0.90263158 0.95000000
#> 
#> $g
#>  [1] 38.962963 21.074074 14.925926 11.888889 10.148148  9.037037  8.296296
#>  [8]  7.814815  7.518519  7.407407  7.407407  7.518519  7.814815  8.296296
#> [15]  9.037037 10.148148 11.888889 14.925926 21.074074 38.962963
#> 
#> $eta
#>     shape 
#> 0.4960403 
#>