Joint maximum likelihood for the non-homogeneous Poisson Process

Calculates the MLEs of the parameters (\(\mu\), \(\sigma\), \(\xi\)), and joint asymptotic covariance matrix of these MLEs over a range of thresholds as supplied by the user.

Usage

.Joint_MLE_NHPP(x, u = NULL, k, q1, q2 = 1, par, M)

Arguments

x

vector of data

u

optional vector of thresholds. If not supplied, then k thresholds between quantiles (q1, q2) will be used

k

number of thresholds to consider if u not supplied

q1

lower quantile to consider for threshold

q2

upper quantile to consider for threshold. Default to 1

par

starting values for the optimization

M

number of superpositions or 'blocks' / 'years' the process corresponds to. It affects the estimation of \(mu\) and \(sigma\), but these can be changed post-hoc to correspond to any number)

Value

a list with components

mle matrix of MLEs above the supplied thresholds; columns are (\(\mu\), \(\sigma\), \(\xi\))

Cov.all joint asymptotic covariance matrix of all MLEs

Cov.mu joint asymptotic covariance matrix of MLEs for \(\mu\)

Cov.sig joint asymptotic covariance matrix of MLEs for \(\sigma\)

Cov.xi joint asymptotic covariance matrix of MLEs for \(\xi\)

Author

Jennifer L. Wadsworth