Extended generalised Pareto log likelihood

This function provides the log-likelihood and quantiles for the different extended generalized Pareto distributions. All families share the same tail index as the generalized Pareto, and reduce to the latter when kappa=1.

Usage

egp.ll(
  xdat,
  thresh,
  par,
  model = c("pt-beta", "pt-gamma", "pt-power", "gj-tnorm", "gj-beta", "exptilt",
    "logist")
)

egp.retlev(
  xdat,
  thresh,
  par,
  model = c("pt-beta", "pt-gamma", "pt-power", "gj-tnorm", "gj-beta", "exptilt",
    "logist"),
  p,
  confint = FALSE,
  plot = TRUE
)

Arguments

xdat

vector of observations, greater than the threshold

thresh

threshold value

par

parameter vector (\(\kappa\), \(\sigma\), \(\xi\)).

model

a string indicating which extended family to fit

p

extreme event probability; p must be greater than the rate of exceedance for the calculation to make sense. See Details.

plot

logical; if TRUE, a plot of the return levels