Fit an extended generalized Pareto distribution of Naveau et al.

Deprecated function name to fit an extended generalized Pareto family. The user should call fit.extgp instead.

Usage

egp2.fit(
  data,
  model = 1,
  method = c("mle", "pwm"),
  init,
  censoring = c(0, Inf),
  rounded = 0,
  CI = FALSE,
  R = 1000,
  ncpus = 1,
  plots = TRUE
)

Arguments

data

data vector.

model

integer ranging from 0 to 4 indicating the model to select (see extgp).

method

string; either 'mle' for maximum likelihood, or 'pwm' for probability weighted moments, or both.

init

vector of initial values, comprising of \(p\), \(\kappa\), \(\delta\), \(\sigma\), \(\xi\) (in that order) for the optimization. All parameters may not appear depending on model.

censoring

numeric vector of length 2 containing the lower and upper bound for censoring; censoring=c(0,Inf) is equivalent to no censoring.

rounded

numeric giving the instrumental precision (and rounding of the data), with default of 0.

R

integer; number of bootstrap replications.

ncpus

integer; number of CPUs for parallel calculations (default: 1).

plots

logical; whether to produce histogram and density plots.

See also