Extended generalised Pareto families of Naveau et al. (2016)

Density function, distribution function, quantile function and random generation for the extended generalized Pareto distribution (GPD) with scale and shape parameters.

Arguments

q

vector of quantiles

x

vector of observations

p

vector of probabilities

n

sample size

prob

mixture probability for model type 4

kappa

shape parameter for type 1, 3 and 4

delta

additional parameter for type 2, 3 and 4

sigma

scale parameter

xi

shape parameter

type

integer between 0 to 5 giving the model choice

step

function of step size for discretization with default 0, corresponding to continuous quantiles

log

logical; should the log-density be returned (default to FALSE)?

unifsamp

sample of uniform; if provided, the data will be used in place of new uniform random variates

censoring

numeric vector of length 2 containing the lower and upper bound for censoring

Details

The extended generalized Pareto families proposed in Naveau et al. (2016) retain the tail index of the distribution while being compliant with the theoretical behavior of extreme low rainfall. There are five proposals, the first one being equivalent to the GP distribution.

type 0 corresponds to uniform carrier, \(G(u)=u\).

type 1 corresponds to a three parameters family, with carrier \(G(u)=u^\kappa\).

type 2 corresponds to a three parameters family, with carrier \(G(u)=1-V_\delta((1-u)^\delta)\).

type 3 corresponds to a four parameters family, with carrier $$G(u)=1-V_\delta((1-u)^\delta))^{\kappa/2}$$.

type 4 corresponds to a five parameter model (a mixture of type 2, with \(G(u)=pu^\kappa + (1-p)*u^\delta\)

Usage

pextgp(q, prob=NA, kappa=NA, delta=NA, sigma=NA, xi=NA, type=1)

dextgp(x, prob=NA, kappa=NA, delta=NA, sigma=NA, xi=NA, type=1, log=FALSE)

qextgp(p, prob=NA, kappa=NA, delta=NA, sigma=NA, xi=NA, type=1)

rextgp(n, prob=NA, kappa=NA, delta=NA, sigma=NA, xi=NA, type=1, unifsamp=NULL, censoring=c(0,Inf))

References

Naveau, P., R. Huser, P. Ribereau, and A. Hannart (2016), Modeling jointly low, moderate, and heavy rainfall intensities without a threshold selection, Water Resour. Res., 52, 2753-2769, doi:10.1002/2015WR018552.

Author

Raphael Huser and Philippe Naveau