Generalized extreme value distribution

Likelihood, score function and information matrix, bias, approximate ancillary statistics and sample space derivative for the generalized extreme value distribution

Arguments

par

vector of loc, scale and shape

dat

sample vector

method

string indicating whether to use the expected ('exp') or the observed ('obs' - the default) information matrix.

V

vector calculated by gev.Vfun

n

sample size

p

vector of probabilities

Usage

gev.ll(par, dat)
gev.ll.optim(par, dat)
gev.score(par, dat)
gev.infomat(par, dat, method = c('obs','exp'))
gev.retlev(par, p)
gev.bias(par, n)
gev.Fscore(par, dat, method=c('obs','exp'))
gev.Vfun(par, dat)
gev.phi(par, dat, V)
gev.dphi(par, dat, V)

Functions

gev.ll: log likelihood

gev.ll.optim: negative log likelihood parametrized in terms of location, log(scale) and shape in order to perform unconstrained optimization

gev.score: score vector

gev.infomat: observed or expected information matrix

gev.retlev: return level, corresponding to the \((1-p)\)th quantile

gev.bias: Cox-Snell first order bias

gev.Fscore: Firth's modified score equation

gev.Vfun: vector implementing conditioning on approximate ancillary statistics for the TEM

gev.phi: canonical parameter in the local exponential family approximation

gev.dphi: derivative matrix of the canonical parameter in the local exponential family approximation

References

Firth, D. (1993). Bias reduction of maximum likelihood estimates, Biometrika, 80(1), 27–38.

Coles, S. (2001). An Introduction to Statistical Modeling of Extreme Values, Springer, 209 p.

Cox, D. R. and E. J. Snell (1968). A general definition of residuals, Journal of the Royal Statistical Society: Series B (Methodological), 30, 248–275.

Cordeiro, G. M. and R. Klein (1994). Bias correction in ARMA models, Statistics and Probability Letters, 19(3), 169–176.