Profile log-likelihood for the generalized extreme value distribution

This function calculates the profile likelihood along with two small-sample corrections based on Severini's (1999) empirical covariance and the Fraser and Reid tangent exponential model approximation.

Usage

gev.pll(
  psi,
  param = c("loc", "scale", "shape", "quant", "Nmean", "Nquant"),
  mod = "profile",
  dat,
  N = NULL,
  p = NULL,
  q = NULL,
  correction = TRUE,
  plot = TRUE,
  ...
)

Arguments

psi

parameter vector over which to profile (unidimensional)

param

string indicating the parameter to profile over

mod

string indicating the model, one of profile, tem or modif.See Details.

dat

sample vector

N

size of block over which to take maxima. Required only for param Nmean and Nquant.

p

tail probability. Required only for param quant.

q

probability level of quantile. Required only for param Nquant.

correction

logical indicating whether to use spline.corr to smooth the tem approximation.

plot

logical; should the profile likelihood be displayed? Default to TRUE

...

additional arguments such as output from call to Vfun if mode='tem'.

Value

a list with components

mle: maximum likelihood estimate

psi.max: maximum profile likelihood estimate

param: string indicating the parameter to profile over

std.error: standard error of psi.max

psi: vector of parameter \(psi\) given in psi

pll: values of the profile log likelihood at psi

maxpll: value of maximum profile log likelihood

In addition, if mod includes tem

normal:maximum likelihood estimate and standard error of the interest parameter \(psi\)

r:values of likelihood root corresponding to \(\psi\)

q:vector of likelihood modifications

rstar:modified likelihood root vector

rstar.old:uncorrected modified likelihood root vector

tem.psimax:maximum of the tangent exponential model likelihood

In addition, if mod includes modif

tem.mle: maximum of tangent exponential modified profile log likelihood

tem.profll: values of the modified profile log likelihood at psi

tem.maxpll: value of maximum modified profile log likelihood

empcov.mle: maximum of Severini's empirical covariance modified profile log likelihood

empcov.profll: values of the modified profile log likelihood at psi

empcov.maxpll: value of maximum modified profile log likelihood

Details

The two additional mod available are tem, the tangent exponential model (TEM) approximation and modif for the penalized profile likelihood based on \(p^*\) approximation proposed by Severini. For the latter, the penalization is based on the TEM or an empirical covariance adjustment term.

References

Fraser, D. A. S., Reid, N. and Wu, J. (1999), A simple general formula for tail probabilities for frequentist and Bayesian inference. Biometrika, 86(2), 249--264.

Severini, T. (2000) Likelihood Methods in Statistics. Oxford University Press. ISBN 9780198506508.

Brazzale, A. R., Davison, A. C. and Reid, N. (2007) Applied asymptotics: case studies in small-sample statistics. Cambridge University Press, Cambridge. ISBN 978-0-521-84703-2

Examples

if (FALSE) {
set.seed(123)
dat <- evd::rgev(n = 100, loc = 0, scale = 2, shape = 0.3)
gev.pll(psi = seq(0,0.5, length = 50), param = 'shape', dat = dat)
gev.pll(psi = seq(-1.5, 1.5, length = 50), param = 'loc', dat = dat)
gev.pll(psi = seq(10, 40, by = 0.1), param = 'quant', dat = dat, p = 0.01)
gev.pll(psi = seq(12, 100, by=1), param = 'Nmean', N = 100, dat = dat)
gev.pll(psi = seq(12, 90, by=1), param = 'Nquant', N = 100, dat = dat, q = 0.5)
}