Generalized Pareto distribution (expected shortfall parametrization)

Likelihood, score function and information matrix, approximate ancillary statistics and sample space derivative for the generalized Pareto distribution parametrized in terms of expected shortfall.

The parameter m corresponds to \(\zeta_u\)/(1-\(\alpha\)), where \(\zeta_u\) is the rate of exceedance over the threshold u and \(\alpha\) is the percentile of the expected shortfall. Note that the actual parametrization is in terms of excess expected shortfall, meaning expected shortfall minus threshold.

Arguments

par

vector of length 2 containing \(e_m\) and \(\xi\), respectively the expected shortfall at probability 1/(1-\(\alpha\)) and the shape parameter.

dat

sample vector

m

number of observations of interest for return levels. See Details

tol

numerical tolerance for the exponential model

method

string indicating whether to use the expected ('exp') or the observed ('obs' - the default) information matrix.

nobs

number of observations

V

vector calculated by gpde.Vfun

Details

The observed information matrix was calculated from the Hessian using symbolic calculus in Sage.

Usage

gpde.ll(par, dat, m, tol=1e-5)
gpde.ll.optim(par, dat, m, tol=1e-5)
gpde.score(par, dat, m)
gpde.infomat(par, dat, m, method = c('obs', 'exp'), nobs = length(dat))
gpde.Vfun(par, dat, m)
gpde.phi(par, dat, V, m)
gpde.dphi(par, dat, V, m)

Functions

gpde.ll: log likelihood

gpde.ll.optim: negative log likelihood parametrized in terms of log expected shortfall and shape in order to perform unconstrained optimization

gpde.score: score vector

gpde.infomat: observed information matrix for GPD parametrized in terms of rate of expected shortfall and shape

gpde.Vfun: vector implementing conditioning on approximate ancillary statistics for the TEM

gpde.phi: canonical parameter in the local exponential family approximation

gpde.dphi: derivative matrix of the canonical parameter in the local exponential family approximation

Author

Leo Belzile