Generalized Pareto distribution

Density function, distribution function, quantile function and random number generation for the generalized Pareto distribution.

Usage

pgp(q, loc = 0, scale = 1, shape = 0, lower.tail = TRUE, log.p = FALSE)

dgp(x, loc = 0, scale = 1, shape = 0, log = FALSE)

qgp(p, loc = 0, scale = 1, shape = 0, lower.tail = TRUE)

rgp(n, loc = 0, scale = 1, shape = 0)

Arguments

loc

location parameter.

scale

scale parameter, strictly positive.

shape

shape parameter.

lower.tail

logical; if TRUE (default), the lower tail probability \(\Pr(X \leq x)\) is returned.

log.p, log

logical; if FALSE (default), values are returned on the probability scale.

x, q

vector of quantiles

p

vector of probabilities

n

scalar number of observations

References

Coles, S. G. (2001) An Introduction to Statistical Modeling of Extreme Values, Springer-Verlag, London. doi:10.1007/978-1-4471-3675-0_3