Interpret bivariate threshold exceedance models

This is an adaptation of the evir package interpret.gpdbiv function. interpret.fbvpot was adapted to deal with the output of a call to fbvpot from the evd and to handle families other than the logistic distribution. The likelihood derivation comes from expression 2.10 in Smith et al. (1997).

Usage

ibvpot(fitted, q, silent = FALSE)

Arguments

fitted

the output of fbvpot or a list. See Details.

q

a vector of quantiles to consider, on the data scale. Must be greater than the thresholds.

silent

boolean; whether to print the interpretation of the result. Default to FALSE.

Value

an invisible numeric vector containing marginal, joint and conditional exceedance probabilities.

Details

The list fitted must contain

model a string; see bvevd for options

param a named vector containing the parameters of the model, as well as parameters scale1, shape1,scale2 and shape2, corresponding to marginal GPD parameters.

threshold a vector of length 2 containing the two thresholds.

pat the proportion of observations above the corresponding threshold

References

Smith, Tawn and Coles (1997), Markov chain models for threshold exceedances. Biometrika, 84(2), 249--268.

See also

Author

Leo Belzile, adapting original S code by Alexander McNeil

Examples

y <- evd::rgpd(1000,1,1,1)
x <- y*rmevspec(n=1000,d=2,sigma=cbind(c(0,0.5),c(0.5,0)),model='hr')
mod <- evd::fbvpot(x,threshold = c(1,1),model = 'hr',likelihood ='censored')
ibvpot(mod, c(20,20))
#> Bivariate POT model: hr 
#> Thresholds: 1 1 
#> Extreme levels of interest (x,y): 20 20 
#> P(X > x) = 0.01737574 
#> P(Y > y) = 0.02125021 
#> P(X > x, Y > y) = 0.005673096 
#> P(X > x) * P(Y > y) = 0.0003692381 
#> P(Y > y | X > x) = 0.3264953 
#> P(X > x | Y > y) = 0.2669666