Estimation of the bivariate angular dependence function of Wadsworth and Tawn (2013)

Estimation of the bivariate angular dependence function of Wadsworth and Tawn (2013)

Usage

lambdadep(
  dat,
  qu = 0.95,
  method = c("hill", "mle", "bayes"),
  plot = TRUE,
  level = 0.95,
  ...
)

Arguments

dat

an \(n\) by \(2\) matrix of multivariate observations

qu

quantile level on uniform scale at which to threshold data. Default to 0.95

method

string indicating the estimation method

plot

logical indicating whether to return the graph of lambda

level

level for confidence intervals, default to 0.95

...

additional arguments, used for backward compatibility

The confidence intervals are based on normal quantiles. The standard errors for the hill are based on the asymptotic covariance and that of the mle derived using the delta-method. Bayesian posterior predictive interval estimates are obtained using ratio-of-uniform sampling with flat priors: the shape parameters are constrained to lie within the triangle, as are frequentist point estimates which are adjusted post-inference.

Value

a plot of the lambda function if plot=TRUE, plus an invisible list with components

w the sequence of angles in (0,1) at which the lambda values are evaluated

lambda point estimates of lambda

lower.confint level% confidence interval for lambda (lower bound)

upper.confint level% confidence interval for lambda (upper bound)

Examples

set.seed(12)
dat <- mev::rmev(n = 1000, d = 2, model = "log", param = 0.1)
lambdadep(dat, method = 'hill')
#> Error in lambdadep(dat, method = "hill"): object 'ci_level' not found
if (FALSE) { # \dontrun{
lambdadep(dat, method = 'bayes')
lambdadep(dat, method = 'mle')
# With independent observations
dat <- matrix(runif(n = 2000), ncol = 2)
lambdadep(dat, method = 'hill')
} # }