Likelihood for multivariate peaks over threshold models

Likelihood for the various parametric limiting models over region determined by $$\{y \in F: \max_{j=1}^D \sigma_j \frac{y^\xi_j-1}{\xi_j}+\mu_j > u\};$$ where \(\mu\) is loc, \(\sigma\) is scale and \(\xi\) is shape.

Usage

likmgp(
  dat,
  thresh,
  loc,
  scale,
  shape,
  par,
  model = c("log", "br", "xstud"),
  likt = c("mgp", "pois", "binom"),
  lambdau = 1,
  ...
)

Arguments

dat

matrix of observations

thresh

functional threshold for the maximum

loc

vector of location parameter for the marginal generalized Pareto distribution

scale

vector of scale parameter for the marginal generalized Pareto distribution

shape

vector of shape parameter for the marginal generalized Pareto distribution

par

list of parameters: alpha for the logistic model, Lambda for the Brown–Resnick model or else Sigma and df for the extremal Student.

model

string indicating the model family, one of "log", "neglog", "br" or "xstud"

likt

string indicating the type of likelihood, with an additional contribution for the non-exceeding components: one of "mgp", "binom" and "pois".

lambdau

vector of marginal rate of marginal threshold exceedance.

...

additional arguments (see Details)

Value

the value of the log-likelihood with attributes expme, giving the exponent measure

Details

Optional arguments can be passed to the function via ...

cl cluster instance created by makeCluster (default to NULL)

ncors number of cores for parallel computing of the likelihood

mmax maximum per column

B1 number of replicates for quasi Monte Carlo integral for the exponent measure

genvec1 generating vector for the quasi Monte Carlo routine (exponent measure), associated with B1

Note

The location and scale parameters are not identifiable unless one of them is fixed.