Multivariate Normal distribution sampler

Sampler derived using the eigendecomposition of the covariance matrix Sigma. The function uses the Armadillo random normal generator

Usage

mvrnorm(n, mu, Sigma)

Arguments

n

sample size

mu

mean vector. Will set the dimension

Sigma

a square covariance matrix, of same dimension as mu. No sanity check is performed to validate that the matrix is p.s.d., so use at own risk

Value

an n sample from a multivariate Normal distribution

Examples

mvrnorm(n=10, mu=c(0,2), Sigma=diag(2))
#>             [,1]      [,2]
#>  [1,] -0.8868628 0.2654812
#>  [2,]  0.6565412 1.8267407
#>  [3,]  0.4882522 2.9154757
#>  [4,] -1.3397256 3.6705300
#>  [5,] -0.4402110 3.7359878
#>  [6,] -1.0323680 2.2615054
#>  [7,]  0.8875935 0.9728781
#>  [8,] -1.7057297 2.1672866
#>  [9,] -0.9111461 1.4518746
#> [10,]  1.5760637 1.3225222