Nutrient data

Interview component of survey 'What we eat in America'. These are extracted from the 2015–2016 National Health and Nutrition Examination Survey (NHANES, https://wwwn.cdc.gov/nchs/nhanes/Default.aspx) report and consist of the total nutrients for all food and beverage intake ingested over a 24 hours period.

Usage

nutrients

Format

A data frame with 9544 rows and 38 variables:

prot

proteins (in grams)

carb

carbonhydrate (in gram)

sugr

total sugars (in gram)

fibe

dietary fibers (in grams)

tfat

total fat (in grams)

sfat

saturated fat (in grams)

mfat

monounsaturated fat (in grams)

pfat

polyunsaturated fat (in grams)

chol

cholesterol (in milligrams)

atoc

vitamin E as alpha-tocopherol (in milligrams)

ret

retinol (in micrograms)

vara

Vitamin A as retinol activity equivalents (in micrograms).

acar

alpha-carotene (in micrograms)

bcar

beta-carotene (in micrograms)

cryp

beta-cryptoxanthin (in micrograms)

lyco

lycopene (in micrograms)

lz

lutein and zeaxanthin (in micrograms).

vb1

thiamin (vitamin B1, in milligrams)

vb2

riboflavin (vitamin B2, in milligrams)

niac

niacin (in milligrams)

vb6

vitamin B5 (in milligrams)

fola

total folate (in micrograms)

fa

folic acid (in micrograms)

ff

food folate (in micrograms)

chl

total choline (in milligrams)

vb12

vitamin B12 (in micrograms)

vc

vitamin C (in milligrams)

vd

vitamin D (comprising D2 and D3, in micrograms)

vk

vitamin K (in micrograms)

calc

calcium (in milligrams)

phos

phosphorus (in milligrams)

magn

magnesium (in milligrams)

iron

iron (in milligrams)

zinc

zinc (in milligrams)

copp

copper (in milligrams)

sodi

sodium (in milligrams)

pota

potassium (in milligrams)

sele

selenium (in micrograms)

Source

National Center for Health Statistics, now available from the Wayback Machine via https://web.archive.org/web/20201029113801/https://wwwn.cdc.gov/Nchs/Nhanes/2015-2016/DR1TOT_I.XPT

Details

Note that the sample design oversampled specific population targets and that only respondants are provided. The website contains more information about sampling weights. There are multiple missing records.

Note

These data are subject to a data user agreement, available at https://www.cdc.gov/nchs/policy/data-user-agreement.html