Random variate generation for Dirichlet distribution on \(S_{d}\)

A function to sample Dirichlet random variables, based on the representation as ratios of Gamma. Note that the RNG will generate on the full simplex and the sum to one constraint is respected here

Usage

rdir(n, alpha, normalize = TRUE)

Arguments

n

sample size

alpha

vector of parameter

normalize

boolean. If FALSE, the function returns Gamma variates with parameter alpha.

Value

sample of dimension d (size of alpha) from the Dirichlet distribution.

Examples

rdir(n=100, alpha=c(0.5,0.5,2),TRUE)
#>                [,1]         [,2]      [,3]
#>   [1,] 7.369649e-02 0.0765378033 0.8497657
#>   [2,] 2.311068e-01 0.4353898871 0.3335033
#>   [3,] 1.552036e-01 0.0783630225 0.7664334
#>   [4,] 2.182047e-01 0.0426221842 0.7391732
#>   [5,] 7.140462e-01 0.1021241124 0.1838297
#>   [6,] 4.954060e-02 0.0022511560 0.9482082
#>   [7,] 3.418142e-02 0.2011327504 0.7646858
#>   [8,] 1.313665e-02 0.0843731177 0.9024902
#>   [9,] 3.055985e-01 0.0349449043 0.6594566
#>  [10,] 2.318918e-01 0.0925978215 0.6755104
#>  [11,] 7.504584e-03 0.0771184493 0.9153770
#>  [12,] 7.147811e-03 0.0251158934 0.9677363
#>  [13,] 6.339870e-01 0.0030634436 0.3629496
#>  [14,] 8.855816e-02 0.2048948574 0.7065470
#>  [15,] 6.796217e-03 0.0010661235 0.9921377
#>  [16,] 1.984315e-03 0.2801299603 0.7178857
#>  [17,] 2.278089e-03 0.1308096919 0.8669122
#>  [18,] 2.506997e-02 0.1946339132 0.7802961
#>  [19,] 2.383513e-02 0.1333190629 0.8428458
#>  [20,] 1.111688e-01 0.0733469955 0.8154842
#>  [21,] 6.198885e-01 0.0720251022 0.3080864
#>  [22,] 4.910645e-01 0.0690836947 0.4398518
#>  [23,] 1.698968e-01 0.2915551808 0.5385480
#>  [24,] 3.177091e-01 0.0315200359 0.6507709
#>  [25,] 1.069836e-01 0.1473328591 0.7456835
#>  [26,] 3.064791e-02 0.0204866398 0.9488654
#>  [27,] 9.131042e-02 0.0844671051 0.8242225
#>  [28,] 9.901256e-03 0.7091638378 0.2809349
#>  [29,] 1.101397e-01 0.1973826652 0.6924776
#>  [30,] 1.856709e-01 0.0329121903 0.7814169
#>  [31,] 6.687670e-02 0.0998275613 0.8332957
#>  [32,] 3.440934e-02 0.0338974309 0.9316932
#>  [33,] 7.663968e-05 0.2268024272 0.7731209
#>  [34,] 7.127407e-02 0.4046925226 0.5240334
#>  [35,] 5.389910e-02 0.0519251114 0.8941758
#>  [36,] 2.590589e-02 0.0568036681 0.9172904
#>  [37,] 9.498001e-02 0.0181748562 0.8868451
#>  [38,] 7.617966e-01 0.0041816480 0.2340217
#>  [39,] 1.076328e-01 0.5866685442 0.3056987
#>  [40,] 8.562890e-02 0.1395148617 0.7748562
#>  [41,] 1.907166e-01 0.0568336421 0.7524498
#>  [42,] 2.616550e-01 0.0124969250 0.7258481
#>  [43,] 4.313003e-01 0.0253662961 0.5433334
#>  [44,] 9.399444e-03 0.1568176826 0.8337829
#>  [45,] 3.677795e-01 0.2164950949 0.4157254
#>  [46,] 2.475853e-01 0.0041190776 0.7482957
#>  [47,] 9.289539e-02 0.0750808246 0.8320238
#>  [48,] 2.572678e-01 0.0181135487 0.7246186
#>  [49,] 3.862775e-01 0.0862304637 0.5274921
#>  [50,] 3.478451e-04 0.2530317691 0.7466204
#>  [51,] 2.841905e-02 0.1649836882 0.8065973
#>  [52,] 6.876566e-03 0.0528202418 0.9403032
#>  [53,] 1.472265e-01 0.1134206562 0.7393529
#>  [54,] 6.609456e-02 0.4709766905 0.4629288
#>  [55,] 5.336126e-01 0.0144474098 0.4519400
#>  [56,] 1.666306e-01 0.0381576436 0.7952117
#>  [57,] 8.154915e-02 0.7013425741 0.2171083
#>  [58,] 9.355152e-02 0.0033538285 0.9030947
#>  [59,] 8.135511e-03 0.4556697392 0.5361947
#>  [60,] 2.606364e-01 0.0055309842 0.7338327
#>  [61,] 8.950617e-02 0.0891867471 0.8213071
#>  [62,] 2.695722e-03 0.8073763926 0.1899279
#>  [63,] 1.643465e-02 0.3419713225 0.6415940
#>  [64,] 4.221305e-02 0.1137491508 0.8440378
#>  [65,] 1.156145e-04 0.7251124422 0.2747719
#>  [66,] 1.991125e-02 0.1082286717 0.8718601
#>  [67,] 2.297183e-01 0.2105580712 0.5597236
#>  [68,] 2.014650e-03 0.0899986970 0.9079867
#>  [69,] 4.896003e-02 0.0194882319 0.9315517
#>  [70,] 2.591126e-04 0.3471344026 0.6526065
#>  [71,] 5.704946e-03 0.4023671740 0.5919279
#>  [72,] 2.592707e-01 0.1259765895 0.6147527
#>  [73,] 6.163806e-02 0.0259790864 0.9123829
#>  [74,] 1.325899e-01 0.0015947716 0.8658154
#>  [75,] 1.939124e-01 0.2708384203 0.5352491
#>  [76,] 1.173954e-02 0.0063440203 0.9819164
#>  [77,] 1.056518e-02 0.4140896119 0.5753452
#>  [78,] 2.920152e-01 0.0113226302 0.6966622
#>  [79,] 2.673810e-01 0.0185421528 0.7140769
#>  [80,] 9.046118e-02 0.0002462465 0.9092926
#>  [81,] 5.296684e-02 0.1210221461 0.8260110
#>  [82,] 9.411097e-06 0.5387013687 0.4612892
#>  [83,] 3.022372e-04 0.3152238190 0.6844739
#>  [84,] 4.895891e-04 0.0003463612 0.9991640
#>  [85,] 9.633849e-03 0.3949001041 0.5954660
#>  [86,] 1.360569e-01 0.0014228178 0.8625202
#>  [87,] 1.411284e-02 0.1243983192 0.8614888
#>  [88,] 2.822394e-02 0.0308136310 0.9409624
#>  [89,] 1.537854e-01 0.0610251617 0.7851895
#>  [90,] 1.363112e-01 0.1926706587 0.6710182
#>  [91,] 6.910669e-01 0.0131130571 0.2958201
#>  [92,] 1.200728e-01 0.0201652134 0.8597620
#>  [93,] 2.094298e-01 0.5009609511 0.2896093
#>  [94,] 4.403166e-02 0.0683066188 0.8876617
#>  [95,] 8.838625e-02 0.0735689354 0.8380448
#>  [96,] 4.751124e-01 0.1803503826 0.3445372
#>  [97,] 3.903883e-01 0.0027770676 0.6068346
#>  [98,] 2.490775e-01 0.0519544858 0.6989680
#>  [99,] 9.759119e-02 0.1425127388 0.7598961
#> [100,] 1.011836e-01 0.0941766302 0.8046398
rdir(n=100, alpha=c(3,1,2),FALSE)
#>             [,1]         [,2]      [,3]
#>   [1,] 0.5191856 0.5657185102 1.3476165
#>   [2,] 3.0572454 0.7587096401 0.4957867
#>   [3,] 0.6712289 2.8103192174 1.3739499
#>   [4,] 1.7766274 0.5149119876 5.7753507
#>   [5,] 5.1980112 0.7468086972 0.8712667
#>   [6,] 0.7832317 0.3207621807 1.8251169
#>   [7,] 1.8164491 1.5091178701 1.4199056
#>   [8,] 3.8437073 0.6332100661 0.1707316
#>   [9,] 3.0454307 3.1648358410 3.1156613
#>  [10,] 5.8064077 0.1407429512 3.6765402
#>  [11,] 1.8641454 0.8455789518 7.4137553
#>  [12,] 1.1736300 1.0091973208 0.7443287
#>  [13,] 2.9713201 1.6507309538 1.7703679
#>  [14,] 4.1716465 1.0172656245 1.8056013
#>  [15,] 2.9324668 0.1310734689 1.4503016
#>  [16,] 5.2781363 1.8206847439 3.9430276
#>  [17,] 2.2138287 0.4456008300 4.5642070
#>  [18,] 3.4872290 1.0653057445 1.3334971
#>  [19,] 0.5402611 1.3250083198 0.7437200
#>  [20,] 4.9117105 0.7281726592 2.2953126
#>  [21,] 5.2280549 1.9847455145 4.3790928
#>  [22,] 5.2264994 0.5332433678 0.7037596
#>  [23,] 1.3405926 0.4362851616 5.1107455
#>  [24,] 3.4540471 1.3620054950 1.3822771
#>  [25,] 1.5716870 3.1221060284 3.6317457
#>  [26,] 2.6902309 0.3619814708 4.3218787
#>  [27,] 6.0405311 1.1078187527 1.7560398
#>  [28,] 5.9088640 0.7410845114 0.3513223
#>  [29,] 1.7903857 0.4182348277 0.1108663
#>  [30,] 4.0502163 1.1828106707 1.4690982
#>  [31,] 3.9763782 1.7447769643 4.7056762
#>  [32,] 1.5329652 4.0111959302 1.0538613
#>  [33,] 1.9134197 1.6225686887 1.4834122
#>  [34,] 3.8662030 0.2951403905 4.5781610
#>  [35,] 1.4242328 3.0645625208 0.6588286
#>  [36,] 2.3109077 0.5853459076 0.8717445
#>  [37,] 3.8799819 0.1762865452 0.4648718
#>  [38,] 1.0383487 0.6933322538 0.6367161
#>  [39,] 1.7879741 1.0049423177 4.4566953
#>  [40,] 4.0967402 0.6254848029 5.0710935
#>  [41,] 4.3214311 0.6297074165 1.6134154
#>  [42,] 2.0034124 1.2979098231 3.2159645
#>  [43,] 1.6896058 0.7033967821 1.5912472
#>  [44,] 1.1215118 1.0039404008 1.2300292
#>  [45,] 2.5165104 0.5122206085 1.3831104
#>  [46,] 3.8222020 0.2321090862 1.2181405
#>  [47,] 6.8946794 1.4351358269 1.4746009
#>  [48,] 1.0387056 0.5334038538 4.0083240
#>  [49,] 4.5091487 0.1375620674 2.0511536
#>  [50,] 1.3306141 0.5830652704 3.5368835
#>  [51,] 1.7000354 0.1623763747 0.9035040
#>  [52,] 4.5675359 1.6294349226 0.5030574
#>  [53,] 3.7579045 4.3537183223 0.5061478
#>  [54,] 4.5876322 0.5801321390 0.3782676
#>  [55,] 7.0260941 0.1275199147 1.7001996
#>  [56,] 4.4090710 0.4754320172 6.3471173
#>  [57,] 2.1176563 1.7518338924 0.9398455
#>  [58,] 1.5590015 0.4105680200 4.1161434
#>  [59,] 1.9336569 1.3302181298 5.0938611
#>  [60,] 0.9604132 0.3157753431 1.0847309
#>  [61,] 3.2304251 1.8045651408 2.4803022
#>  [62,] 1.1478336 1.5282867316 1.4989679
#>  [63,] 3.8511378 1.1209671502 1.3281819
#>  [64,] 3.1761339 3.1254713993 1.8640920
#>  [65,] 3.5362214 1.1038261523 2.0277034
#>  [66,] 0.8296751 0.0716374256 1.2411745
#>  [67,] 1.6655442 7.8322391665 0.7619994
#>  [68,] 4.4486161 2.6607105374 2.4260625
#>  [69,] 2.7995609 3.8071390025 1.2952952
#>  [70,] 6.9166534 0.0219043313 1.6299354
#>  [71,] 2.9564156 3.0130172097 3.4486462
#>  [72,] 2.3180321 0.4031472426 6.5958056
#>  [73,] 3.0984318 0.4431342734 1.4009734
#>  [74,] 4.1619789 1.0643301617 3.3788266
#>  [75,] 2.2406345 2.3717957308 0.5126196
#>  [76,] 2.8405126 1.3682749829 1.0031464
#>  [77,] 3.0856193 0.8970408785 3.0025300
#>  [78,] 3.2910129 0.8657499801 0.5113335
#>  [79,] 4.0228616 1.4653744747 1.9375124
#>  [80,] 0.4335552 2.0972002653 1.9818001
#>  [81,] 0.4600798 0.9791579351 1.3472679
#>  [82,] 1.7154897 1.4873921369 1.1390135
#>  [83,] 1.4581621 0.4943336982 1.6054939
#>  [84,] 7.2077960 0.6215677808 1.7869377
#>  [85,] 2.3611707 2.5012458704 0.4119828
#>  [86,] 1.4592564 1.5373245435 0.3379988
#>  [87,] 0.9967300 0.3453843085 1.5476422
#>  [88,] 0.7037616 3.2810737506 1.0323901
#>  [89,] 7.5508363 1.2662776511 1.9617336
#>  [90,] 4.4722093 0.3799931302 4.5241559
#>  [91,] 1.5447284 0.0721678259 1.3333920
#>  [92,] 0.8987009 0.8667604376 1.5147813
#>  [93,] 1.5747097 0.2764275411 1.4067947
#>  [94,] 1.8487052 0.3943475888 1.3526483
#>  [95,] 2.3353823 3.3625770511 0.8735704
#>  [96,] 4.9237932 0.7460144623 1.0115178
#>  [97,] 6.2148871 1.9870104106 0.5484934
#>  [98,] 1.6386062 0.9349983750 0.5562714
#>  [99,] 1.6367624 0.4635335813 1.6701824
#> [100,] 2.4702622 0.0006304142 1.4764322