Random samples from spectral distributions of multivariate extreme value models.

Generate from \(Q_i\), the spectral measure of a given multivariate extreme value model based on the L1 norm.

Usage

rmevspec(
  n,
  d,
  param,
  sigma,
  model = c("log", "neglog", "bilog", "negbilog", "hr", "br", "xstud", "smith",
    "schlather", "ct", "sdir", "dirmix", "pairbeta", "pairexp", "wdirbs", "wexpbs",
    "maxlin"),
  weights = NULL,
  vario = NULL,
  coord = NULL,
  grid = FALSE,
  dist = NULL,
  ...
)

Arguments

n

number of observations

d

dimension of sample

param

parameter vector for the logistic, bilogistic, negative bilogistic and extremal Dirichlet (Coles and Tawn) model. Parameter matrix for the Dirichlet mixture. Degree of freedoms for extremal student model. See Details.

sigma

covariance matrix for Brown-Resnick and extremal Student-t distributions. Symmetric matrix of squared coefficients \(\lambda^2\) for the Husler-Reiss model, with zero diagonal elements.

model

for multivariate extreme value distributions, users can choose between 1-parameter logistic and negative logistic, asymmetric logistic and negative logistic, bilogistic, Husler-Reiss, extremal Dirichlet model (Coles and Tawn) or the Dirichlet mixture. Spatial models include the Brown-Resnick, Smith, Schlather and extremal Student max-stable processes. Max linear models are also supported

weights

vector of length m for the m mixture components that sum to one. For the "maxlin" model, weights should be a matrix with d columns that represent the weight of the components and whose column sum to one (if provided, this argument overrides asy).

vario

semivariogram function whose first argument must be distance. Used only if provided in conjunction with coord and if sigma is missing

coord

d by k matrix of coordinates, used as input in the variogram vario or as parameter for the Smith model. If grid is TRUE, unique entries should be supplied.

grid

Logical. TRUE if the coordinates are two-dimensional grid points (spatial models).

dist

symmetric matrix of pairwise distances. Default to NULL.

...

additional arguments for the vario function

Value

an n by d exact sample from the corresponding multivariate extreme value model

Details

The vector param differs depending on the model

log: one dimensional parameter greater than 1

neglog: one dimensional positive parameter

bilog: d-dimensional vector of parameters in \([0,1]\)

negbilog: d-dimensional vector of negative parameters

ct, dir, negdir: d-dimensional vector of positive (a)symmetry parameters. Alternatively, a \(d+1\) vector consisting of the d Dirichlet parameters and the last entry is an index of regular variation in \((0, 1]\) treated as scale

xstud: one dimensional parameter corresponding to degrees of freedom alpha

dirmix: d by m-dimensional matrix of positive (a)symmetry parameters

pairbeta, pairexp: d(d-1)/2+1 vector of parameters, containing the concentration parameter and the coefficients of the pairwise beta, in lexicographical order e.g., \(\beta_{1,2}, \beta_{1,3}, \ldots\)

wdirbs, wexpbs: 2d vector of d concentration parameters followed by the d Dirichlet parameters

Note

This functionality can be useful to generate for example Pareto processes with marginal exceedances.

References

Dombry, Engelke and Oesting (2016). Exact simulation of max-stable processes, Biometrika, 103(2), 303–317.

Boldi (2009). A note on the representation of parametric models for multivariate extremes. Extremes 12, 211–218.

Author

Leo Belzile

Examples

set.seed(1)
rmevspec(n=100, d=3, param=2.5, model='log')
#>              [,1]       [,2]       [,3]
#>   [1,] 0.25523061 0.53956467 0.20520472
#>   [2,] 0.49950727 0.24182974 0.25866299
#>   [3,] 0.09690181 0.84429025 0.05880794
#>   [4,] 0.14968357 0.20081837 0.64949806
#>   [5,] 0.31644664 0.33776951 0.34578384
#>   [6,] 0.40585410 0.24744829 0.34669761
#>   [7,] 0.34190440 0.42640123 0.23169438
#>   [8,] 0.30139404 0.37250351 0.32610245
#>   [9,] 0.49383789 0.28298163 0.22318048
#>  [10,] 0.34311249 0.35963094 0.29725656
#>  [11,] 0.31731829 0.38568985 0.29699186
#>  [12,] 0.41654853 0.45631933 0.12713214
#>  [13,] 0.31718297 0.52808696 0.15473008
#>  [14,] 0.54776772 0.12712686 0.32510542
#>  [15,] 0.46090048 0.22475579 0.31434372
#>  [16,] 0.35805293 0.25018043 0.39176664
#>  [17,] 0.35613354 0.18080080 0.46306566
#>  [18,] 0.18238254 0.51989655 0.29772091
#>  [19,] 0.11837443 0.56335068 0.31827490
#>  [20,] 0.38439249 0.32848402 0.28712349
#>  [21,] 0.42625231 0.30540881 0.26833888
#>  [22,] 0.22199672 0.54359840 0.23440488
#>  [23,] 0.35487225 0.39470031 0.25042744
#>  [24,] 0.31688048 0.28913121 0.39398832
#>  [25,] 0.22325021 0.43791465 0.33883514
#>  [26,] 0.23476446 0.41455357 0.35068197
#>  [27,] 0.74496221 0.10283155 0.15220624
#>  [28,] 0.47982957 0.31272030 0.20745013
#>  [29,] 0.41166667 0.18877346 0.39955988
#>  [30,] 0.66666884 0.18768873 0.14564243
#>  [31,] 0.12366956 0.11220365 0.76412678
#>  [32,] 0.32197227 0.28813342 0.38989431
#>  [33,] 0.15962112 0.70588650 0.13449238
#>  [34,] 0.29869877 0.24533328 0.45596795
#>  [35,] 0.24024526 0.44230272 0.31745202
#>  [36,] 0.30665521 0.20767501 0.48566977
#>  [37,] 0.44398224 0.23863558 0.31738219
#>  [38,] 0.24663838 0.40233133 0.35103029
#>  [39,] 0.14727118 0.59193895 0.26078987
#>  [40,] 0.43487698 0.25690635 0.30821667
#>  [41,] 0.30079447 0.25862662 0.44057891
#>  [42,] 0.38958126 0.53196925 0.07844949
#>  [43,] 0.34865895 0.32371936 0.32762169
#>  [44,] 0.19671246 0.44371484 0.35957271
#>  [45,] 0.56992506 0.32221843 0.10785651
#>  [46,] 0.21847715 0.29121012 0.49031273
#>  [47,] 0.51656197 0.29594124 0.18749679
#>  [48,] 0.19568837 0.23702560 0.56728603
#>  [49,] 0.17223595 0.40102142 0.42674263
#>  [50,] 0.42278884 0.18851859 0.38869257
#>  [51,] 0.42398733 0.15527388 0.42073879
#>  [52,] 0.30155520 0.31716222 0.38128258
#>  [53,] 0.09895022 0.06726794 0.83378184
#>  [54,] 0.44898423 0.28522578 0.26578999
#>  [55,] 0.39170668 0.34687159 0.26142172
#>  [56,] 0.30607697 0.34678660 0.34713643
#>  [57,] 0.51628394 0.24591389 0.23780217
#>  [58,] 0.34496410 0.24772504 0.40731086
#>  [59,] 0.57322565 0.28482020 0.14195415
#>  [60,] 0.36871176 0.21152938 0.41975885
#>  [61,] 0.22628225 0.29972802 0.47398973
#>  [62,] 0.24251173 0.45970980 0.29777847
#>  [63,] 0.24876396 0.42245127 0.32878477
#>  [64,] 0.56463666 0.22430134 0.21106200
#>  [65,] 0.48047458 0.14728921 0.37223621
#>  [66,] 0.10638141 0.64800554 0.24561304
#>  [67,] 0.83747489 0.07691375 0.08561136
#>  [68,] 0.48220264 0.31078974 0.20700763
#>  [69,] 0.19974626 0.46422541 0.33602833
#>  [70,] 0.38055666 0.24986962 0.36957372
#>  [71,] 0.18743929 0.66535329 0.14720742
#>  [72,] 0.27367659 0.32656767 0.39975574
#>  [73,] 0.50398049 0.27241178 0.22360773
#>  [74,] 0.12338522 0.12671130 0.74990348
#>  [75,] 0.70914175 0.18156176 0.10929649
#>  [76,] 0.43260934 0.28681220 0.28057846
#>  [77,] 0.18353903 0.21512067 0.60134030
#>  [78,] 0.16809848 0.70330193 0.12859959
#>  [79,] 0.41701506 0.33345375 0.24953118
#>  [80,] 0.25020582 0.10897845 0.64081573
#>  [81,] 0.67845151 0.15387600 0.16767249
#>  [82,] 0.21689551 0.38560145 0.39750304
#>  [83,] 0.26816456 0.44299181 0.28884363
#>  [84,] 0.22761305 0.53424681 0.23814015
#>  [85,] 0.29582570 0.41851609 0.28565820
#>  [86,] 0.22475633 0.26652366 0.50872001
#>  [87,] 0.52009405 0.33180504 0.14810091
#>  [88,] 0.37922289 0.49366545 0.12711166
#>  [89,] 0.34930689 0.36252369 0.28816942
#>  [90,] 0.19079161 0.29952436 0.50968403
#>  [91,] 0.47653091 0.21370614 0.30976295
#>  [92,] 0.22922311 0.22059555 0.55018134
#>  [93,] 0.50271619 0.27451214 0.22277167
#>  [94,] 0.16524229 0.34637661 0.48838110
#>  [95,] 0.35898734 0.42981014 0.21120252
#>  [96,] 0.25364711 0.32514562 0.42120727
#>  [97,] 0.22926294 0.48058674 0.29015032
#>  [98,] 0.16556028 0.32971386 0.50472587
#>  [99,] 0.52464681 0.25236862 0.22298456
#> [100,] 0.17165090 0.38219625 0.44615285
rmevspec(n=100, d=3, param=2.5, model='neglog')
#>              [,1]       [,2]       [,3]
#>   [1,] 0.29324755 0.27371670 0.43303575
#>   [2,] 0.24266201 0.30189134 0.45544665
#>   [3,] 0.31491897 0.09349562 0.59158541
#>   [4,] 0.27106403 0.45647367 0.27246230
#>   [5,] 0.30114810 0.25663858 0.44221331
#>   [6,] 0.49332040 0.23112621 0.27555339
#>   [7,] 0.39640444 0.49076605 0.11282951
#>   [8,] 0.30549790 0.23227172 0.46223038
#>   [9,] 0.44474084 0.27205092 0.28320824
#>  [10,] 0.51738810 0.33769578 0.14491612
#>  [11,] 0.43672504 0.36623100 0.19704396
#>  [12,] 0.27525384 0.18181453 0.54293163
#>  [13,] 0.45892746 0.23603767 0.30503487
#>  [14,] 0.48476158 0.28697691 0.22826151
#>  [15,] 0.23435381 0.29250756 0.47313863
#>  [16,] 0.22361187 0.40423480 0.37215333
#>  [17,] 0.52183070 0.32149883 0.15667047
#>  [18,] 0.46947056 0.27772310 0.25280634
#>  [19,] 0.28068026 0.27629024 0.44302949
#>  [20,] 0.31469913 0.36943222 0.31586865
#>  [21,] 0.30805674 0.17890266 0.51304060
#>  [22,] 0.55781581 0.20928548 0.23289871
#>  [23,] 0.38233221 0.40150490 0.21616289
#>  [24,] 0.41240845 0.17701829 0.41057326
#>  [25,] 0.40947319 0.16696131 0.42356550
#>  [26,] 0.17878224 0.08348366 0.73773411
#>  [27,] 0.35439864 0.49684478 0.14875658
#>  [28,] 0.22309828 0.34663669 0.43026503
#>  [29,] 0.18760515 0.28025888 0.53213597
#>  [30,] 0.13633821 0.39188672 0.47177506
#>  [31,] 0.41812736 0.12801475 0.45385789
#>  [32,] 0.19467090 0.22831348 0.57701562
#>  [33,] 0.23529724 0.26011478 0.50458798
#>  [34,] 0.30747184 0.38316536 0.30936280
#>  [35,] 0.31844673 0.29558943 0.38596383
#>  [36,] 0.47195816 0.44898731 0.07905453
#>  [37,] 0.23654438 0.58608835 0.17736726
#>  [38,] 0.34617156 0.41380491 0.24002353
#>  [39,] 0.27549295 0.45452960 0.26997744
#>  [40,] 0.44315336 0.31232459 0.24452205
#>  [41,] 0.21926632 0.40420856 0.37652512
#>  [42,] 0.43340922 0.35066458 0.21592620
#>  [43,] 0.10612122 0.42053157 0.47334721
#>  [44,] 0.45643463 0.26662252 0.27694284
#>  [45,] 0.40455405 0.52270899 0.07273696
#>  [46,] 0.42284508 0.26074728 0.31640765
#>  [47,] 0.42349328 0.25871417 0.31779255
#>  [48,] 0.09953614 0.59841573 0.30204813
#>  [49,] 0.31080589 0.38601611 0.30317800
#>  [50,] 0.22193529 0.44474529 0.33331942
#>  [51,] 0.63648363 0.15918551 0.20433087
#>  [52,] 0.30111364 0.39681544 0.30207093
#>  [53,] 0.36523614 0.35988737 0.27487649
#>  [54,] 0.16016202 0.41581293 0.42402504
#>  [55,] 0.40146904 0.22720406 0.37132690
#>  [56,] 0.18868136 0.36875695 0.44256169
#>  [57,] 0.45191540 0.17257494 0.37550966
#>  [58,] 0.09354807 0.53620331 0.37024862
#>  [59,] 0.24169399 0.35790696 0.40039905
#>  [60,] 0.24063069 0.47306686 0.28630245
#>  [61,] 0.35072203 0.20657151 0.44270647
#>  [62,] 0.23644323 0.31757645 0.44598032
#>  [63,] 0.11377660 0.57925497 0.30696843
#>  [64,] 0.35397526 0.28189151 0.36413323
#>  [65,] 0.33767102 0.30057717 0.36175182
#>  [66,] 0.47681037 0.29273929 0.23045034
#>  [67,] 0.36148092 0.45283351 0.18568557
#>  [68,] 0.33060925 0.33090638 0.33848437
#>  [69,] 0.37124002 0.26291557 0.36584441
#>  [70,] 0.27271834 0.53023050 0.19705116
#>  [71,] 0.30243256 0.15987047 0.53769698
#>  [72,] 0.20119119 0.62642157 0.17238724
#>  [73,] 0.05610546 0.31400037 0.62989417
#>  [74,] 0.15443324 0.47144442 0.37412234
#>  [75,] 0.40902752 0.34735879 0.24361369
#>  [76,] 0.19754064 0.26151378 0.54094557
#>  [77,] 0.53891618 0.18529816 0.27578567
#>  [78,] 0.41188223 0.38225169 0.20586608
#>  [79,] 0.18473567 0.40259145 0.41267288
#>  [80,] 0.19063865 0.26506958 0.54429177
#>  [81,] 0.26923934 0.36639670 0.36436396
#>  [82,] 0.16515065 0.34390390 0.49094545
#>  [83,] 0.40108530 0.48616676 0.11274794
#>  [84,] 0.32501678 0.26945166 0.40553156
#>  [85,] 0.29233200 0.25614788 0.45152012
#>  [86,] 0.23576741 0.35543645 0.40879613
#>  [87,] 0.52754559 0.29771201 0.17474241
#>  [88,] 0.28031868 0.31097141 0.40870991
#>  [89,] 0.26963035 0.47377489 0.25659476
#>  [90,] 0.49633602 0.43858928 0.06507470
#>  [91,] 0.30124339 0.32311604 0.37564057
#>  [92,] 0.12143910 0.45008436 0.42847654
#>  [93,] 0.25784598 0.10338145 0.63877257
#>  [94,] 0.39487542 0.29363435 0.31149023
#>  [95,] 0.61931011 0.04807043 0.33261945
#>  [96,] 0.36704236 0.35944486 0.27351278
#>  [97,] 0.33603453 0.21460757 0.44935790
#>  [98,] 0.32267496 0.31008608 0.36723896
#>  [99,] 0.18219448 0.65648506 0.16132046
#> [100,] 0.32880516 0.43063171 0.24056314
rmevspec(n=100, d=4, param=c(0.2,0.1,0.9,0.5), model='bilog')
#>                [,1]         [,2]        [,3]         [,4]
#>   [1,] 5.234418e-01 2.825618e-01 0.091749316 1.022470e-01
#>   [2,] 4.000961e-01 2.862000e-01 0.089627773 2.240762e-01
#>   [3,] 8.147276e-02 1.148620e-01 0.017689832 7.859754e-01
#>   [4,] 2.992166e-01 2.737143e-01 0.282904647 1.441645e-01
#>   [5,] 4.405556e-04 3.018120e-04 0.999147683 1.099497e-04
#>   [6,] 4.729032e-01 2.862873e-01 0.107838118 1.329714e-01
#>   [7,] 2.397379e-01 3.325465e-01 0.051819134 3.758964e-01
#>   [8,] 4.971702e-05 4.370568e-05 0.999875779 3.079866e-05
#>   [9,] 3.053156e-01 2.678775e-01 0.257221575 1.695853e-01
#>  [10,] 4.151696e-01 2.845956e-01 0.017161818 2.830729e-01
#>  [11,] 3.408676e-01 3.840805e-01 0.031575710 2.434762e-01
#>  [12,] 5.020323e-01 2.852426e-01 0.055035061 1.576900e-01
#>  [13,] 2.720881e-01 3.805112e-01 0.136558376 2.108423e-01
#>  [14,] 2.113526e-01 3.505668e-01 0.043092849 3.949878e-01
#>  [15,] 3.342656e-01 2.856227e-01 0.193797959 1.863138e-01
#>  [16,] 9.380495e-02 1.255324e-01 0.508745335 2.719173e-01
#>  [17,] 2.138964e-01 3.572308e-01 0.113382847 3.154900e-01
#>  [18,] 4.583654e-01 3.059282e-01 0.017803665 2.179027e-01
#>  [19,] 4.388056e-01 3.502183e-01 0.022948523 1.880276e-01
#>  [20,] 3.096968e-01 3.311045e-01 0.020221560 3.389772e-01
#>  [21,] 3.025482e-01 3.483917e-01 0.101417812 2.476422e-01
#>  [22,] 3.072719e-01 3.720219e-01 0.075032805 2.456734e-01
#>  [23,] 3.387522e-01 3.684183e-01 0.059841039 2.329884e-01
#>  [24,] 3.318002e-01 2.872838e-01 0.218446281 1.624697e-01
#>  [25,] 2.927621e-01 3.166867e-01 0.109533992 2.810172e-01
#>  [26,] 3.356358e-01 2.227075e-01 0.021327794 4.203289e-01
#>  [27,] 2.392905e-01 3.712061e-01 0.059487103 3.300163e-01
#>  [28,] 4.610092e-01 3.220538e-01 0.023876364 1.930606e-01
#>  [29,] 4.944411e-02 8.239843e-02 0.007940288 8.602172e-01
#>  [30,] 8.753428e-02 1.069626e-01 0.121743850 6.837593e-01
#>  [31,] 3.568704e-01 2.174275e-01 0.014905967 4.107961e-01
#>  [32,] 1.707729e-01 1.435619e-01 0.606911638 7.875365e-02
#>  [33,] 3.606348e-01 2.967578e-01 0.144366469 1.982410e-01
#>  [34,] 3.278830e-01 2.549662e-01 0.118852627 2.982982e-01
#>  [35,] 1.454120e-01 2.541724e-01 0.036764104 5.636514e-01
#>  [36,] 2.852044e-02 4.271625e-02 0.002791577 9.259717e-01
#>  [37,] 2.492278e-01 2.953158e-01 0.155068342 3.003880e-01
#>  [38,] 2.420126e-01 3.463577e-01 0.070861497 3.407682e-01
#>  [39,] 2.398732e-01 3.096548e-01 0.052373112 3.980989e-01
#>  [40,] 3.226722e-01 3.712857e-01 0.063628385 2.424138e-01
#>  [41,] 3.120426e-01 3.582267e-01 0.067203531 2.625272e-01
#>  [42,] 2.925666e-01 4.586201e-01 0.065323141 1.834901e-01
#>  [43,] 2.838861e-01 3.425302e-01 0.038371330 3.352124e-01
#>  [44,] 3.276374e-01 2.492006e-01 0.017260608 4.059015e-01
#>  [45,] 3.548533e-02 3.875469e-02 0.886150427 3.960955e-02
#>  [46,] 2.824921e-01 3.013307e-01 0.132938839 2.832384e-01
#>  [47,] 2.313276e-01 2.270500e-01 0.019976182 5.216462e-01
#>  [48,] 3.482360e-01 3.287156e-01 0.025820795 2.972276e-01
#>  [49,] 3.417187e-01 3.491408e-01 0.073605946 2.355346e-01
#>  [50,] 4.648467e-01 3.158384e-01 0.017898104 2.014168e-01
#>  [51,] 2.704358e-01 2.997379e-01 0.230875005 1.989513e-01
#>  [52,] 3.582416e-01 4.349754e-01 0.069293011 1.374900e-01
#>  [53,] 1.965003e-01 2.549836e-01 0.431867060 1.166490e-01
#>  [54,] 2.529928e-01 3.500727e-01 0.105323619 2.916108e-01
#>  [55,] 2.884183e-01 4.144635e-01 0.018753558 2.783646e-01
#>  [56,] 4.856594e-01 3.278061e-01 0.027248292 1.592862e-01
#>  [57,] 3.228237e-01 3.202787e-01 0.049143322 3.077542e-01
#>  [58,] 2.836506e-01 3.571933e-01 0.077279737 2.818763e-01
#>  [59,] 1.417132e-06 1.410414e-06 0.999989145 8.027882e-06
#>  [60,] 4.334936e-01 3.779352e-01 0.050174686 1.383965e-01
#>  [61,] 2.455830e-01 2.127516e-01 0.415145818 1.265195e-01
#>  [62,] 4.227721e-01 3.544517e-01 0.054577858 1.681983e-01
#>  [63,] 4.189232e-01 3.814654e-01 0.021040709 1.785707e-01
#>  [64,] 1.062556e-01 1.351672e-01 0.645830521 1.127467e-01
#>  [65,] 2.825446e-01 3.398268e-01 0.110105375 2.675233e-01
#>  [66,] 3.328575e-01 4.297617e-01 0.046631188 1.907496e-01
#>  [67,] 2.588589e-01 2.080499e-01 0.015979754 5.171115e-01
#>  [68,] 2.578896e-01 3.002217e-01 0.096779678 3.451090e-01
#>  [69,] 3.182007e-01 3.158479e-01 0.113238907 2.527124e-01
#>  [70,] 1.688772e-01 3.161627e-01 0.054221351 4.607388e-01
#>  [71,] 2.899607e-01 3.570217e-01 0.012091503 3.409261e-01
#>  [72,] 1.416092e-01 1.734128e-01 0.624515045 6.046290e-02
#>  [73,] 4.110583e-01 3.162779e-01 0.030990175 2.416736e-01
#>  [74,] 1.999856e-01 3.921502e-01 0.049635406 3.582287e-01
#>  [75,] 4.099284e-01 2.890438e-01 0.051979526 2.490483e-01
#>  [76,] 2.882774e-01 3.863854e-01 0.063767786 2.615694e-01
#>  [77,] 3.076414e-01 2.981711e-01 0.015294100 3.788934e-01
#>  [78,] 2.091336e-04 2.460646e-04 0.999051305 4.934973e-04
#>  [79,] 3.424804e-01 3.844728e-01 0.054041933 2.190049e-01
#>  [80,] 4.453471e-01 3.305633e-01 0.020734182 2.033554e-01
#>  [81,] 4.078170e-01 3.538033e-01 0.062018819 1.763608e-01
#>  [82,] 3.218940e-01 3.493563e-01 0.039717130 2.890326e-01
#>  [83,] 3.913929e-01 3.612719e-01 0.095098604 1.522366e-01
#>  [84,] 3.245563e-01 3.824731e-01 0.045941462 2.470291e-01
#>  [85,] 9.461340e-02 1.398897e-01 0.031890555 7.336063e-01
#>  [86,] 1.954649e-01 3.805011e-01 0.099903213 3.241308e-01
#>  [87,] 3.491957e-01 3.474568e-01 0.174528237 1.288192e-01
#>  [88,] 3.307423e-01 5.166804e-01 0.028539277 1.240380e-01
#>  [89,] 2.979614e-01 3.815906e-01 0.111423171 2.090248e-01
#>  [90,] 3.258936e-01 3.531354e-01 0.136493865 1.844771e-01
#>  [91,] 2.188505e-01 2.307762e-01 0.082866839 4.675065e-01
#>  [92,] 3.118764e-01 2.987570e-01 0.012965719 3.764009e-01
#>  [93,] 2.982466e-01 3.542589e-01 0.031713263 3.157812e-01
#>  [94,] 4.092440e-01 3.028485e-01 0.020915236 2.669923e-01
#>  [95,] 4.795289e-01 3.480204e-01 0.035202608 1.372481e-01
#>  [96,] 3.816164e-01 4.121650e-01 0.083712141 1.225064e-01
#>  [97,] 2.942777e-01 3.228676e-01 0.067318791 3.155359e-01
#>  [98,] 2.344488e-02 2.721367e-02 0.001785496 9.475559e-01
#>  [99,] 2.946552e-01 3.625995e-01 0.026715439 3.160299e-01
#> [100,] 7.716272e-04 1.156103e-03 0.996882465 1.189805e-03
rmevspec(n=100, d=2, param=c(0.8,1.2), model='ct') #Dirichlet model
#>               [,1]       [,2]
#>   [1,] 0.766977261 0.23302274
#>   [2,] 0.519930388 0.48006961
#>   [3,] 0.820744261 0.17925574
#>   [4,] 0.419979154 0.58002085
#>   [5,] 0.257032610 0.74296739
#>   [6,] 0.269467110 0.73053289
#>   [7,] 0.712717139 0.28728286
#>   [8,] 0.946163061 0.05383694
#>   [9,] 0.148200178 0.85179982
#>  [10,] 0.932529683 0.06747032
#>  [11,] 0.500902590 0.49909741
#>  [12,] 0.005820861 0.99417914
#>  [13,] 0.837504514 0.16249549
#>  [14,] 0.206313419 0.79368658
#>  [15,] 0.061469923 0.93853008
#>  [16,] 0.740252796 0.25974720
#>  [17,] 0.906899052 0.09310095
#>  [18,] 0.182580411 0.81741959
#>  [19,] 0.105128366 0.89487163
#>  [20,] 0.210196940 0.78980306
#>  [21,] 0.787398754 0.21260125
#>  [22,] 0.426760240 0.57323976
#>  [23,] 0.572848582 0.42715142
#>  [24,] 0.841393884 0.15860612
#>  [25,] 0.906165356 0.09383464
#>  [26,] 0.818251952 0.18174805
#>  [27,] 0.530767747 0.46923225
#>  [28,] 0.221712806 0.77828719
#>  [29,] 0.252408617 0.74759138
#>  [30,] 0.598097185 0.40190282
#>  [31,] 0.123636850 0.87636315
#>  [32,] 0.919883457 0.08011654
#>  [33,] 0.585294663 0.41470534
#>  [34,] 0.846797386 0.15320261
#>  [35,] 0.002788189 0.99721181
#>  [36,] 0.456231460 0.54376854
#>  [37,] 0.944342132 0.05565787
#>  [38,] 0.720228473 0.27977153
#>  [39,] 0.056174747 0.94382525
#>  [40,] 0.024535819 0.97546418
#>  [41,] 0.407063474 0.59293653
#>  [42,] 0.644443921 0.35555608
#>  [43,] 0.099297136 0.90070286
#>  [44,] 0.187755511 0.81224449
#>  [45,] 0.322924817 0.67707518
#>  [46,] 0.185870580 0.81412942
#>  [47,] 0.478976953 0.52102305
#>  [48,] 0.024740507 0.97525949
#>  [49,] 0.683467211 0.31653279
#>  [50,] 0.209576356 0.79042364
#>  [51,] 0.946265322 0.05373468
#>  [52,] 0.637082835 0.36291717
#>  [53,] 0.964131589 0.03586841
#>  [54,] 0.673060393 0.32693961
#>  [55,] 0.763185392 0.23681461
#>  [56,] 0.346217017 0.65378298
#>  [57,] 0.430650977 0.56934902
#>  [58,] 0.636382521 0.36361748
#>  [59,] 0.922134342 0.07786566
#>  [60,] 0.663137544 0.33686246
#>  [61,] 0.617025311 0.38297469
#>  [62,] 0.333686878 0.66631312
#>  [63,] 0.012037393 0.98796261
#>  [64,] 0.636069187 0.36393081
#>  [65,] 0.784411071 0.21558893
#>  [66,] 0.122183026 0.87781697
#>  [67,] 0.757663137 0.24233686
#>  [68,] 0.964361390 0.03563861
#>  [69,] 0.220866868 0.77913313
#>  [70,] 0.399517393 0.60048261
#>  [71,] 0.188922139 0.81107786
#>  [72,] 0.466985876 0.53301412
#>  [73,] 0.054373722 0.94562628
#>  [74,] 0.303629555 0.69637045
#>  [75,] 0.416163629 0.58383637
#>  [76,] 0.930149695 0.06985031
#>  [77,] 0.883962304 0.11603770
#>  [78,] 0.191816520 0.80818348
#>  [79,] 0.243036649 0.75696335
#>  [80,] 0.715270062 0.28472994
#>  [81,] 0.635416344 0.36458366
#>  [82,] 0.672644246 0.32735575
#>  [83,] 0.038507119 0.96149288
#>  [84,] 0.488487293 0.51151271
#>  [85,] 0.882744144 0.11725586
#>  [86,] 0.644603088 0.35539691
#>  [87,] 0.391452089 0.60854791
#>  [88,] 0.880067026 0.11993297
#>  [89,] 0.671738180 0.32826182
#>  [90,] 0.825962377 0.17403762
#>  [91,] 0.586047446 0.41395255
#>  [92,] 0.928503950 0.07149605
#>  [93,] 0.199077589 0.80092241
#>  [94,] 0.720293639 0.27970636
#>  [95,] 0.345133839 0.65486616
#>  [96,] 0.543824284 0.45617572
#>  [97,] 0.848125157 0.15187484
#>  [98,] 0.971971982 0.02802802
#>  [99,] 0.482082874 0.51791713
#> [100,] 0.821567570 0.17843243
rmevspec(n=100, d=2, param=c(0.8,1.2,0.5), model='sdir') #with additional scale parameter
#>              [,1]      [,2]
#>   [1,] 0.31033028 0.6896697
#>   [2,] 0.21549227 0.7845077
#>   [3,] 0.64379436 0.3562056
#>   [4,] 0.26228207 0.7377179
#>   [5,] 0.66730331 0.3326967
#>   [6,] 0.50393433 0.4960657
#>   [7,] 0.67898667 0.3210133
#>   [8,] 0.69084895 0.3091511
#>   [9,] 0.40941193 0.5905881
#>  [10,] 0.62447557 0.3755244
#>  [11,] 0.47668604 0.5233140
#>  [12,] 0.31821825 0.6817818
#>  [13,] 0.45675225 0.5432478
#>  [14,] 0.28707504 0.7129250
#>  [15,] 0.59858006 0.4014199
#>  [16,] 0.79763964 0.2023604
#>  [17,] 0.19540249 0.8045975
#>  [18,] 0.70771239 0.2922876
#>  [19,] 0.70270616 0.2972938
#>  [20,] 0.49247456 0.5075254
#>  [21,] 0.60353584 0.3964642
#>  [22,] 0.89026684 0.1097332
#>  [23,] 0.47379389 0.5262061
#>  [24,] 0.43835902 0.5616410
#>  [25,] 0.74896478 0.2510352
#>  [26,] 0.37018946 0.6298105
#>  [27,] 0.71948491 0.2805151
#>  [28,] 0.60971315 0.3902869
#>  [29,] 0.51047985 0.4895202
#>  [30,] 0.15470757 0.8452924
#>  [31,] 0.60518550 0.3948145
#>  [32,] 0.72024357 0.2797564
#>  [33,] 0.64419590 0.3558041
#>  [34,] 0.41682518 0.5831748
#>  [35,] 0.33755441 0.6624456
#>  [36,] 0.08474767 0.9152523
#>  [37,] 0.62662745 0.3733725
#>  [38,] 0.31253529 0.6874647
#>  [39,] 0.05208558 0.9479144
#>  [40,] 0.53211240 0.4678876
#>  [41,] 0.54413106 0.4558689
#>  [42,] 0.55617317 0.4438268
#>  [43,] 0.39162534 0.6083747
#>  [44,] 0.68205993 0.3179401
#>  [45,] 0.70906490 0.2909351
#>  [46,] 0.74462746 0.2553725
#>  [47,] 0.36179558 0.6382044
#>  [48,] 0.60678767 0.3932123
#>  [49,] 0.45447010 0.5455299
#>  [50,] 0.73598393 0.2640161
#>  [51,] 0.43936710 0.5606329
#>  [52,] 0.52850042 0.4714996
#>  [53,] 0.80706413 0.1929359
#>  [54,] 0.65530137 0.3446986
#>  [55,] 0.18108883 0.8189112
#>  [56,] 0.56141792 0.4385821
#>  [57,] 0.54249638 0.4575036
#>  [58,] 0.21837390 0.7816261
#>  [59,] 0.59875813 0.4012419
#>  [60,] 0.21104405 0.7889560
#>  [61,] 0.51684138 0.4831586
#>  [62,] 0.58272519 0.4172748
#>  [63,] 0.59016229 0.4098377
#>  [64,] 0.40267059 0.5973294
#>  [65,] 0.85192388 0.1480761
#>  [66,] 0.44773633 0.5522637
#>  [67,] 0.65412018 0.3458798
#>  [68,] 0.23174180 0.7682582
#>  [69,] 0.32200372 0.6779963
#>  [70,] 0.46879534 0.5312047
#>  [71,] 0.53628372 0.4637163
#>  [72,] 0.75381695 0.2461830
#>  [73,] 0.16153925 0.8384608
#>  [74,] 0.25919633 0.7408037
#>  [75,] 0.65118218 0.3488178
#>  [76,] 0.33871641 0.6612836
#>  [77,] 0.51842896 0.4815710
#>  [78,] 0.76258268 0.2374173
#>  [79,] 0.81832970 0.1816703
#>  [80,] 0.45832607 0.5416739
#>  [81,] 0.53351175 0.4664883
#>  [82,] 0.49816844 0.5018316
#>  [83,] 0.80159363 0.1984064
#>  [84,] 0.51998032 0.4800197
#>  [85,] 0.63228176 0.3677182
#>  [86,] 0.76257052 0.2374295
#>  [87,] 0.44368209 0.5563179
#>  [88,] 0.56832605 0.4316739
#>  [89,] 0.43475333 0.5652467
#>  [90,] 0.31041012 0.6895899
#>  [91,] 0.26613044 0.7338696
#>  [92,] 0.57336625 0.4266337
#>  [93,] 0.56504137 0.4349586
#>  [94,] 0.63011638 0.3698836
#>  [95,] 0.53320745 0.4667925
#>  [96,] 0.59810146 0.4018985
#>  [97,] 0.29552933 0.7044707
#>  [98,] 0.46947712 0.5305229
#>  [99,] 0.44598246 0.5540175
#> [100,] 0.01266773 0.9873323
#Variogram gamma(h) = scale*||h||^alpha
#NEW: Variogram must take distance as argument
vario <- function(x, scale=0.5, alpha=0.8){ scale*x^alpha }
#grid specification
grid.coord <- as.matrix(expand.grid(runif(4), runif(4)))
rmevspec(n=100, vario=vario,coord=grid.coord, model='br')
#>              [,1]       [,2]       [,3]       [,4]        [,5]        [,6]
#>   [1,] 0.06689096 0.03419209 0.03318673 0.03265432 0.111367396 0.070053704
#>   [2,] 0.03086720 0.04408257 0.02793249 0.03271036 0.142888898 0.127601795
#>   [3,] 0.05837757 0.07155454 0.08657438 0.05408623 0.040579061 0.017302465
#>   [4,] 0.02664661 0.05210343 0.04629848 0.07474279 0.064414789 0.068330277
#>   [5,] 0.04965216 0.03774968 0.04503100 0.01963997 0.039613277 0.066867282
#>   [6,] 0.11652286 0.07084491 0.06842788 0.05420961 0.054682960 0.059878304
#>   [7,] 0.08398108 0.07394477 0.04236507 0.11222111 0.021982919 0.024872621
#>   [8,] 0.10705461 0.06837333 0.04720954 0.05175896 0.041544067 0.033927833
#>   [9,] 0.03326864 0.03380278 0.02778525 0.04881958 0.096003551 0.132837781
#>  [10,] 0.07058902 0.09069471 0.10007206 0.09166841 0.041898711 0.027965118
#>  [11,] 0.08707355 0.09305639 0.09369712 0.04858379 0.037569628 0.025192460
#>  [12,] 0.07953866 0.04737645 0.05466171 0.05127489 0.105758353 0.037444847
#>  [13,] 0.05123618 0.04925586 0.04886829 0.03604944 0.041953444 0.052460932
#>  [14,] 0.11817538 0.03674466 0.07227198 0.03107762 0.064204526 0.038765027
#>  [15,] 0.03976101 0.07474998 0.05971531 0.05098980 0.040514363 0.023821186
#>  [16,] 0.05287921 0.06290842 0.11463779 0.05440868 0.028921668 0.050155947
#>  [17,] 0.04075717 0.04701808 0.05463522 0.03473375 0.038988302 0.130722969
#>  [18,] 0.08788141 0.03464054 0.04754382 0.02779754 0.110552187 0.075771342
#>  [19,] 0.05058153 0.11886693 0.10392100 0.12202055 0.012437841 0.021095601
#>  [20,] 0.05394497 0.05538243 0.06533078 0.06080829 0.021069505 0.021136370
#>  [21,] 0.03976095 0.07824552 0.08053730 0.08589859 0.083357963 0.057382220
#>  [22,] 0.02952896 0.02654403 0.04685083 0.02467981 0.055123779 0.084285122
#>  [23,] 0.04919485 0.05425649 0.05437582 0.03504857 0.117660064 0.089717143
#>  [24,] 0.04806103 0.06493378 0.05889482 0.09569195 0.059349081 0.074858224
#>  [25,] 0.04707499 0.01412120 0.02183442 0.01857180 0.137622440 0.086399337
#>  [26,] 0.04654989 0.03212901 0.04004362 0.01942299 0.123792846 0.136359707
#>  [27,] 0.03805898 0.06016921 0.03478189 0.06611485 0.044439848 0.112540688
#>  [28,] 0.06287118 0.08597742 0.05770598 0.06099794 0.065196133 0.079668092
#>  [29,] 0.03894916 0.09354297 0.08741095 0.09306205 0.021533527 0.071198936
#>  [30,] 0.04035504 0.06320604 0.05578812 0.04573531 0.033193089 0.120564695
#>  [31,] 0.10120327 0.09789382 0.08595099 0.04649517 0.062359880 0.026541138
#>  [32,] 0.14557244 0.10441656 0.09503395 0.08418097 0.029231186 0.019887941
#>  [33,] 0.03498315 0.08878373 0.08065437 0.09329052 0.029451314 0.040770598
#>  [34,] 0.10368187 0.06119633 0.08201309 0.04528505 0.062831309 0.037624623
#>  [35,] 0.06211542 0.06570026 0.04913898 0.05057755 0.085152544 0.097056719
#>  [36,] 0.07806922 0.05362261 0.06088991 0.06762484 0.006260225 0.008688362
#>  [37,] 0.07259119 0.04751101 0.03974238 0.04621308 0.090587503 0.121061719
#>  [38,] 0.11425335 0.07179963 0.10663713 0.04374535 0.042322846 0.117005845
#>  [39,] 0.05459001 0.11429988 0.07208161 0.06676541 0.019351369 0.023659717
#>  [40,] 0.06154665 0.06733607 0.07268184 0.07588464 0.043087022 0.038042381
#>  [41,] 0.08193535 0.04318060 0.07800689 0.03261937 0.124940982 0.069613823
#>  [42,] 0.05593877 0.04263143 0.05256958 0.04220667 0.123914508 0.034876262
#>  [43,] 0.05613057 0.05159957 0.04710942 0.11598820 0.040303191 0.043911879
#>  [44,] 0.02291009 0.07888612 0.09302431 0.08468766 0.091171878 0.045947635
#>  [45,] 0.17446774 0.05642234 0.06472130 0.05437733 0.021939641 0.037533525
#>  [46,] 0.02357889 0.07397200 0.07759234 0.05277672 0.053401490 0.088459350
#>  [47,] 0.09194226 0.06105818 0.06112316 0.04698784 0.054775911 0.025730714
#>  [48,] 0.03250091 0.03810956 0.03667593 0.02551200 0.082677921 0.108970623
#>  [49,] 0.02470848 0.01863890 0.01297900 0.01705689 0.129698401 0.124993635
#>  [50,] 0.10002126 0.07165276 0.08915416 0.06543639 0.147608214 0.059448972
#>  [51,] 0.04412875 0.08679948 0.10905534 0.06259254 0.077630689 0.041141853
#>  [52,] 0.08253843 0.06326659 0.07691979 0.08159403 0.028570275 0.070512871
#>  [53,] 0.06771678 0.10038046 0.12764365 0.08464900 0.046736449 0.020657278
#>  [54,] 0.06414520 0.05308667 0.04915185 0.08961482 0.077728287 0.131127589
#>  [55,] 0.05439877 0.04802028 0.07011323 0.04536229 0.024562790 0.014644135
#>  [56,] 0.04409033 0.03412504 0.03441326 0.01345967 0.051351094 0.133317523
#>  [57,] 0.02261741 0.01406023 0.01817256 0.03982258 0.105148814 0.179680495
#>  [58,] 0.14018514 0.07974515 0.06532877 0.09939565 0.059162075 0.048404096
#>  [59,] 0.02236944 0.03296625 0.03143109 0.05481348 0.064374860 0.138359181
#>  [60,] 0.09569230 0.06666663 0.07542055 0.04971782 0.039004394 0.039532122
#>  [61,] 0.07352438 0.05511767 0.06986788 0.03693586 0.047549860 0.048893088
#>  [62,] 0.08426267 0.09386105 0.10200179 0.09519754 0.026429992 0.043762383
#>  [63,] 0.07714083 0.02746741 0.04997823 0.01824553 0.056370056 0.059418771
#>  [64,] 0.09014822 0.08331757 0.08467930 0.10960795 0.046002978 0.036243738
#>  [65,] 0.04971860 0.03447879 0.05081704 0.04024663 0.057488914 0.091777858
#>  [66,] 0.01066341 0.02048838 0.01975950 0.03756315 0.141983663 0.157637386
#>  [67,] 0.09596550 0.05637678 0.03360388 0.06173045 0.038954613 0.094700088
#>  [68,] 0.04053253 0.11013084 0.09817441 0.07116805 0.039475217 0.071688402
#>  [69,] 0.06302977 0.10369452 0.15286955 0.11614525 0.035467940 0.024622688
#>  [70,] 0.03539838 0.04839409 0.02831056 0.04361557 0.129489639 0.090517517
#>  [71,] 0.04231799 0.06861479 0.07995154 0.07210228 0.038895720 0.026714945
#>  [72,] 0.03038653 0.05082206 0.02061560 0.03831902 0.053767363 0.169065656
#>  [73,] 0.07187430 0.07446468 0.06025998 0.12823471 0.061468592 0.062919938
#>  [74,] 0.02476916 0.04596970 0.04772669 0.04026831 0.037831787 0.048960092
#>  [75,] 0.06110332 0.05526528 0.04277558 0.06613928 0.097293921 0.091404599
#>  [76,] 0.03226037 0.04334027 0.04252718 0.05125084 0.065929208 0.119866523
#>  [77,] 0.07653351 0.07626848 0.12442443 0.05352846 0.055138545 0.048674429
#>  [78,] 0.01799870 0.11812769 0.05240235 0.07895159 0.051949487 0.141649679
#>  [79,] 0.03757034 0.02853987 0.05150110 0.03839747 0.077871022 0.092706550
#>  [80,] 0.04129248 0.11964948 0.08307582 0.14941824 0.013622213 0.018952534
#>  [81,] 0.06684759 0.07714047 0.04468848 0.05537510 0.063689691 0.070771085
#>  [82,] 0.10533520 0.02788220 0.04140260 0.04449217 0.043692957 0.140841266
#>  [83,] 0.03424467 0.04112391 0.03072092 0.05917354 0.103800649 0.102024210
#>  [84,] 0.05035008 0.06630094 0.08696284 0.08314390 0.079607233 0.094492318
#>  [85,] 0.14142692 0.04947832 0.06316566 0.05985803 0.037013878 0.045762397
#>  [86,] 0.05836660 0.02930582 0.04743397 0.01755171 0.067421191 0.029703120
#>  [87,] 0.01465868 0.03240319 0.05399937 0.01940803 0.099594006 0.072988207
#>  [88,] 0.03242832 0.02854760 0.02214813 0.03309069 0.043571778 0.078935553
#>  [89,] 0.06531630 0.03396199 0.04621183 0.04668298 0.035699820 0.078205161
#>  [90,] 0.12401670 0.03922597 0.06444008 0.07168167 0.060038079 0.075570291
#>  [91,] 0.06973414 0.09944907 0.09201671 0.03408471 0.050518583 0.054493065
#>  [92,] 0.04859699 0.07881168 0.04730818 0.06288750 0.053355840 0.035967619
#>  [93,] 0.15739054 0.08130814 0.10482662 0.14334469 0.082196760 0.042586607
#>  [94,] 0.10917969 0.06705071 0.09957202 0.07481598 0.057454642 0.069435393
#>  [95,] 0.10981410 0.09642054 0.07334845 0.12141358 0.027316439 0.028705023
#>  [96,] 0.13282046 0.06252981 0.14236358 0.09268017 0.034203463 0.030526834
#>  [97,] 0.07425055 0.09996400 0.06400684 0.10226127 0.019200848 0.019949773
#>  [98,] 0.06148466 0.04527085 0.04494470 0.05214484 0.037673237 0.122899331
#>  [99,] 0.03498493 0.10779134 0.04540544 0.10737902 0.044595209 0.055634188
#> [100,] 0.06844374 0.07285118 0.06620756 0.05733647 0.045666607 0.049151349
#>               [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#>   [1,] 0.088307293 0.06360732 0.07374251 0.06222877 0.04967870 0.04553231
#>   [2,] 0.093872300 0.21625838 0.03554230 0.04027237 0.02997280 0.04211399
#>   [3,] 0.025726354 0.01887072 0.06898192 0.11269288 0.05918224 0.05127564
#>   [4,] 0.105543944 0.06633212 0.06620221 0.06503106 0.06957536 0.07561958
#>   [5,] 0.053854957 0.03462130 0.07472186 0.07195271 0.04471942 0.07293247
#>   [6,] 0.056619302 0.08931144 0.02503046 0.10239502 0.04878060 0.07368762
#>   [7,] 0.025362218 0.04806644 0.06587229 0.08083157 0.04731412 0.09433969
#>   [8,] 0.035587464 0.04808314 0.08084808 0.05364273 0.07972111 0.09958454
#>   [9,] 0.101611205 0.10496795 0.05315824 0.04841794 0.06056182 0.05064766
#>  [10,] 0.034861903 0.01802961 0.06378785 0.09248419 0.07624177 0.08155168
#>  [11,] 0.039099348 0.03404817 0.10545821 0.09118941 0.08173939 0.05327714
#>  [12,] 0.065561453 0.04877166 0.09776142 0.03996460 0.05466622 0.04652512
#>  [13,] 0.045868035 0.04249563 0.07220381 0.05259224 0.06525488 0.05069730
#>  [14,] 0.052423843 0.02145448 0.13400416 0.05435167 0.08063056 0.06096676
#>  [15,] 0.024959434 0.03590055 0.07787913 0.10396910 0.07610284 0.07763951
#>  [16,] 0.032282969 0.03850844 0.04555028 0.09260075 0.06203372 0.07642634
#>  [17,] 0.083293230 0.12877754 0.05193839 0.08193131 0.03819897 0.06467384
#>  [18,] 0.074258096 0.09285575 0.03510357 0.04741671 0.06002504 0.05348566
#>  [19,] 0.014693425 0.03825061 0.04302352 0.07532977 0.05163827 0.09872028
#>  [20,] 0.026586797 0.02171389 0.03480608 0.08949961 0.08182927 0.11540050
#>  [21,] 0.041743018 0.06279773 0.03767976 0.04703137 0.08565990 0.06586384
#>  [22,] 0.053425769 0.11480926 0.09902310 0.04693573 0.06628878 0.02624715
#>  [23,] 0.049661833 0.09159126 0.06756021 0.05253572 0.06091696 0.03020827
#>  [24,] 0.057776659 0.08535216 0.04264843 0.08883543 0.04252896 0.08155803
#>  [25,] 0.201546769 0.08002422 0.08058945 0.02484652 0.02287369 0.02108534
#>  [26,] 0.128935985 0.08586659 0.03910911 0.04123044 0.03362182 0.04604369
#>  [27,] 0.086377392 0.12456417 0.05601516 0.05535416 0.02899277 0.07947340
#>  [28,] 0.091273208 0.09248775 0.08092056 0.05177609 0.04467140 0.04172500
#>  [29,] 0.034830468 0.10549377 0.03260673 0.03909557 0.04149893 0.07747203
#>  [30,] 0.097966297 0.10359742 0.03927308 0.07533827 0.04101642 0.10986357
#>  [31,] 0.057876173 0.03809728 0.07549750 0.06850732 0.06993457 0.05169158
#>  [32,] 0.016750967 0.02804387 0.04650547 0.06587340 0.05927118 0.05722157
#>  [33,] 0.043242760 0.05310526 0.09089322 0.05408916 0.05908290 0.05559775
#>  [34,] 0.037191481 0.04288185 0.03918567 0.04873803 0.05059213 0.08075541
#>  [35,] 0.082101180 0.08627753 0.02392722 0.06109757 0.04271556 0.06294119
#>  [36,] 0.006442646 0.00847324 0.05004185 0.15573008 0.04428619 0.10137263
#>  [37,] 0.079792440 0.07526864 0.11033758 0.03618461 0.04676890 0.03295379
#>  [38,] 0.060527600 0.11383848 0.04666290 0.05293509 0.03417329 0.02620246
#>  [39,] 0.026006423 0.02346083 0.05539198 0.08945129 0.11035843 0.07943521
#>  [40,] 0.048310196 0.04280741 0.08853253 0.09779936 0.04611352 0.08317004
#>  [41,] 0.125291166 0.04383351 0.12865510 0.01987303 0.03794226 0.02414713
#>  [42,] 0.070688389 0.06151732 0.08313041 0.05319871 0.06564291 0.04160468
#>  [43,] 0.048147417 0.06191795 0.07880701 0.07200137 0.04939406 0.09822058
#>  [44,] 0.042172224 0.04797150 0.03931723 0.10758664 0.07662451 0.04539466
#>  [45,] 0.035956608 0.03879927 0.08759229 0.06991707 0.09045861 0.04319258
#>  [46,] 0.111623519 0.06924258 0.04138012 0.04858470 0.05491170 0.05716484
#>  [47,] 0.024394283 0.02240480 0.05966423 0.09516356 0.06854462 0.08105169
#>  [48,] 0.142608856 0.11361878 0.04147478 0.05397158 0.03969859 0.08746064
#>  [49,] 0.143541358 0.12060214 0.03813359 0.04424991 0.05305308 0.05151165
#>  [50,] 0.062873200 0.04992731 0.04747315 0.03687904 0.04366612 0.05632581
#>  [51,] 0.036907464 0.03675597 0.08445890 0.06086569 0.05862902 0.06308498
#>  [52,] 0.079301621 0.04075070 0.03812464 0.07518043 0.09073669 0.05070183
#>  [53,] 0.032001712 0.04029305 0.03913476 0.10531227 0.11044776 0.08236746
#>  [54,] 0.164194199 0.07562990 0.04443740 0.03035345 0.03246706 0.04365937
#>  [55,] 0.016730054 0.01893080 0.19894864 0.04063841 0.08327119 0.05931388
#>  [56,] 0.072119043 0.07705038 0.09968767 0.05136149 0.03811585 0.03423552
#>  [57,] 0.142893332 0.10649909 0.06062636 0.02873621 0.02257614 0.04857731
#>  [58,] 0.068670872 0.03905182 0.05204331 0.06005186 0.04955136 0.05526497
#>  [59,] 0.084413506 0.26586680 0.03345470 0.03580784 0.02656921 0.03465226
#>  [60,] 0.043714443 0.03217195 0.09676824 0.05640480 0.08988418 0.06122066
#>  [61,] 0.050055150 0.05224983 0.07639312 0.06811839 0.05683464 0.06025877
#>  [62,] 0.038134368 0.04318523 0.07996943 0.05590230 0.07851538 0.04846637
#>  [63,] 0.063774282 0.04842956 0.05844334 0.06865524 0.10546217 0.04312702
#>  [64,] 0.043083074 0.04435567 0.05362946 0.09635835 0.05987523 0.05573043
#>  [65,] 0.058760959 0.08419518 0.06077448 0.06551982 0.07444275 0.06635108
#>  [66,] 0.135956096 0.17366894 0.03628848 0.03398895 0.03747314 0.04463895
#>  [67,] 0.038795221 0.05293417 0.05495337 0.05340394 0.06907844 0.08611556
#>  [68,] 0.064504335 0.04682699 0.04501224 0.04416583 0.08775959 0.07148344
#>  [69,] 0.028036418 0.02213899 0.06554821 0.06542467 0.05218192 0.07333962
#>  [70,] 0.089105656 0.11544731 0.04932489 0.05114790 0.05273989 0.05345377
#>  [71,] 0.029002107 0.03660796 0.06304548 0.07858591 0.07674641 0.07569943
#>  [72,] 0.121128554 0.23332538 0.02067680 0.03795576 0.02558041 0.05370734
#>  [73,] 0.076758927 0.06358456 0.04093528 0.02483879 0.04598549 0.04643736
#>  [74,] 0.085882111 0.07526337 0.05295390 0.08063293 0.06098478 0.10432496
#>  [75,] 0.085419586 0.04846566 0.03517607 0.07052796 0.06354022 0.07837547
#>  [76,] 0.084144699 0.18796538 0.03322904 0.03833249 0.05090169 0.06468637
#>  [77,] 0.036241037 0.05011957 0.03144267 0.08585213 0.07856750 0.10635661
#>  [78,] 0.140383474 0.10856393 0.02317142 0.03852356 0.03966407 0.05028578
#>  [79,] 0.085708672 0.08357473 0.06703092 0.07067469 0.06179158 0.03013245
#>  [80,] 0.013619317 0.01913382 0.02394167 0.10623419 0.06430348 0.10807157
#>  [81,] 0.042673929 0.08267075 0.06847688 0.03690568 0.04500475 0.04349234
#>  [82,] 0.090290562 0.11776291 0.08250866 0.02674470 0.05120480 0.02580627
#>  [83,] 0.055652851 0.12423910 0.05462187 0.07786176 0.04111654 0.05688982
#>  [84,] 0.107349780 0.04716152 0.02353983 0.06151356 0.04226506 0.05927588
#>  [85,] 0.050827233 0.08038947 0.06973033 0.06298697 0.06202684 0.05422164
#>  [86,] 0.042732510 0.03026000 0.08154807 0.06299476 0.05280577 0.05075487
#>  [87,] 0.117489537 0.06166321 0.03391884 0.03996799 0.04233250 0.07435753
#>  [88,] 0.060869625 0.11994392 0.06868526 0.07772355 0.05729606 0.05369423
#>  [89,] 0.046899834 0.09888445 0.06875105 0.03322664 0.07990918 0.05826881
#>  [90,] 0.064171640 0.06006206 0.04193523 0.07063682 0.04757005 0.05025724
#>  [91,] 0.054690407 0.09386209 0.06522462 0.05007859 0.05471525 0.03820220
#>  [92,] 0.043287938 0.03908146 0.07071513 0.05339708 0.06342713 0.10537769
#>  [93,] 0.063757877 0.04598519 0.05606507 0.03655072 0.02686345 0.02488903
#>  [94,] 0.068580768 0.10363098 0.04672524 0.03419812 0.05635151 0.04844006
#>  [95,] 0.037577882 0.03097003 0.06937210 0.04420894 0.05182266 0.07029491
#>  [96,] 0.030029376 0.04870494 0.05703682 0.04803826 0.04053069 0.03945640
#>  [97,] 0.020894483 0.02527045 0.09440453 0.08062377 0.11874694 0.05701472
#>  [98,] 0.085730726 0.14048417 0.03344449 0.04857705 0.04404587 0.02905032
#>  [99,] 0.051806097 0.04846867 0.04865368 0.05620416 0.06712611 0.07066217
#> [100,] 0.029682193 0.06667454 0.06157017 0.05303083 0.07914909 0.06434674
#>             [,13]      [,14]      [,15]      [,16]
#>   [1,] 0.10883386 0.05220354 0.08056719 0.02695331
#>   [2,] 0.04507902 0.02259803 0.04215490 0.02605260
#>   [3,] 0.08016678 0.06727938 0.14246537 0.04488448
#>   [4,] 0.07252341 0.04543081 0.05291330 0.04829186
#>   [5,] 0.05062053 0.11892700 0.11726816 0.10182823
#>   [6,] 0.01920461 0.05622671 0.02830732 0.07587039
#>   [7,] 0.08334214 0.05192232 0.03878090 0.10480074
#>   [8,] 0.03761312 0.06649568 0.07744985 0.07110595
#>   [9,] 0.07045203 0.03230511 0.04623030 0.05913017
#>  [10,] 0.05881656 0.02809705 0.06365036 0.05959099
#>  [11,] 0.08140534 0.02588619 0.06814672 0.03457713
#>  [12,] 0.08689815 0.06582981 0.06345925 0.05450739
#>  [13,] 0.08072013 0.15250718 0.08480123 0.07303543
#>  [14,] 0.04882024 0.05795008 0.06984134 0.05831766
#>  [15,] 0.04868310 0.09581821 0.07140862 0.09808786
#>  [16,] 0.05485839 0.07454587 0.08885501 0.07042652
#>  [17,] 0.03713834 0.05352414 0.05653104 0.05713772
#>  [18,] 0.05099454 0.07892135 0.05874793 0.06400451
#>  [19,] 0.03247720 0.07614890 0.09518378 0.04561079
#>  [20,] 0.03207877 0.08224366 0.06986582 0.16830327
#>  [21,] 0.04254955 0.06002039 0.05424492 0.07722698
#>  [22,] 0.12175167 0.06696266 0.07726422 0.06027914
#>  [23,] 0.06204528 0.06719638 0.07615580 0.04187535
#>  [24,] 0.03912851 0.06660707 0.05689099 0.03688487
#>  [25,] 0.12339385 0.04059053 0.03579241 0.04363304
#>  [26,] 0.05876941 0.04075477 0.07743558 0.04993456
#>  [27,] 0.03411830 0.06118979 0.03510144 0.08270796
#>  [28,] 0.05525824 0.04437988 0.04869053 0.03640059
#>  [29,] 0.03804111 0.05998271 0.04611408 0.11916699
#>  [30,] 0.02746302 0.04105005 0.04058695 0.06500262
#>  [31,] 0.05477846 0.05184250 0.06347587 0.04785448
#>  [32,] 0.02052181 0.08538570 0.04918831 0.09291469
#>  [33,] 0.05793498 0.09118984 0.05438556 0.07254488
#>  [34,] 0.04920851 0.07603263 0.08754224 0.09523977
#>  [35,] 0.04350423 0.04752948 0.04345020 0.09671439
#>  [36,] 0.05895940 0.08626978 0.06788870 0.14538033
#>  [37,] 0.09170954 0.02900050 0.05001645 0.03026066
#>  [38,] 0.04161766 0.05884365 0.03434823 0.03508649
#>  [39,] 0.07488784 0.06307831 0.08330275 0.04387893
#>  [40,] 0.08397273 0.05280059 0.05954696 0.03836806
#>  [41,] 0.10360142 0.02519264 0.04370030 0.01746643
#>  [42,] 0.10455773 0.03325201 0.07984521 0.05442541
#>  [43,] 0.08396855 0.04869538 0.06101286 0.04279199
#>  [44,] 0.03734426 0.05697247 0.05587831 0.07411050
#>  [45,] 0.05022813 0.04935506 0.06456481 0.06047371
#>  [46,] 0.04661510 0.07047451 0.05746348 0.07275865
#>  [47,] 0.05649779 0.08171371 0.09312028 0.07582698
#>  [48,] 0.04127645 0.04719191 0.05685476 0.05139671
#>  [49,] 0.05223928 0.06029029 0.05920305 0.04910034
#>  [50,] 0.03002289 0.04053498 0.04587959 0.05309617
#>  [51,] 0.03654356 0.06688897 0.04746039 0.08705642
#>  [52,] 0.04986374 0.05950439 0.05644126 0.05599270
#>  [53,] 0.03461244 0.03914828 0.04306112 0.02583752
#>  [54,] 0.02438637 0.04115337 0.03482010 0.04404436
#>  [55,] 0.12242666 0.06113896 0.08232466 0.05917526
#>  [56,] 0.13533344 0.06357568 0.06088359 0.05688041
#>  [57,] 0.06489309 0.03450445 0.04826449 0.06292744
#>  [58,] 0.04450415 0.05044086 0.03488384 0.05331606
#>  [59,] 0.05287183 0.03494546 0.05420906 0.03289504
#>  [60,] 0.04448657 0.07600584 0.08373444 0.04957505
#>  [61,] 0.09869799 0.07176924 0.07494008 0.05879405
#>  [62,] 0.07813316 0.03910412 0.05130427 0.04176995
#>  [63,] 0.11084745 0.05841323 0.08018086 0.07404601
#>  [64,] 0.03632111 0.06273999 0.04557610 0.05233084
#>  [65,] 0.07858109 0.08478320 0.06729528 0.03476834
#>  [66,] 0.05025371 0.02129621 0.04483237 0.03350768
#>  [67,] 0.03243457 0.08128368 0.07252861 0.07714114
#>  [68,] 0.03492983 0.05926861 0.07970004 0.03517964
#>  [69,] 0.05065072 0.05226621 0.04381634 0.05076716
#>  [70,] 0.04431747 0.04900669 0.07536570 0.04436498
#>  [71,] 0.06989778 0.08403274 0.08438733 0.07339757
#>  [72,] 0.01956250 0.04882929 0.02614128 0.05011645
#>  [73,] 0.07317414 0.05778718 0.04822910 0.06304696
#>  [74,] 0.03473032 0.08407263 0.05475048 0.12087880
#>  [75,] 0.06839668 0.04121659 0.06015928 0.03474051
#>  [76,] 0.03669424 0.04724522 0.05498684 0.04663963
#>  [77,] 0.03334020 0.04536214 0.04463127 0.05351901
#>  [78,] 0.01885210 0.04902711 0.03101379 0.03943528
#>  [79,] 0.06724372 0.07761561 0.07369223 0.05594903
#>  [80,] 0.02156770 0.07173637 0.06893427 0.07644685
#>  [81,] 0.06454728 0.08409562 0.08886169 0.06475866
#>  [82,] 0.06487434 0.04576198 0.06707621 0.02432318
#>  [83,] 0.03235822 0.06279570 0.05276539 0.07061085
#>  [84,] 0.03941110 0.06196739 0.03069298 0.06596558
#>  [85,] 0.06531940 0.05757831 0.04657711 0.05363748
#>  [86,] 0.08968594 0.13306288 0.04595729 0.16041550
#>  [87,] 0.03638051 0.11311216 0.09093036 0.09679587
#>  [88,] 0.08167644 0.07071888 0.10694093 0.06372903
#>  [89,] 0.06997220 0.06181265 0.08873040 0.08746670
#>  [90,] 0.06706438 0.05874632 0.04709498 0.05748851
#>  [91,] 0.07824455 0.06256927 0.04242856 0.05968819
#>  [92,] 0.06719682 0.07429647 0.06671947 0.08957301
#>  [93,] 0.03767022 0.02743364 0.02211001 0.04702144
#>  [94,] 0.04117866 0.03480836 0.04730884 0.04126903
#>  [95,] 0.07901188 0.04049128 0.05035312 0.06887905
#>  [96,] 0.06926098 0.06225026 0.06079418 0.04877377
#>  [97,] 0.06392050 0.04724089 0.06630341 0.04594703
#>  [98,] 0.07758943 0.06474143 0.06048845 0.05143044
#>  [99,] 0.04479673 0.07933847 0.05630716 0.08084665
#> [100,] 0.07385705 0.06045183 0.06275027 0.08883039
## Example with Dirichlet mixture
alpha.mat <- cbind(c(2,1,1),c(1,2,1),c(1,1,2))
rmevspec(n=100, param=alpha.mat, weights=rep(1/3,3), model='dirmix')
#>               [,1]        [,2]        [,3]
#>   [1,] 0.051624206 0.565315301 0.383060493
#>   [2,] 0.292282001 0.304269583 0.403448416
#>   [3,] 0.324185845 0.659988135 0.015826020
#>   [4,] 0.279996338 0.006084283 0.713919378
#>   [5,] 0.410676189 0.165214714 0.424109096
#>   [6,] 0.166331501 0.143378493 0.690290006
#>   [7,] 0.335044320 0.296851263 0.368104417
#>   [8,] 0.001103986 0.032358488 0.966537526
#>   [9,] 0.675424888 0.071382835 0.253192277
#>  [10,] 0.048144959 0.704544504 0.247310537
#>  [11,] 0.934558255 0.044922077 0.020519668
#>  [12,] 0.081308762 0.072628956 0.846062281
#>  [13,] 0.599748312 0.309001432 0.091250256
#>  [14,] 0.348993343 0.251433428 0.399573229
#>  [15,] 0.148495327 0.357389492 0.494115181
#>  [16,] 0.196381104 0.184024611 0.619594285
#>  [17,] 0.046903129 0.357654274 0.595442598
#>  [18,] 0.462578486 0.482724556 0.054696957
#>  [19,] 0.100272177 0.568400869 0.331326954
#>  [20,] 0.360603828 0.335437516 0.303958655
#>  [21,] 0.218019779 0.735225570 0.046754651
#>  [22,] 0.041809629 0.314788075 0.643402295
#>  [23,] 0.325269550 0.045354086 0.629376364
#>  [24,] 0.631135021 0.047730547 0.321134432
#>  [25,] 0.275284121 0.015849536 0.708866342
#>  [26,] 0.075310954 0.880030070 0.044658976
#>  [27,] 0.044726679 0.408560062 0.546713259
#>  [28,] 0.013920534 0.391235378 0.594844088
#>  [29,] 0.224853494 0.490862200 0.284284307
#>  [30,] 0.145032035 0.660696539 0.194271425
#>  [31,] 0.113766205 0.193104624 0.693129171
#>  [32,] 0.322618734 0.464373161 0.213008105
#>  [33,] 0.090716753 0.674220958 0.235062289
#>  [34,] 0.836458208 0.140142906 0.023398886
#>  [35,] 0.051427549 0.716845629 0.231726821
#>  [36,] 0.482117593 0.175474504 0.342407903
#>  [37,] 0.095364188 0.415084609 0.489551203
#>  [38,] 0.256249575 0.110867586 0.632882839
#>  [39,] 0.482912445 0.367510022 0.149577532
#>  [40,] 0.095829924 0.287681536 0.616488540
#>  [41,] 0.417702262 0.117541517 0.464756221
#>  [42,] 0.076540656 0.383311212 0.540148131
#>  [43,] 0.543017439 0.403889759 0.053092802
#>  [44,] 0.473818077 0.525157116 0.001024807
#>  [45,] 0.476077135 0.234624285 0.289298580
#>  [46,] 0.125483155 0.516540314 0.357976531
#>  [47,] 0.811099304 0.080464442 0.108436254
#>  [48,] 0.498960667 0.377833833 0.123205500
#>  [49,] 0.325289274 0.477180263 0.197530464
#>  [50,] 0.074661996 0.489954957 0.435383047
#>  [51,] 0.328643862 0.263208238 0.408147900
#>  [52,] 0.056983719 0.647784806 0.295231475
#>  [53,] 0.052310138 0.407313823 0.540376039
#>  [54,] 0.164416326 0.001008764 0.834574910
#>  [55,] 0.373081570 0.208051527 0.418866903
#>  [56,] 0.277973935 0.359816087 0.362209978
#>  [57,] 0.411820124 0.131151996 0.457027881
#>  [58,] 0.237658542 0.003739442 0.758602016
#>  [59,] 0.293020242 0.101459461 0.605520298
#>  [60,] 0.053555279 0.100687095 0.845757625
#>  [61,] 0.534509819 0.232428425 0.233061755
#>  [62,] 0.623489122 0.307689633 0.068821245
#>  [63,] 0.000872198 0.012842629 0.986285173
#>  [64,] 0.063523258 0.705060342 0.231416400
#>  [65,] 0.584545647 0.322473981 0.092980372
#>  [66,] 0.137737078 0.629636476 0.232626446
#>  [67,] 0.409524927 0.479651294 0.110823778
#>  [68,] 0.643473687 0.003279862 0.353246451
#>  [69,] 0.292754204 0.210159584 0.497086213
#>  [70,] 0.456357891 0.129145054 0.414497055
#>  [71,] 0.622010756 0.114667705 0.263321539
#>  [72,] 0.502693236 0.359310612 0.137996152
#>  [73,] 0.160116786 0.223763841 0.616119373
#>  [74,] 0.292684174 0.254071655 0.453244172
#>  [75,] 0.134368927 0.859244303 0.006386770
#>  [76,] 0.554294338 0.329626352 0.116079310
#>  [77,] 0.212524571 0.151609938 0.635865490
#>  [78,] 0.185234997 0.631332050 0.183432953
#>  [79,] 0.440873843 0.410828500 0.148297656
#>  [80,] 0.812155361 0.069109039 0.118735600
#>  [81,] 0.017699534 0.883738429 0.098562037
#>  [82,] 0.150730297 0.501074082 0.348195621
#>  [83,] 0.449489271 0.526265715 0.024245014
#>  [84,] 0.033392967 0.572462804 0.394144228
#>  [85,] 0.609137021 0.145780761 0.245082218
#>  [86,] 0.322946932 0.091147566 0.585905502
#>  [87,] 0.182731735 0.318284388 0.498983876
#>  [88,] 0.117953165 0.432439921 0.449606914
#>  [89,] 0.859238406 0.031548816 0.109212778
#>  [90,] 0.407984008 0.275249138 0.316766854
#>  [91,] 0.787141834 0.006553961 0.206304205
#>  [92,] 0.295450221 0.143463312 0.561086467
#>  [93,] 0.045156248 0.651384335 0.303459417
#>  [94,] 0.016928195 0.882938485 0.100133320
#>  [95,] 0.296193730 0.047605779 0.656200492
#>  [96,] 0.157383957 0.818592557 0.024023486
#>  [97,] 0.524836450 0.402186653 0.072976897
#>  [98,] 0.235766285 0.246037258 0.518196458
#>  [99,] 0.215101228 0.574004880 0.210893892
#> [100,] 0.225255254 0.049861721 0.724883025