Hill's estimator of the shape parameter

Given a sample of positive observations, calculate the tail index or shape parameter. The shape estimate returned is positive.

Usage

shape.hill(xdat, k)

Arguments

xdat

vector of positive observations

k

vector of order statistics; if missing, a vector going from 10 to sample size minus one.

Value

a data frame with the number of order statistics k and the shape parameter estimate shape, or a single numeric value if k is a scalar integer.

References

Hill, B.M. (1975). A simple general approach to inference about the tail of a distribution. Annals of Statistics, 3, 1163-1173.

Examples

xdat <- mev::rgp(n = 200, loc = 1, scale = 0.5, shape = 0.5)
shape.hill(xdat)
#>       k     shape
#> 1    10 0.4426683
#> 2    11 0.5633855
#> 3    12 0.5472549
#> 4    13 0.5485229
#> 5    14 0.6467767
#> 6    15 0.7228104
#> 7    16 0.6865752
#> 8    17 0.6627097
#> 9    18 0.7200167
#> 10   19 0.6978114
#> 11   20 0.6784871
#> 12   21 0.6640905
#> 13   22 0.6695848
#> 14   23 0.6626214
#> 15   24 0.6445207
#> 16   25 0.6307411
#> 17   26 0.6091874
#> 18   27 0.6105774
#> 19   28 0.5927844
#> 20   29 0.5731796
#> 21   30 0.5686297
#> 22   31 0.5666772
#> 23   32 0.5648807
#> 24   33 0.5495421
#> 25   34 0.5391372
#> 26   35 0.5617625
#> 27   36 0.5522086
#> 28   37 0.5423674
#> 29   38 0.5376236
#> 30   39 0.5470458
#> 31   40 0.5392985
#> 32   41 0.5296147
#> 33   42 0.5231792
#> 34   43 0.5167968
#> 35   44 0.5070945
#> 36   45 0.5132877
#> 37   46 0.5161882
#> 38   47 0.5062529
#> 39   48 0.5350228
#> 40   49 0.5597543
#> 41   50 0.5599716
#> 42   51 0.5644356
#> 43   52 0.5589574
#> 44   53 0.5625257
#> 45   54 0.5568498
#> 46   55 0.5469943
#> 47   56 0.5443387
#> 48   57 0.5350962
#> 49   58 0.5421817
#> 50   59 0.5342857
#> 51   60 0.5256150
#> 52   61 0.5181152
#> 53   62 0.5128090
#> 54   63 0.5052956
#> 55   64 0.5067161
#> 56   65 0.5028907
#> 57   66 0.5033724
#> 58   67 0.5022940
#> 59   68 0.5078908
#> 60   69 0.5061425
#> 61   70 0.5148218
#> 62   71 0.5186885
#> 63   72 0.5170225
#> 64   73 0.5244256
#> 65   74 0.5184143
#> 66   75 0.5213267
#> 67   76 0.5247316
#> 68   77 0.5319622
#> 69   78 0.5262654
#> 70   79 0.5247686
#> 71   80 0.5254467
#> 72   81 0.5235410
#> 73   82 0.5171936
#> 74   83 0.5150814
#> 75   84 0.5362610
#> 76   85 0.5392294
#> 77   86 0.5372442
#> 78   87 0.5418629
#> 79   88 0.5371465
#> 80   89 0.5398455
#> 81   90 0.5368086
#> 82   91 0.5366398
#> 83   92 0.5399349
#> 84   93 0.5380976
#> 85   94 0.5330399
#> 86   95 0.5295831
#> 87   96 0.5278283
#> 88   97 0.5329346
#> 89   98 0.5301547
#> 90   99 0.5254141
#> 91  100 0.5217999
#> 92  101 0.5211613
#> 93  102 0.5186722
#> 94  103 0.5155115
#> 95  104 0.5170153
#> 96  105 0.5129363
#> 97  106 0.5092004
#> 98  107 0.5115093
#> 99  108 0.5093249
#> 100 109 0.5133900
#> 101 110 0.5150191
#> 102 111 0.5107469
#> 103 112 0.5140772
#> 104 113 0.5122474
#> 105 114 0.5091909
#> 106 115 0.5201185
#> 107 116 0.5186661
#> 108 117 0.5192002
#> 109 118 0.5260075
#> 110 119 0.5237335
#> 111 120 0.5206732
#> 112 121 0.5171877
#> 113 122 0.5190841
#> 114 123 0.5236810
#> 115 124 0.5284869
#> 116 125 0.5315033
#> 117 126 0.5375400
#> 118 127 0.5355941
#> 119 128 0.5343819
#> 120 129 0.5367225
#> 121 130 0.5345219
#> 122 131 0.5314048
#> 123 132 0.5281245
#> 124 133 0.5253632
#> 125 134 0.5229487
#> 126 135 0.5199706
#> 127 136 0.5257043
#> 128 137 0.5253617
#> 129 138 0.5247728
#> 130 139 0.5219988
#> 131 140 0.5200289
#> 132 141 0.5183074
#> 133 142 0.5166262
#> 134 143 0.5153115
#> 135 144 0.5143395
#> 136 145 0.5155066
#> 137 146 0.5122767
#> 138 147 0.5125582
#> 139 148 0.5101666
#> 140 149 0.5078397
#> 141 150 0.5100148
#> 142 151 0.5086600
#> 143 152 0.5078504
#> 144 153 0.5101571
#> 145 154 0.5071452
#> 146 155 0.5074129
#> 147 156 0.5074886
#> 148 157 0.5065422
#> 149 158 0.5039974
#> 150 159 0.5037112
#> 151 160 0.5017031
#> 152 161 0.5035377
#> 153 162 0.5015425
#> 154 163 0.5027260
#> 155 164 0.5022197
#> 156 165 0.5049175
#> 157 166 0.5025792
#> 158 167 0.5027807
#> 159 168 0.5036395
#> 160 169 0.5033837
#> 161 170 0.5022055
#> 162 171 0.5016286
#> 163 172 0.5011482
#> 164 173 0.4988054
#> 165 174 0.4973985
#> 166 175 0.5007446
#> 167 176 0.5002069
#> 168 177 0.4993009
#> 169 178 0.5074279
#> 170 179 0.5076236
#> 171 180 0.5115662
#> 172 181 0.5194843
#> 173 182 0.5169448
#> 174 183 0.5193606
#> 175 184 0.5187695
#> 176 185 0.5257670
#> 177 186 0.5232059
#> 178 187 0.5253074
#> 179 188 0.5226451
#> 180 189 0.5265903
#> 181 190 0.5281111
#> 182 191 0.5271696
#> 183 192 0.5270476
#> 184 193 0.5254860
#> 185 194 0.5239062
#> 186 195 0.5215288
#> 187 196 0.5235409
#> 188 197 0.5216290
#> 189 198 0.5228225
#> 190 199 0.5278483