Lower-trimmed Hill estimator for the shape parameter

Given a sample of Pareto-tailed samples (positive tail index), compute the lower-trimmed Hill estimator. If \(k0=k\), the estimator reduces to Hill's estimator for the shape index

Usage

shape.lthill(xdat, k, k0 = k, sorted = FALSE, ...)

Arguments

xdat

[numeric] vector of positive observations

k

[integer] number of order statistics for the threshold

k0

[integer] vector of number of largest order statistics, no greater than k

sorted

[logical] if TRUE, data are assumed to be sorted in decreasing order.

...

additional arguments for other routines (notably vectorize)

Value

a scalar with the shape parameter estimate if k0 is a scalar, otherwise a data frame with columns k0 for the number of exceedances and shape for the tail index.

References

Bladt, M., Albrecher, H. & Beirlant, J. (2020) Threshold selection and trimming in extremes. Extremes, 23, 629-665 . doi:10.1007/s10687-020-00385-0

Examples

# Pareto sample
n <- 200
xdat <- 10/(1 - runif(n)) - 10
shape.lthill(xdat = xdat, k = 100, k0 = 5:100)
#>     k0    shape
#> 1    5 1.086956
#> 2    6 1.083652
#> 3    7 1.087468
#> 4    8 1.092676
#> 5    9 1.099353
#> 6   10 1.105379
#> 7   11 1.113381
#> 8   12 1.122984
#> 9   13 1.133318
#> 10  14 1.142943
#> 11  15 1.150032
#> 12  16 1.157003
#> 13  17 1.163310
#> 14  18 1.170349
#> 15  19 1.176650
#> 16  20 1.183040
#> 17  21 1.189657
#> 18  22 1.194514
#> 19  23 1.200038
#> 20  24 1.205283
#> 21  25 1.211072
#> 22  26 1.217385
#> 23  27 1.223413
#> 24  28 1.229756
#> 25  29 1.235242
#> 26  30 1.239296
#> 27  31 1.243772
#> 28  32 1.248267
#> 29  33 1.251850
#> 30  34 1.254364
#> 31  35 1.257297
#> 32  36 1.260367
#> 33  37 1.262907
#> 34  38 1.264004
#> 35  39 1.265197
#> 36  40 1.266630
#> 37  41 1.268136
#> 38  42 1.269165
#> 39  43 1.270147
#> 40  44 1.271415
#> 41  45 1.272850
#> 42  46 1.274460
#> 43  47 1.275954
#> 44  48 1.277686
#> 45  49 1.279546
#> 46  50 1.281539
#> 47  51 1.283109
#> 48  52 1.284426
#> 49  53 1.285805
#> 50  54 1.287310
#> 51  55 1.288306
#> 52  56 1.289059
#> 53  57 1.289868
#> 54  58 1.290790
#> 55  59 1.291711
#> 56  60 1.292848
#> 57  61 1.294180
#> 58  62 1.295439
#> 59  63 1.296699
#> 60  64 1.297596
#> 61  65 1.298685
#> 62  66 1.299812
#> 63  67 1.301013
#> 64  68 1.302388
#> 65  69 1.303594
#> 66  70 1.304761
#> 67  71 1.305754
#> 68  72 1.306575
#> 69  73 1.307328
#> 70  74 1.307876
#> 71  75 1.308602
#> 72  76 1.309434
#> 73  77 1.310394
#> 74  78 1.311330
#> 75  79 1.312144
#> 76  80 1.313093
#> 77  81 1.313861
#> 78  82 1.314596
#> 79  83 1.314501
#> 80  84 1.314390
#> 81  85 1.314165
#> 82  86 1.313951
#> 83  87 1.313853
#> 84  88 1.313885
#> 85  89 1.313907
#> 86  90 1.314044
#> 87  91 1.314284
#> 88  92 1.314350
#> 89  93 1.314237
#> 90  94 1.314090
#> 91  95 1.314081
#> 92  96 1.314160
#> 93  97 1.314330
#> 94  98 1.314285
#> 95  99 1.314183
#> 96 100 1.314213