Semi-parametric marginal transformation to uniform

The function spunif transforms a matrix or vector of data x to the pseudo-uniform scale using a semiparametric transform. Data below the threshold are transformed to pseudo-uniforms using a rank transform, while data above the threshold are assumed to follow a generalized Pareto distribution. The parameters of the latter are estimated using maximum likelihood if either scale = NULL or shape = NULL.

Usage

spunif(x, thresh, scale = NULL, shape = NULL)

Arguments

x

matrix or vector of data

thresh

vector of marginal thresholds

scale

vector of marginal scale parameters for the generalized Pareto

shape

vector of marginal shape parameters for the generalized Pareto

Value

a matrix or vector of the same dimension as x, with pseudo-uniform observations

Author

Leo Belzile

Examples

x <- rmev(1000, d = 3, param = 2, model = 'log')
thresh <- apply(x, 2, quantile, 0.95)
spunif(x, thresh)
#>                [,1]        [,2]        [,3]
#>    [1,] 0.829170829 0.179820180 0.328671329
#>    [2,] 0.860139860 0.826173826 0.824175824
#>    [3,] 0.443556444 0.182817183 0.171828172
#>    [4,] 0.922077922 0.714285714 0.873126873
#>    [5,] 0.616383616 0.688311688 0.624375624
#>    [6,] 0.479520480 0.718281718 0.590409590
#>    [7,] 0.964688190 0.904095904 0.981821385
#>    [8,] 0.060939061 0.101898102 0.009990010
#>    [9,] 0.835164835 0.687312687 0.397602398
#>   [10,] 0.187812188 0.682317682 0.222777223
#>   [11,] 0.756243756 0.457542458 0.611388611
#>   [12,] 0.782217782 0.852147852 0.895104895
#>   [13,] 0.138861139 0.183816184 0.206793207
#>   [14,] 0.384615385 0.593406593 0.182817183
#>   [15,] 0.674325674 0.592407592 0.722277722
#>   [16,] 0.901098901 0.847152847 0.635364635
#>   [17,] 0.769230769 0.907092907 0.856143856
#>   [18,] 0.209790210 0.239760240 0.315684316
#>   [19,] 0.503496503 0.689310689 0.637362637
#>   [20,] 0.890109890 0.950907031 0.908091908
#>   [21,] 0.848151848 0.914085914 0.790209790
#>   [22,] 0.003996004 0.240759241 0.035964036
#>   [23,] 0.409590410 0.087912088 0.107892108
#>   [24,] 0.650349650 0.423576424 0.721278721
#>   [25,] 0.660339660 0.800199800 0.737262737
#>   [26,] 0.652347652 0.849150849 0.474525475
#>   [27,] 0.791208791 0.521478521 0.370629371
#>   [28,] 0.301698302 0.807192807 0.682317682
#>   [29,] 0.843156843 0.667332667 0.495504496
#>   [30,] 0.198801199 0.758241758 0.445554446
#>   [31,] 0.820179820 0.410589411 0.651348651
#>   [32,] 0.851148851 0.599400599 0.690309690
#>   [33,] 0.323676324 0.155844156 0.304695305
#>   [34,] 0.757242757 0.881118881 0.788211788
#>   [35,] 0.779220779 0.340659341 0.492507493
#>   [36,] 0.855144855 0.775224775 0.544455544
#>   [37,] 0.598401598 0.809190809 0.610389610
#>   [38,] 0.160839161 0.558441558 0.031968032
#>   [39,] 0.291708292 0.484515485 0.336663337
#>   [40,] 0.078921079 0.046953047 0.472527473
#>   [41,] 0.386613387 0.253746254 0.264735265
#>   [42,] 0.447552448 0.228771229 0.241758242
#>   [43,] 0.068931069 0.014985015 0.058941059
#>   [44,] 0.177822178 0.243756244 0.030969031
#>   [45,] 0.481518482 0.378621379 0.475524476
#>   [46,] 0.441558442 0.605394605 0.844155844
#>   [47,] 0.921078921 0.384615385 0.870129870
#>   [48,] 0.277722278 0.200799201 0.835164835
#>   [49,] 0.583416583 0.143856144 0.354645355
#>   [50,] 0.325674326 0.040959041 0.828171828
#>   [51,] 0.403596404 0.617382617 0.516483516
#>   [52,] 0.496503497 0.187812188 0.305694306
#>   [53,] 0.979170427 0.889110889 0.949050949
#>   [54,] 0.718281718 0.053946054 0.281718282
#>   [55,] 0.560439560 0.832167832 0.535464535
#>   [56,] 0.005994006 0.217782218 0.013986014
#>   [57,] 0.444555445 0.213786214 0.340659341
#>   [58,] 0.897102897 0.900099900 0.923076923
#>   [59,] 0.061938062 0.199800200 0.065934066
#>   [60,] 0.966748139 0.927072927 0.946053946
#>   [61,] 0.798201798 0.509490509 0.560439560
#>   [62,] 0.046953047 0.181818182 0.186813187
#>   [63,] 0.986865455 0.974579932 0.950431000
#>   [64,] 0.947052947 0.660339660 0.567432567
#>   [65,] 0.809190809 0.556443556 0.481518482
#>   [66,] 0.064935065 0.162837163 0.436563437
#>   [67,] 0.552447552 0.760239760 0.531468531
#>   [68,] 0.309690310 0.011988012 0.069930070
#>   [69,] 0.353646354 0.491508492 0.599400599
#>   [70,] 0.290709291 0.526473526 0.307692308
#>   [71,] 0.269730270 0.111888112 0.205794206
#>   [72,] 0.335664336 0.147852148 0.059940060
#>   [73,] 0.602397602 0.216783217 0.600399600
#>   [74,] 0.609390609 0.952286726 0.702297702
#>   [75,] 0.648351648 0.553446553 0.201798202
#>   [76,] 0.072927073 0.097902098 0.102897103
#>   [77,] 0.985020281 0.977760636 0.972766276
#>   [78,] 0.421578422 0.979438135 0.794205794
#>   [79,] 0.567432567 0.633366633 0.829170829
#>   [80,] 0.079920080 0.380619381 0.095904096
#>   [81,] 0.442557443 0.302697303 0.739260739
#>   [82,] 0.172827173 0.254745255 0.540459540
#>   [83,] 0.837162837 0.771228771 0.843156843
#>   [84,] 0.938061938 0.860139860 0.887112887
#>   [85,] 0.276723277 0.517482517 0.420579421
#>   [86,] 0.531468531 0.719280719 0.572427572
#>   [87,] 0.380619381 0.080919081 0.719280719
#>   [88,] 0.555444555 0.194805195 0.372627373
#>   [89,] 0.746253746 0.877122877 0.758241758
#>   [90,] 0.482517483 0.394605395 0.446553447
#>   [91,] 0.772227772 0.805194805 0.676323676
#>   [92,] 0.205794206 0.511488511 0.394605395
#>   [93,] 0.457542458 0.505494505 0.605394605
#>   [94,] 0.869130869 0.500499500 0.619380619
#>   [95,] 0.505494505 0.759240759 0.393606394
#>   [96,] 0.400599401 0.459540460 0.344655345
#>   [97,] 0.575424575 0.209790210 0.273726274
#>   [98,] 0.121878122 0.226773227 0.274725275
#>   [99,] 0.709290709 0.708291708 0.762237762
#>  [100,] 0.904095904 0.863136863 0.866133866
#>  [101,] 0.657342657 0.958403973 0.520479520
#>  [102,] 0.422577423 0.876123876 0.192807193
#>  [103,] 0.706293706 0.602397602 0.821178821
#>  [104,] 0.333666334 0.376623377 0.140859141
#>  [105,] 0.755244755 0.727272727 0.199800200
#>  [106,] 0.143856144 0.311688312 0.966999516
#>  [107,] 0.182817183 0.261738262 0.111888112
#>  [108,] 0.071928072 0.072927073 0.229770230
#>  [109,] 0.452547453 0.898101898 0.395604396
#>  [110,] 0.029970030 0.036963037 0.138861139
#>  [111,] 0.515484515 0.743256743 0.641358641
#>  [112,] 0.972399699 0.966325367 0.966448023
#>  [113,] 0.412587413 0.342657343 0.462537463
#>  [114,] 0.812187812 0.794205794 0.692307692
#>  [115,] 0.493506494 0.661338661 0.450549451
#>  [116,] 0.541458541 0.467532468 0.174825175
#>  [117,] 0.341658342 0.084915085 0.021978022
#>  [118,] 0.611388611 0.566433566 0.245754246
#>  [119,] 0.726273726 0.710289710 0.912087912
#>  [120,] 0.836163836 0.755244755 0.797202797
#>  [121,] 0.524475524 0.475524476 0.280719281
#>  [122,] 0.749250749 0.456543457 0.258741259
#>  [123,] 0.816183816 0.894105894 0.679320679
#>  [124,] 0.789210789 0.823176823 0.832167832
#>  [125,] 0.146853147 0.579420579 0.148851149
#>  [126,] 0.431568432 0.539460539 0.554445554
#>  [127,] 0.631368631 0.546453546 0.905094905
#>  [128,] 0.151848152 0.215784216 0.167832168
#>  [129,] 0.044955045 0.145854146 0.170829171
#>  [130,] 0.725274725 0.731268731 0.756243756
#>  [131,] 0.581418581 0.626373626 0.834165834
#>  [132,] 0.410589411 0.586413586 0.539460539
#>  [133,] 0.058941059 0.005994006 0.018981019
#>  [134,] 0.311688312 0.678321678 0.323676324
#>  [135,] 0.220779221 0.041958042 0.310689311
#>  [136,] 0.065934066 0.403596404 0.296703297
#>  [137,] 0.874125874 0.778221778 0.979165375
#>  [138,] 0.100899101 0.409590410 0.128871129
#>  [139,] 0.246753247 0.365634366 0.346653347
#>  [140,] 0.800199800 0.781218781 0.951208378
#>  [141,] 0.729270729 0.924075924 0.961824853
#>  [142,] 0.680319680 0.393606394 0.650349650
#>  [143,] 0.931068931 0.963799509 0.943056943
#>  [144,] 0.372627373 0.066933067 0.525474525
#>  [145,] 0.596403596 0.867132867 0.596403596
#>  [146,] 0.888111888 0.872127872 0.644355644
#>  [147,] 0.274725275 0.051948052 0.066933067
#>  [148,] 0.983056214 0.932067932 0.921078921
#>  [149,] 0.817182817 0.899100899 0.952281498
#>  [150,] 0.475524476 0.833166833 0.631368631
#>  [151,] 0.562437562 0.267732268 0.451548452
#>  [152,] 0.407592408 0.186813187 0.424575425
#>  [153,] 0.051948052 0.399600400 0.414585415
#>  [154,] 0.507492507 0.756243756 0.589410589
#>  [155,] 0.774225774 0.600399600 0.766233766
#>  [156,] 0.681318681 0.757242757 0.932067932
#>  [157,] 0.165834166 0.034965035 0.002997003
#>  [158,] 0.207792208 0.355644356 0.151848152
#>  [159,] 0.484515485 0.098901099 0.145854146
#>  [160,] 0.438561439 0.434565435 0.373626374
#>  [161,] 0.404595405 0.415584416 0.254745255
#>  [162,] 0.926073926 0.787212787 0.792207792
#>  [163,] 0.120879121 0.032967033 0.146853147
#>  [164,] 0.951200113 0.627372627 0.666333666
#>  [165,] 0.256743257 0.010989011 0.763236763
#>  [166,] 0.453546454 0.584415584 0.629370629
#>  [167,] 0.349650350 0.776223776 0.449550450
#>  [168,] 0.411588412 0.326673327 0.377622378
#>  [169,] 0.280719281 0.278721279 0.216783217
#>  [170,] 0.830169830 0.936063936 0.678321678
#>  [171,] 0.289710290 0.453546454 0.526473526
#>  [172,] 0.346653347 0.811188811 0.913086913
#>  [173,] 0.458541459 0.411588412 0.249750250
#>  [174,] 0.367632368 0.722277722 0.291708292
#>  [175,] 0.669330669 0.297702298 0.640359640
#>  [176,] 0.020979021 0.001998002 0.020979021
#>  [177,] 0.845154845 0.871128871 0.812187812
#>  [178,] 0.847152847 0.482517483 0.602397602
#>  [179,] 0.530469530 0.220779221 0.493506494
#>  [180,] 0.157842158 0.144855145 0.165834166
#>  [181,] 0.283716284 0.736263736 0.169830170
#>  [182,] 0.491508492 0.612387612 0.161838162
#>  [183,] 0.420579421 0.221778222 0.116883117
#>  [184,] 0.047952048 0.374625375 0.112887113
#>  [185,] 0.486513487 0.879120879 0.461538462
#>  [186,] 0.906093906 0.742257742 0.859140859
#>  [187,] 0.497502498 0.545454545 0.815184815
#>  [188,] 0.784215784 0.738261738 0.097902098
#>  [189,] 0.231768232 0.543456543 0.326673327
#>  [190,] 0.610389610 0.799200799 0.429570430
#>  [191,] 0.302697303 0.580419580 0.435564436
#>  [192,] 0.226773227 0.572427572 0.542457542
#>  [193,] 0.702297702 0.571428571 0.820179820
#>  [194,] 0.385614386 0.283716284 0.329670330
#>  [195,] 0.185814186 0.085914086 0.916083916
#>  [196,] 0.028971029 0.636363636 0.033966034
#>  [197,] 0.190809191 0.561438561 0.507492507
#>  [198,] 0.692307692 0.685314685 0.176823177
#>  [199,] 0.547452547 0.578421578 0.288711289
#>  [200,] 0.125874126 0.168831169 0.004995005
#>  [201,] 0.137862138 0.470529471 0.232767233
#>  [202,] 0.416583417 0.485514486 0.473526474
#>  [203,] 0.550449550 0.330669331 0.184815185
#>  [204,] 0.892107892 0.818181818 0.840159840
#>  [205,] 0.903096903 0.828171828 0.787212787
#>  [206,] 0.118881119 0.677322677 0.001998002
#>  [207,] 0.312687313 0.025974026 0.073926074
#>  [208,] 0.979304222 0.988681789 0.977492188
#>  [209,] 0.850149850 0.857142857 0.782217782
#>  [210,] 0.173826174 0.280719281 0.221778222
#>  [211,] 0.796203796 0.446553447 0.953416510
#>  [212,] 0.532467532 0.865134865 0.517482517
#>  [213,] 0.298701299 0.076923077 0.276723277
#>  [214,] 0.682317682 0.564435564 0.729270729
#>  [215,] 0.347652348 0.574425574 0.638361638
#>  [216,] 0.946053946 0.979715139 0.968962590
#>  [217,] 0.950886235 0.888111888 0.935064935
#>  [218,] 0.647352647 0.132867133 0.557442557
#>  [219,] 0.200799201 0.222777223 0.299700300
#>  [220,] 0.436563437 0.534465534 0.319680320
#>  [221,] 0.854145854 0.765234765 0.942057942
#>  [222,] 0.994361657 0.998372242 0.993062317
#>  [223,] 0.582417582 0.670329670 0.614385614
#>  [224,] 0.179820180 0.102897103 0.051948052
#>  [225,] 0.760239760 0.128871129 0.410589411
#>  [226,] 0.579420579 0.256743257 0.584415584
#>  [227,] 0.573426573 0.316683317 0.330669331
#>  [228,] 0.504495504 0.432567433 0.327672328
#>  [229,] 0.587412587 0.601398601 0.837162837
#>  [230,] 0.695304695 0.929070929 0.769230769
#>  [231,] 0.721278721 0.921078921 0.607392607
#>  [232,] 0.180819181 0.163836164 0.166833167
#>  [233,] 0.009990010 0.018981019 0.101898102
#>  [234,] 0.111888112 0.137862138 0.072927073
#>  [235,] 0.235764236 0.189810190 0.057942058
#>  [236,] 0.194805195 0.631368631 0.705294705
#>  [237,] 0.483516484 0.803196803 0.799200799
#>  [238,] 0.414585415 0.210789211 0.267732268
#>  [239,] 0.871128871 0.590409590 0.730269730
#>  [240,] 0.108891109 0.065934066 0.337662338
#>  [241,] 0.645354645 0.720279720 0.431568432
#>  [242,] 0.268731269 0.133866134 0.217782218
#>  [243,] 0.612387612 0.792207792 0.300699301
#>  [244,] 0.199800200 0.659340659 0.399600400
#>  [245,] 0.873126873 0.656343656 0.767232767
#>  [246,] 0.639360639 0.354645355 0.415584416
#>  [247,] 0.840159840 0.839160839 0.872127872
#>  [248,] 0.265734266 0.156843157 0.550449550
#>  [249,] 0.970830785 0.969326997 0.958057487
#>  [250,] 0.520479520 0.549450549 0.556443556
#>  [251,] 0.293706294 0.430569431 0.720279720
#>  [252,] 0.105894106 0.000999001 0.064935065
#>  [253,] 0.132867133 0.573426573 0.110889111
#>  [254,] 0.574425574 0.648351648 0.879120879
#>  [255,] 0.270729271 0.705294705 0.698301698
#>  [256,] 0.710289710 0.377622378 0.627372627
#>  [257,] 0.517482517 0.477522478 0.838161838
#>  [258,] 0.506493506 0.681318681 0.718281718
#>  [259,] 0.345654346 0.211788212 0.008991009
#>  [260,] 0.624375624 0.386613387 0.322677323
#>  [261,] 0.253746254 0.385614386 0.543456543
#>  [262,] 0.388611389 0.510489510 0.712287712
#>  [263,] 0.010989011 0.012987013 0.029970030
#>  [264,] 0.448551449 0.004995005 0.187812188
#>  [265,] 0.152847153 0.074925075 0.484515485
#>  [266,] 0.978625481 0.950511674 0.990313804
#>  [267,] 0.715284715 0.649350649 0.341658342
#>  [268,] 0.168831169 0.229770230 0.528471528
#>  [269,] 0.485514486 0.441558442 0.583416583
#>  [270,] 0.272727273 0.675324675 0.255744256
#>  [271,] 0.113886114 0.191808192 0.672327672
#>  [272,] 0.013986014 0.038961039 0.313686314
#>  [273,] 0.123876124 0.023976024 0.713286713
#>  [274,] 0.273726274 0.313686314 0.302697303
#>  [275,] 0.424575425 0.064935065 0.266733267
#>  [276,] 0.382617383 0.610389610 0.703296703
#>  [277,] 0.543456543 0.177822178 0.508491508
#>  [278,] 0.872127872 0.885114885 0.861138861
#>  [279,] 0.474525475 0.115884116 0.092907093
#>  [280,] 0.928071928 0.843156843 0.871128871
#>  [281,] 0.752247752 0.813186813 0.689310689
#>  [282,] 0.661338661 0.499500500 0.545454545
#>  [283,] 0.203796204 0.300699301 0.776223776
#>  [284,] 0.242757243 0.408591409 0.714285714
#>  [285,] 0.656343656 0.325674326 0.727272727
#>  [286,] 0.184815185 0.324675325 0.425574426
#>  [287,] 0.433566434 0.528471528 0.412587413
#>  [288,] 0.662337662 0.657342657 0.701298701
#>  [289,] 0.510489510 0.425574426 0.240759241
#>  [290,] 0.867132867 0.908091908 0.817182817
#>  [291,] 0.249750250 0.608391608 0.662337662
#>  [292,] 0.738261738 0.480519481 0.428571429
#>  [293,] 0.238761239 0.075924076 0.225774226
#>  [294,] 0.623376623 0.662337662 0.831168831
#>  [295,] 0.588411588 0.702297702 0.944055944
#>  [296,] 0.708291708 0.692307692 0.548451548
#>  [297,] 0.310689311 0.227772228 0.082917083
#>  [298,] 0.239760240 0.063936064 0.125874126
#>  [299,] 0.626373626 0.699300699 0.471528472
#>  [300,] 0.548451548 0.557442557 0.725274725
#>  [301,] 0.429570430 0.875124875 0.262737263
#>  [302,] 0.728271728 0.750249750 0.779220779
#>  [303,] 0.134865135 0.244755245 0.141858142
#>  [304,] 0.264735265 0.414585415 0.577422577
#>  [305,] 0.344655345 0.761238761 0.437562438
#>  [306,] 0.917082917 0.978432026 0.886113886
#>  [307,] 0.968917846 0.938061938 0.878121878
#>  [308,] 0.285714286 0.045954046 0.212787213
#>  [309,] 0.086913087 0.152847153 0.568431568
#>  [310,] 0.039960040 0.352647353 0.419580420
#>  [311,] 0.018981019 0.042957043 0.463536464
#>  [312,] 0.949050949 0.982789276 0.930069930
#>  [313,] 0.351648352 0.402597403 0.503496503
#>  [314,] 0.781218781 0.777222777 0.962837148
#>  [315,] 0.927072927 0.939060939 0.955001887
#>  [316,] 0.742257742 0.769230769 0.553446553
#>  [317,] 0.750249750 0.845154845 0.825174825
#>  [318,] 0.785214785 0.903096903 0.805194805
#>  [319,] 0.133866134 0.439560440 0.726273726
#>  [320,] 0.878121878 0.589410589 0.371628372
#>  [321,] 0.328671329 0.207792208 0.056943057
#>  [322,] 0.833166833 0.568431568 0.903096903
#>  [323,] 0.528471528 0.635364635 0.317682318
#>  [324,] 0.375624376 0.532467532 0.411588412
#>  [325,] 0.033966034 0.159840160 0.034965035
#>  [326,] 0.381618382 0.554445554 0.223776224
#>  [327,] 0.212787213 0.044955045 0.246753247
#>  [328,] 0.263736264 0.450549451 0.489510490
#>  [329,] 0.905094905 0.901098901 0.842157842
#>  [330,] 0.092907093 0.008991009 0.529470529
#>  [331,] 0.542457542 0.492507493 0.144855145
#>  [332,] 0.569430569 0.435564436 0.744255744
#>  [333,] 0.365634366 0.233766234 0.655344655
#>  [334,] 0.296703297 0.266733267 0.129870130
#>  [335,] 0.155844156 0.766233766 0.318681319
#>  [336,] 0.535464535 0.103896104 0.765234765
#>  [337,] 0.511488511 0.645354645 0.279720280
#>  [338,] 0.377622378 0.310689311 0.364635365
#>  [339,] 0.131868132 0.504495504 0.269730270
#>  [340,] 0.909090909 0.864135864 0.857142857
#>  [341,] 0.278721279 0.341658342 0.380619381
#>  [342,] 0.208791209 0.314685315 0.442557443
#>  [343,] 0.225774226 0.031968032 0.534465534
#>  [344,] 0.834165834 0.494505495 0.432567433
#>  [345,] 0.620379620 0.821178821 0.652347652
#>  [346,] 0.373626374 0.417582418 0.261738262
#>  [347,] 0.426573427 0.238761239 0.194805195
#>  [348,] 0.389610390 0.747252747 0.510489510
#>  [349,] 0.939060939 0.837162837 0.917082917
#>  [350,] 0.440559441 0.322677323 0.440559441
#>  [351,] 0.258741259 0.632367632 0.237762238
#>  [352,] 0.853146853 0.824175824 0.736263736
#>  [353,] 0.332667333 0.124875125 0.405594406
#>  [354,] 0.940059940 0.947052947 0.909090909
#>  [355,] 0.912087912 0.397602398 0.434565435
#>  [356,] 0.795204795 0.607392607 0.555444555
#>  [357,] 0.186813187 0.668331668 0.119880120
#>  [358,] 0.521478521 0.184815185 0.311688312
#>  [359,] 0.460539461 0.654345654 0.153846154
#>  [360,] 0.161838162 0.445554446 0.228771229
#>  [361,] 0.055944056 0.003996004 0.024975025
#>  [362,] 0.245754246 0.318681319 0.094905095
#>  [363,] 0.363636364 0.519480519 0.827172827
#>  [364,] 0.162837163 0.055944056 0.010989011
#>  [365,] 0.705294705 0.895104895 0.890109890
#>  [366,] 0.703296703 0.597402597 0.789210789
#>  [367,] 0.934065934 0.949050949 0.896103896
#>  [368,] 0.063936064 0.407592408 0.038961039
#>  [369,] 0.052947053 0.002997003 0.087912088
#>  [370,] 0.144855145 0.763236763 0.754245754
#>  [371,] 0.141858142 0.583416583 0.179820180
#>  [372,] 0.637362637 0.092907093 0.162837163
#>  [373,] 0.685314685 0.303696304 0.303696304
#>  [374,] 0.594405594 0.120879121 0.374625375
#>  [375,] 0.735264735 0.241758242 0.423576424
#>  [376,] 0.329670330 0.286713287 0.361638362
#>  [377,] 0.210789211 0.148851149 0.132867133
#>  [378,] 0.566433566 0.820179820 0.579420579
#>  [379,] 0.932067932 0.934065934 0.865134865
#>  [380,] 0.494505495 0.317682318 0.282717283
#>  [381,] 0.392607393 0.404595405 0.231768232
#>  [382,] 0.665334665 0.919080919 0.753246753
#>  [383,] 0.266733267 0.424575425 0.263736264
#>  [384,] 0.688311688 0.339660340 0.684315684
#>  [385,] 0.807192807 0.652347652 0.755244755
#>  [386,] 0.035964036 0.203796204 0.409590410
#>  [387,] 0.374625375 0.873126873 0.809190809
#>  [388,] 0.677322677 0.272727273 0.499500500
#>  [389,] 0.679320679 0.022977023 0.000999001
#>  [390,] 0.733266733 0.831168831 0.741258741
#>  [391,] 0.480519481 0.611388611 0.349650350
#>  [392,] 0.739260739 0.208791209 0.456543457
#>  [393,] 0.376623377 0.419580420 0.173826174
#>  [394,] 0.175824176 0.058941059 0.773226773
#>  [395,] 0.689310689 0.287712288 0.697302697
#>  [396,] 0.659340659 0.717282717 0.647352647
#>  [397,] 0.634365634 0.798201798 0.919080919
#>  [398,] 0.636363636 0.454545455 0.505494505
#>  [399,] 0.241758242 0.690309690 0.308691309
#>  [400,] 0.585414585 0.265734266 0.250749251
#>  [401,] 0.181818182 0.336663337 0.407592408
#>  [402,] 0.576423576 0.262737263 0.224775225
#>  [403,] 0.405594406 0.451548452 0.242757243
#>  [404,] 0.672327672 0.897102897 0.808191808
#>  [405,] 0.255744256 0.154845155 0.086913087
#>  [406,] 0.354645355 0.150849151 0.164835165
#>  [407,] 0.958297586 0.835164835 0.804195804
#>  [408,] 0.102897103 0.676323676 0.133866134
#>  [409,] 0.262737263 0.840159840 0.122877123
#>  [410,] 0.937062937 0.782217782 0.976166733
#>  [411,] 0.116883117 0.106893107 0.017982018
#>  [412,] 0.740259740 0.598401598 0.454545455
#>  [413,] 0.193806194 0.785214785 0.028971029
#>  [414,] 0.050949051 0.486513487 0.043956044
#>  [415,] 0.006993007 0.282717283 0.063936064
#>  [416,] 0.025974026 0.773226773 0.272727273
#>  [417,] 0.613386613 0.364635365 0.183816184
#>  [418,] 0.401598402 0.604395604 0.877122877
#>  [419,] 0.230769231 0.606393606 0.019980020
#>  [420,] 0.900099900 0.741258741 0.791208791
#>  [421,] 0.537462537 0.563436563 0.780219780
#>  [422,] 0.534465534 0.321678322 0.459540460
#>  [423,] 0.084915085 0.062937063 0.724275724
#>  [424,] 0.731268731 0.230769231 0.551448551
#>  [425,] 0.419580420 0.616383616 0.606393606
#>  [426,] 0.956972354 0.972782476 0.956585622
#>  [427,] 0.211788212 0.387612388 0.244755245
#>  [428,] 0.119880120 0.329670330 0.195804196
#>  [429,] 0.951233074 0.969193145 0.928071928
#>  [430,] 0.022977023 0.569430569 0.465534466
#>  [431,] 0.306693307 0.223776224 0.196803197
#>  [432,] 0.736263736 0.804195804 0.124875125
#>  [433,] 0.292707293 0.214785215 0.074925075
#>  [434,] 0.174825175 0.047952048 0.160839161
#>  [435,] 0.490509491 0.538461538 0.547452547
#>  [436,] 0.748251748 0.346653347 0.109890110
#>  [437,] 0.254745255 0.429570430 0.297702298
#>  [438,] 0.466533467 0.440559441 0.620379620
#>  [439,] 0.099900100 0.091908092 0.115884116
#>  [440,] 0.563436563 0.493506494 0.561438561
#>  [441,] 0.348651349 0.466533467 0.563436563
#>  [442,] 0.366633367 0.017982018 0.363636364
#>  [443,] 0.880119880 0.815184815 0.922077922
#>  [444,] 0.924075924 0.829170829 0.711288711
#>  [445,] 0.578421578 0.701298701 0.530469530
#>  [446,] 0.287712288 0.577422577 0.559440559
#>  [447,] 0.110889111 0.170829171 0.562437562
#>  [448,] 0.603396603 0.753246753 0.443556444
#>  [449,] 0.808191808 0.442557443 0.868131868
#>  [450,] 0.171828172 0.174825175 0.098901099
#>  [451,] 0.032967033 0.052947053 0.180819181
#>  [452,] 0.036963037 0.281718282 0.041958042
#>  [453,] 0.861138861 0.783216783 0.839160839
#>  [454,] 0.780219780 0.672327672 0.786213786
#>  [455,] 0.568431568 0.585414585 0.574425574
#>  [456,] 0.096903097 0.125874126 0.044955045
#>  [457,] 0.478521479 0.508491508 0.733266733
#>  [458,] 0.261738262 0.323676324 0.732267732
#>  [459,] 0.862137862 0.967599967 0.810189810
#>  [460,] 0.815184815 0.481518482 0.595404595
#>  [461,] 0.876123876 0.315684316 0.664335664
#>  [462,] 0.004995005 0.122877123 0.085914086
#>  [463,] 0.859140859 0.902097902 0.669330669
#>  [464,] 0.339660340 0.086913087 0.342657343
#>  [465,] 0.983438924 0.968678883 0.867132867
#>  [466,] 0.966212349 0.795204795 0.798201798
#>  [467,] 0.691308691 0.825174825 0.523476523
#>  [468,] 0.305694306 0.119880120 0.015984016
#>  [469,] 0.271728272 0.396603397 0.343656344
#>  [470,] 0.551448551 0.816183816 0.275724276
#>  [471,] 0.980647697 0.983439353 0.985101985
#>  [472,] 0.139860140 0.375624376 0.152847153
#>  [473,] 0.570429570 0.961470038 0.709290709
#>  [474,] 0.802197802 0.806193806 0.731268731
#>  [475,] 0.676323676 0.655344655 0.628371628
#>  [476,] 0.057942058 0.250749251 0.234765235
#>  [477,] 0.943056943 0.658341658 0.609390609
#>  [478,] 0.368631369 0.056943057 0.036963037
#>  [479,] 0.995289300 0.997822758 0.994453056
#>  [480,] 0.920079920 0.834165834 0.851148851
#>  [481,] 0.396603397 0.514485514 0.042957043
#>  [482,] 0.053946054 0.288711289 0.306693307
#>  [483,] 0.500499500 0.171828172 0.578421578
#>  [484,] 0.303696304 0.479520480 0.444555445
#>  [485,] 0.525474525 0.363636364 0.439560440
#>  [486,] 0.584415584 0.185814186 0.210789211
#>  [487,] 0.340659341 0.416583417 0.301698302
#>  [488,] 0.902097902 0.567432567 0.734265734
#>  [489,] 0.806193806 0.784215784 0.742257742
#>  [490,] 0.248751249 0.550449550 0.379620380
#>  [491,] 0.722277722 0.695304695 0.685314685
#>  [492,] 0.034965035 0.006993007 0.006993007
#>  [493,] 0.418581419 0.749250749 0.927072927
#>  [494,] 0.142857143 0.140859141 0.588411588
#>  [495,] 0.295704296 0.257742258 0.757242757
#>  [496,] 0.557442557 0.562437562 0.350649351
#>  [497,] 0.919080919 0.696303696 0.699300699
#>  [498,] 0.968709524 0.944055944 0.939060939
#>  [499,] 0.350649351 0.338661339 0.564435564
#>  [500,] 0.723276723 0.945054945 0.819180819
#>  [501,] 0.948051948 0.975787249 0.956200181
#>  [502,] 0.930069930 0.978379490 0.775224775
#>  [503,] 0.713286713 0.694305694 0.772227772
#>  [504,] 0.828171828 0.716283716 0.926073926
#>  [505,] 0.894105894 0.887112887 0.656343656
#>  [506,] 0.319680320 0.334665335 0.491508492
#>  [507,] 0.841158841 0.591408591 0.902097902
#>  [508,] 0.501498501 0.247752248 0.335664336
#>  [509,] 0.390609391 0.192807193 0.538461538
#>  [510,] 0.461538462 0.134865135 0.512487512
#>  [511,] 0.865134865 0.819180819 0.880119880
#>  [512,] 0.539460539 0.673326673 0.293706294
#>  [513,] 0.214785215 0.366633367 0.546453546
#>  [514,] 0.915084915 0.746253746 0.681318681
#>  [515,] 0.895104895 0.975638144 0.936063936
#>  [516,] 0.717282717 0.851148851 0.941058941
#>  [517,] 0.605394605 0.121878122 0.032967033
#>  [518,] 0.810189810 0.754245754 0.760239760
#>  [519,] 0.595404595 0.890109890 0.381618382
#>  [520,] 0.972991893 0.990110018 0.967514622
#>  [521,] 0.463536464 0.697302697 0.716283716
#>  [522,] 0.971343119 0.965085767 0.966423813
#>  [523,] 0.026973027 0.357642358 0.204795205
#>  [524,] 0.472527473 0.853146853 0.496503497
#>  [525,] 0.572427572 0.433566434 0.477522478
#>  [526,] 0.625374625 0.892107892 0.855144855
#>  [527,] 0.999860590 0.999830378 0.999960621
#>  [528,] 0.021978022 0.848151848 0.061938062
#>  [529,] 0.995802747 0.994536958 0.997277447
#>  [530,] 0.201798202 0.161838162 0.252747253
#>  [531,] 0.435564436 0.070929071 0.007992008
#>  [532,] 0.229770230 0.104895105 0.202797203
#>  [533,] 0.787212787 0.613386613 0.533466533
#>  [534,] 0.858141858 0.830169830 0.670329670
#>  [535,] 0.754245754 0.206793207 0.663336663
#>  [536,] 0.095904096 0.158841159 0.218781219
#>  [537,] 0.942057942 0.931068931 0.937062937
#>  [538,] 0.857142857 0.724275724 0.386613387
#>  [539,] 0.916083916 0.469530470 0.521478521
#>  [540,] 0.741258741 0.371628372 0.175824176
#>  [541,] 0.097902098 0.817182817 0.172827173
#>  [542,] 0.646353646 0.506493506 0.388611389
#>  [543,] 0.793206793 0.236763237 0.103896104
#>  [544,] 0.222777223 0.258741259 0.157842158
#>  [545,] 0.962606535 0.933066933 0.933066933
#>  [546,] 0.762237762 0.984801357 0.846153846
#>  [547,] 0.684315684 0.212787213 0.389610390
#>  [548,] 0.371628372 0.196803197 0.661338661
#>  [549,] 0.972890627 0.971219568 0.983295840
#>  [550,] 0.008991009 0.541458541 0.012987013
#>  [551,] 0.554445554 0.941058941 0.657342657
#>  [552,] 0.232767233 0.307692308 0.053946054
#>  [553,] 0.734265734 0.698301698 0.778221778
#>  [554,] 0.342657343 0.201798202 0.213786214
#>  [555,] 0.219780220 0.131868132 0.286713287
#>  [556,] 0.663336663 0.180819181 0.580419580
#>  [557,] 0.228771229 0.019980020 0.265734266
#>  [558,] 0.767232767 0.663336663 0.691308691
#>  [559,] 0.049950050 0.068931069 0.026973027
#>  [560,] 0.413586414 0.373626374 0.150849151
#>  [561,] 0.811188811 0.918081918 0.717282717
#>  [562,] 0.355644356 0.619380619 0.080919081
#>  [563,] 0.846153846 0.637362637 0.667332667
#>  [564,] 0.640359640 0.576423576 0.761238761
#>  [565,] 0.619380619 0.367632368 0.693306693
#>  [566,] 0.918081918 0.930069930 0.940059940
#>  [567,] 0.707292707 0.547452547 0.345654346
#>  [568,] 0.159840160 0.270729271 0.362637363
#>  [569,] 0.863136863 0.732267732 0.918081918
#>  [570,] 0.935064935 0.905094905 0.915084915
#>  [571,] 0.150849151 0.117882118 0.513486513
#>  [572,] 0.668331668 0.389610390 0.076923077
#>  [573,] 0.607392607 0.646353646 0.796203796
#>  [574,] 0.307692308 0.395604396 0.955907208
#>  [575,] 0.999115609 0.996189072 0.995967402
#>  [576,] 0.012987013 0.305694306 0.014985015
#>  [577,] 0.279720280 0.138861139 0.039960040
#>  [578,] 0.533466533 0.188811189 0.453546454
#>  [579,] 0.838161838 0.629370629 0.847152847
#>  [580,] 0.109890110 0.252747253 0.089910090
#>  [581,] 0.107892108 0.126873127 0.803196803
#>  [582,] 0.670329670 0.390609391 0.470529471
#>  [583,] 0.019980020 0.246753247 0.048951049
#>  [584,] 0.993317793 0.979017192 0.967427917
#>  [585,] 0.450549451 0.870129870 0.633366633
#>  [586,] 0.606393606 0.630369630 0.735264735
#>  [587,] 0.632367632 0.704295704 0.593406593
#>  [588,] 0.544455544 0.420579421 0.426573427
#>  [589,] 0.667332667 0.412587413 0.469530470
#>  [590,] 0.911088911 0.909090909 0.957210412
#>  [591,] 0.697302697 0.544455544 0.671328671
#>  [592,] 0.250749251 0.164835165 0.257742258
#>  [593,] 0.545454545 0.512487512 0.447552448
#>  [594,] 0.899100899 0.869130869 0.854145854
#>  [595,] 0.980664221 0.982614677 0.977584848
#>  [596,] 0.135864136 0.295704296 0.355644356
#>  [597,] 0.437562438 0.356643357 0.050949051
#>  [598,] 0.884115884 0.884115884 0.841158841
#>  [599,] 0.983446346 0.956091646 0.931068931
#>  [600,] 0.244755245 0.498501499 0.749250749
#>  [601,] 0.591408591 0.024975025 0.016983017
#>  [602,] 0.936063936 0.960367355 0.889110889
#>  [603,] 0.383616384 0.379620380 0.352647353
#>  [604,] 0.635364635 0.665334665 0.710289710
#>  [605,] 0.600399600 0.129870130 0.083916084
#>  [606,] 0.630369630 0.640359640 0.630369630
#>  [607,] 0.327672328 0.542457542 0.278721279
#>  [608,] 0.428571429 0.039960040 0.052947053
#>  [609,] 0.875124875 0.774225774 0.892107892
#>  [610,] 0.929070929 0.739260739 0.900099900
#>  [611,] 0.516483516 0.548451548 0.348651349
#>  [612,] 0.821178821 0.581418581 0.623376623
#>  [613,] 0.498501499 0.653346653 0.549450549
#>  [614,] 0.317682318 0.277722278 0.836163836
#>  [615,] 0.129870130 0.285714286 0.438561439
#>  [616,] 0.093906094 0.461538462 0.118881119
#>  [617,] 0.540459540 0.421578422 0.490509491
#>  [618,] 0.849150849 0.625374625 0.570429570
#>  [619,] 0.526473526 0.358641359 0.190809191
#>  [620,] 0.294705295 0.202797203 0.406593407
#>  [621,] 0.982230992 0.968094887 0.977334705
#>  [622,] 0.315684316 0.301698302 0.408591409
#>  [623,] 0.464535465 0.495504496 0.613386613
#>  [624,] 0.565434565 0.674325674 0.747252747
#>  [625,] 0.017982018 0.007992008 0.391608392
#>  [626,] 0.589410589 0.118881119 0.149850150
#>  [627,] 0.149850150 0.043956044 0.155844156
#>  [628,] 0.629370629 0.693306693 0.707292707
#>  [629,] 0.527472527 0.248751249 0.626373626
#>  [630,] 0.204795205 0.169830170 0.168831169
#>  [631,] 0.247752248 0.372627373 0.135864136
#>  [632,] 0.395604396 0.624375624 0.486513487
#>  [633,] 0.954030300 0.964039905 0.947052947
#>  [634,] 0.831168831 0.942057942 0.845154845
#>  [635,] 0.765234765 0.529470529 0.802197802
#>  [636,] 0.103896104 0.094905095 0.458541459
#>  [637,] 0.778221778 0.621378621 0.777222777
#>  [638,] 0.987030870 0.993661336 0.986624878
#>  [639,] 0.519480519 0.464535465 0.708291708
#>  [640,] 0.154845155 0.296703297 0.136863137
#>  [641,] 0.690309690 0.218781219 0.881118881
#>  [642,] 0.308691309 0.264735265 0.883116883
#>  [643,] 0.451548452 0.489510490 0.625374625
#>  [644,] 0.014985015 0.112887113 0.005994006
#>  [645,] 0.976078812 0.915084915 0.833166833
#>  [646,] 0.642357642 0.740259740 0.566433566
#>  [647,] 0.127872128 0.359640360 0.214785215
#>  [648,] 0.799200799 0.810189810 0.945054945
#>  [649,] 0.518481518 0.332667333 0.488511489
#>  [650,] 0.217782218 0.447552448 0.067932068
#>  [651,] 0.394605395 0.173826174 0.527472527
#>  [652,] 0.284715285 0.299700300 0.227772228
#>  [653,] 0.844155844 0.814185814 0.636363636
#>  [654,] 0.825174825 0.858141858 0.743256743
#>  [655,] 0.866133866 0.868131868 0.404595405
#>  [656,] 0.456543457 0.911088911 0.615384615
#>  [657,] 0.387612388 0.245754246 0.147852148
#>  [658,] 0.130869131 0.422577423 0.117882118
#>  [659,] 0.549450549 0.292707293 0.105894106
#>  [660,] 0.887112887 0.920079920 0.882117882
#>  [661,] 0.770229770 0.142857143 0.509490509
#>  [662,] 0.783216783 0.536463536 0.888111888
#>  [663,] 0.024975025 0.082917083 0.382617383
#>  [664,] 0.671328671 0.418581419 0.649350649
#>  [665,] 0.488511489 0.772227772 0.466533467
#>  [666,] 0.067932068 0.113886114 0.185814186
#>  [667,] 0.487512488 0.570429570 0.467532468
#>  [668,] 0.147852148 0.587412587 0.332667333
#>  [669,] 0.169830170 0.178821179 0.055944056
#>  [670,] 0.538461538 0.020979021 0.047952048
#>  [671,] 0.590409590 0.748251748 0.120879121
#>  [672,] 0.801198801 0.793206793 0.816183816
#>  [673,] 0.321678322 0.351648352 0.378621379
#>  [674,] 0.822177822 0.896103896 0.910089910
#>  [675,] 0.364635365 0.465534466 0.441558442
#>  [676,] 0.730269730 0.770229770 0.700299700
#>  [677,] 0.651348651 0.728271728 0.515484515
#>  [678,] 0.454545455 0.224775225 0.106893107
#>  [679,] 0.536463536 0.146853147 0.500499500
#>  [680,] 0.968824313 0.946053946 0.986426420
#>  [681,] 0.224775225 0.079920080 0.504495504
#>  [682,] 0.744255744 0.715284715 0.575424575
#>  [683,] 0.038961039 0.488511489 0.285714286
#>  [684,] 0.747252747 0.734265734 0.573426573
#>  [685,] 0.495504496 0.603396603 0.759240759
#>  [686,] 0.042957043 0.276723277 0.093906094
#>  [687,] 0.724275724 0.790209790 0.688311688
#>  [688,] 0.604395604 0.503496503 0.617382617
#>  [689,] 0.944055944 0.948051948 0.995783669
#>  [690,] 0.080919081 0.312687313 0.121878122
#>  [691,] 0.502497502 0.638361638 0.680319680
#>  [692,] 0.085914086 0.219780220 0.113886114
#>  [693,] 0.077922078 0.015984016 0.203796204
#>  [694,] 0.073926074 0.284715285 0.193806194
#>  [695,] 0.580419580 0.398601399 0.418581419
#>  [696,] 0.470529471 0.368631369 0.455544456
#>  [697,] 0.696303696 0.520479520 0.706293706
#>  [698,] 0.233766234 0.767232767 0.421578422
#>  [699,] 0.753246753 0.391608392 0.694305694
#>  [700,] 0.508491508 0.289710290 0.514485514
#>  [701,] 0.286713287 0.497502498 0.677322677
#>  [702,] 0.045954046 0.050949051 0.751248751
#>  [703,] 0.015984016 0.016983017 0.068931069
#>  [704,] 0.362637363 0.448551449 0.764235764
#>  [705,] 0.216783217 0.443556444 0.156843157
#>  [706,] 0.359640360 0.370629371 0.536463536
#>  [707,] 0.699300699 0.801198801 0.660339660
#>  [708,] 0.627372627 0.725274725 0.893106893
#>  [709,] 0.369630370 0.779220779 0.324675325
#>  [710,] 0.002997003 0.290709291 0.079920080
#>  [711,] 0.432567433 0.033966034 0.070929071
#>  [712,] 0.462537463 0.298701299 0.646353646
#>  [713,] 0.370629371 0.263736264 0.253746254
#>  [714,] 0.592407592 0.968819571 0.911088911
#>  [715,] 0.794205794 0.225774226 0.359640360
#>  [716,] 0.476523477 0.136863137 0.383616384
#>  [717,] 0.074925075 0.089910090 0.099900100
#>  [718,] 0.352647353 0.737262737 0.781218781
#>  [719,] 0.465534466 0.347652348 0.295704296
#>  [720,] 0.156843157 0.362637363 0.422577423
#>  [721,] 0.227772228 0.874125874 0.236763237
#>  [722,] 0.644355644 0.650349650 0.347652348
#>  [723,] 0.313686314 0.249750250 0.259740260
#>  [724,] 0.196803197 0.333666334 0.376623377
#>  [725,] 0.675324675 0.664335664 0.622377622
#>  [726,] 0.445554446 0.472527473 0.430569431
#>  [727,] 0.304695305 0.437562438 0.604395604
#>  [728,] 0.633366633 0.618381618 0.277722278
#>  [729,] 0.164835165 0.153846154 0.049950050
#>  [730,] 0.408591409 0.431568432 0.874125874
#>  [731,] 0.192807193 0.234765235 0.715284715
#>  [732,] 0.766233766 0.551448551 0.659340659
#>  [733,] 0.925074925 0.913086913 0.901098901
#>  [734,] 0.615384615 0.279720280 0.417582418
#>  [735,] 0.818181818 0.721278721 0.608391608
#>  [736,] 0.322677323 0.438561439 0.511488511
#>  [737,] 0.513486513 0.789210789 0.369630370
#>  [738,] 0.468531469 0.294705295 0.966765504
#>  [739,] 0.998366327 0.998373851 0.984820996
#>  [740,] 0.318681319 0.535464535 0.643356643
#>  [741,] 0.761238761 0.684315684 0.862137862
#>  [742,] 0.343656344 0.634365634 0.367632368
#>  [743,] 0.700299700 0.400599401 0.785214785
#>  [744,] 0.477522478 0.291708292 0.746253746
#>  [745,] 0.654345654 0.232767233 0.219780220
#>  [746,] 0.334665335 0.026973027 0.251748252
#>  [747,] 0.337662338 0.476523477 0.100899101
#>  [748,] 0.434565435 0.735264735 0.582417582
#>  [749,] 0.597402597 0.846153846 0.654345654
#>  [750,] 0.237762238 0.204795205 0.498501499
#>  [751,] 0.176823177 0.405594406 0.163836164
#>  [752,] 0.586413586 0.413586414 0.402597403
#>  [753,] 0.509490509 0.444555445 0.814185814
#>  [754,] 0.727272727 0.680319680 0.384615385
#>  [755,] 0.288711289 0.462537463 0.239760240
#>  [756,] 0.140859141 0.530469530 0.096903097
#>  [757,] 0.803196803 0.436563437 0.476523477
#>  [758,] 0.827172827 0.628371628 0.696303696
#>  [759,] 0.593406593 0.762237762 0.189810190
#>  [760,] 0.415584416 0.540459540 0.674325674
#>  [761,] 0.163836164 0.501498501 0.325674326
#>  [762,] 0.683316683 0.733266733 0.770229770
#>  [763,] 0.114885115 0.382617383 0.126873127
#>  [764,] 0.864135864 0.856143856 0.683316683
#>  [765,] 0.439560440 0.166833167 0.108891109
#>  [766,] 0.007992008 0.198801199 0.003996004
#>  [767,] 0.826173826 0.910089910 0.904095904
#>  [768,] 0.062937063 0.093906094 0.292707293
#>  [769,] 0.977007010 0.972252093 0.986436686
#>  [770,] 0.112887113 0.077922078 0.243756244
#>  [771,] 0.257742258 0.268731269 0.639360639
#>  [772,] 0.449550450 0.641358641 0.485514486
#>  [773,] 0.236763237 0.251748252 0.312687313
#>  [774,] 0.561438561 0.620379620 0.482517483
#>  [775,] 0.743256743 0.555444555 0.541458541
#>  [776,] 0.043956044 0.160839161 0.502497502
#>  [777,] 0.260739261 0.516483516 0.795204795
#>  [778,] 0.397602398 0.231768232 0.143856144
#>  [779,] 0.473526474 0.842157842 0.658341658
#>  [780,] 0.016983017 0.009990010 0.191808192
#>  [781,] 0.832167832 0.537462537 0.522477522
#>  [782,] 0.886113886 0.827172827 0.849150849
#>  [783,] 0.070929071 0.071928072 0.158841159
#>  [784,] 0.608391608 0.764235764 0.457542458
#>  [785,] 0.040959041 0.027972028 0.209790210
#>  [786,] 0.115884116 0.105894106 0.334665335
#>  [787,] 0.763236763 0.679320679 0.519480519
#>  [788,] 0.614385614 0.157842158 0.114885115
#>  [789,] 0.011988012 0.331668332 0.366633367
#>  [790,] 0.489510490 0.054945055 0.127872128
#>  [791,] 0.910089910 0.935064935 0.980250729
#>  [792,] 0.316683317 0.306693307 0.537462537
#>  [793,] 0.952155574 0.940059940 0.963025696
#>  [794,] 0.814185814 0.666333666 0.558441558
#>  [795,] 0.197802198 0.114885115 0.298701299
#>  [796,] 0.357642358 0.237762238 0.400599401
#>  [797,] 0.056943057 0.048951049 0.357642358
#>  [798,] 0.117882118 0.235764236 0.137862138
#>  [799,] 0.300699301 0.255744256 0.077922078
#>  [800,] 0.559440559 0.195804196 0.256743257
#>  [801,] 0.206793207 0.596403596 0.621378621
#>  [802,] 0.275724276 0.515484515 0.230769231
#>  [803,] 0.338661339 0.141858142 0.247752248
#>  [804,] 0.066933067 0.021978022 0.023976024
#>  [805,] 0.790209790 0.859140859 0.984687718
#>  [806,] 0.649350649 0.691308691 0.768231768
#>  [807,] 0.771228771 0.524475524 0.686313686
#>  [808,] 0.331668332 0.711288711 0.289710290
#>  [809,] 0.075924076 0.057942058 0.054945055
#>  [810,] 0.324675325 0.522477522 0.565434565
#>  [811,] 0.122877123 0.349650350 0.090909091
#>  [812,] 0.891108891 0.700299700 0.413586414
#>  [813,] 0.467532468 0.812187812 0.238761239
#>  [814,] 0.564435564 0.726273726 0.208791209
#>  [815,] 0.893106893 0.588411588 0.813186813
#>  [816,] 0.704295704 0.452547453 0.501498501
#>  [817,] 0.824175824 0.712287712 0.807192807
#>  [818,] 0.001998002 0.614385614 0.046953047
#>  [819,] 0.446553447 0.487512488 0.226773227
#>  [820,] 0.399600400 0.751248751 0.591408591
#>  [821,] 0.215784216 0.081918082 0.233766234
#>  [822,] 0.797202797 0.855144855 0.675324675
#>  [823,] 0.558441558 0.135864136 0.616383616
#>  [824,] 0.124875125 0.139860140 0.075924076
#>  [825,] 0.299700300 0.595404595 0.673326673
#>  [826,] 0.098901099 0.205794206 0.062937063
#>  [827,] 0.094905095 0.100899101 0.396603397
#>  [828,] 0.076923077 0.259740260 0.392607393
#>  [829,] 0.908091908 0.928071928 0.898101898
#>  [830,] 0.529470529 0.474525475 0.468531469
#>  [831,] 0.643356643 0.594405594 0.387612388
#>  [832,] 0.178821179 0.361638362 0.360639361
#>  [833,] 0.621378621 0.496503497 0.479520480
#>  [834,] 0.882117882 0.960870309 0.899100899
#>  [835,] 0.425574426 0.744255744 0.356643357
#>  [836,] 0.856143856 0.786213786 0.524475524
#>  [837,] 0.933066933 0.882117882 0.875124875
#>  [838,] 0.360639361 0.175824176 0.771228771
#>  [839,] 0.281718282 0.455544456 0.480519481
#>  [840,] 0.923076923 0.866133866 0.869130869
#>  [841,] 0.839160839 0.850149850 0.938061938
#>  [842,] 0.166833167 0.381618382 0.198801199
#>  [843,] 0.191808192 0.260739261 0.271728272
#>  [844,] 0.758241758 0.647352647 0.800199800
#>  [845,] 0.601398601 0.651348651 0.723276723
#>  [846,] 0.759240759 0.525474525 0.811188811
#>  [847,] 0.054945055 0.073926074 0.081918082
#>  [848,] 0.512487512 0.130869131 0.060939061
#>  [849,] 0.698301698 0.838161838 0.884115884
#>  [850,] 0.981901271 0.937062937 0.897102897
#>  [851,] 0.804195804 0.780219780 0.894105894
#>  [852,] 0.960497852 0.883116883 0.906093906
#>  [853,] 0.773226773 0.483516484 0.248751249
#>  [854,] 0.658341658 0.518481518 0.587412587
#>  [855,] 0.877122877 0.916083916 0.864135864
#>  [856,] 0.083916084 0.088911089 0.084915085
#>  [857,] 0.969354570 0.989157600 0.985799054
#>  [858,] 0.059940060 0.271728272 0.139860140
#>  [859,] 0.963726090 0.978495859 0.991125893
#>  [860,] 0.694305694 0.523476523 0.914085914
#>  [861,] 0.218781219 0.559440559 0.750249750
#>  [862,] 0.158841159 0.327672328 0.483516484
#>  [863,] 0.775224775 0.473526474 0.863136863
#>  [864,] 0.240759241 0.917082917 0.178821179
#>  [865,] 0.786213786 0.729270729 0.365634366
#>  [866,] 0.195804196 0.090909091 0.497502498
#>  [867,] 0.221778222 0.427572428 0.581418581
#>  [868,] 0.081918082 0.471528472 0.130869131
#>  [869,] 0.522477522 0.623376623 0.188811189
#>  [870,] 0.687312687 0.582417582 0.532467532
#>  [871,] 0.126873127 0.449550450 0.270729271
#>  [872,] 0.282717283 0.360639361 0.506493506
#>  [873,] 0.737262737 0.575424575 0.333666334
#>  [874,] 0.091908092 0.309690310 0.612387612
#>  [875,] 0.048951049 0.037962038 0.464535465
#>  [876,] 0.167832168 0.275724276 0.433566434
#>  [877,] 0.514485514 0.269730270 0.142857143
#>  [878,] 0.577422577 0.788211788 0.885114885
#>  [879,] 0.879120879 0.709290709 0.642357642
#>  [880,] 0.234765235 0.069930070 0.215784216
#>  [881,] 0.883116883 0.615384615 0.632367632
#>  [882,] 0.297702298 0.369630370 0.398601399
#>  [883,] 0.764235764 0.822177822 0.740259740
#>  [884,] 0.885114885 0.797202797 0.783216783
#>  [885,] 0.189810190 0.552447552 0.618381618
#>  [886,] 0.430569431 0.337662338 0.321678322
#>  [887,] 0.673326673 0.193806194 0.752247752
#>  [888,] 0.686313686 0.796203796 0.748251748
#>  [889,] 0.356643357 0.345654346 0.597402597
#>  [890,] 0.768231768 0.802197802 0.934065934
#>  [891,] 0.792207792 0.123876124 0.200799201
#>  [892,] 0.089910090 0.197802198 0.645354645
#>  [893,] 0.868131868 0.880119880 0.850149850
#>  [894,] 0.469530470 0.463536464 0.634365634
#>  [895,] 0.907092907 0.836163836 0.852147852
#>  [896,] 0.471528472 0.401598402 0.154845155
#>  [897,] 0.980944198 0.953761830 0.929070929
#>  [898,] 0.027972028 0.344655345 0.197802198
#>  [899,] 0.945054945 0.943056943 0.925074925
#>  [900,] 0.622377622 0.176823177 0.351648352
#>  [901,] 0.976253670 0.906093906 0.891108891
#>  [902,] 0.223776224 0.172827173 0.159840160
#>  [903,] 0.913086913 0.791208791 0.826173826
#>  [904,] 0.213786214 0.190809191 0.585414585
#>  [905,] 0.823176823 0.922077922 0.830169830
#>  [906,] 0.813186813 0.669330669 0.598401598
#>  [907,] 0.805194805 0.533466533 0.984286076
#>  [908,] 0.153846154 0.609390609 0.552447552
#>  [909,] 0.030969031 0.035964036 0.022977023
#>  [910,] 0.711288711 0.854145854 0.806193806
#>  [911,] 0.459540460 0.083916084 0.576423576
#>  [912,] 0.069930070 0.165834166 0.316683317
#>  [913,] 0.402597403 0.388611389 0.416583417
#>  [914,] 0.745254745 0.468531469 0.487512488
#>  [915,] 0.957063017 0.950756806 0.982403619
#>  [916,] 0.990001713 0.973032364 0.988218079
#>  [917,] 0.037962038 0.308691309 0.235764236
#>  [918,] 0.664335664 0.460539461 0.448551449
#>  [919,] 0.087912088 0.671328671 0.427572428
#>  [920,] 0.842157842 0.886113886 0.924075924
#>  [921,] 0.714285714 0.392607393 0.653346653
#>  [922,] 0.914085914 0.703296703 0.848151848
#>  [923,] 0.776223776 0.713286713 0.390609391
#>  [924,] 0.361638362 0.335664336 0.207792208
#>  [925,] 0.082917083 0.096903097 0.027972028
#>  [926,] 0.031968032 0.099900100 0.045954046
#>  [927,] 0.523476523 0.350649351 0.268731269
#>  [928,] 0.617382617 0.861138861 0.745254745
#>  [929,] 0.941058941 0.923076923 0.948051948
#>  [930,] 0.170829171 0.319680320 0.375624376
#>  [931,] 0.106893107 0.108891109 0.104895105
#>  [932,] 0.852147852 0.912087912 0.668331668
#>  [933,] 0.267732268 0.560439560 0.403596404
#>  [934,] 0.378621379 0.683316683 0.338661339
#>  [935,] 0.985262848 0.982596054 0.987102084
#>  [936,] 0.358641359 0.808191808 0.368631369
#>  [937,] 0.553446553 0.343656344 0.784215784
#>  [938,] 0.889110889 0.723276723 0.959308510
#>  [939,] 0.314685315 0.127872128 0.818181818
#>  [940,] 0.720279720 0.622377622 0.586413586
#>  [941,] 0.716283716 0.752247752 0.665334665
#>  [942,] 0.423576424 0.167832168 0.131868132
#>  [943,] 0.398601399 0.110889111 0.339660340
#>  [944,] 0.145854146 0.095904096 0.011988012
#>  [945,] 0.638361638 0.841158841 0.603396603
#>  [946,] 0.202797203 0.531468531 0.594405594
#>  [947,] 0.777222777 0.644355644 0.858141858
#>  [948,] 0.896103896 0.844155844 0.920079920
#>  [949,] 0.326673327 0.304695305 0.494505495
#>  [950,] 0.183816184 0.507492507 0.078921079
#>  [951,] 0.023976024 0.059940060 0.071928072
#>  [952,] 0.455544456 0.383616384 0.283716284
#>  [953,] 0.417582418 0.527472527 0.569430569
#>  [954,] 0.732267732 0.642357642 0.687312687
#>  [955,] 0.090909091 0.078921079 0.211788212
#>  [956,] 0.406593407 0.151848152 0.793206793
#>  [957,] 0.259740260 0.293706294 0.123876124
#>  [958,] 0.751248751 0.706293706 0.822177822
#>  [959,] 0.393606394 0.426573427 0.385614386
#>  [960,] 0.556443556 0.273726274 0.571428571
#>  [961,] 0.546453546 0.274725275 0.478521479
#>  [962,] 0.379620380 0.320679321 0.353646354
#>  [963,] 0.666333666 0.328671329 0.309690310
#>  [964,] 0.251748252 0.707292707 0.294705295
#>  [965,] 0.788211788 0.926073926 0.907092907
#>  [966,] 0.391608392 0.353646354 0.260739261
#>  [967,] 0.320679321 0.049950050 0.037962038
#>  [968,] 0.188811189 0.893106893 0.091908092
#>  [969,] 0.628371628 0.878121878 0.460539461
#>  [970,] 0.330669331 0.013986014 0.704295704
#>  [971,] 0.041958042 0.028971029 0.134865135
#>  [972,] 0.492507493 0.490509491 0.648351648
#>  [973,] 0.101898102 0.067932068 0.284715285
#>  [974,] 0.898101898 0.730269730 0.738261738
#>  [975,] 0.104895105 0.060939061 0.518481518
#>  [976,] 0.653346653 0.643356643 0.331668332
#>  [977,] 0.678321678 0.348651349 0.290709291
#>  [978,] 0.693306693 0.149850150 0.314685315
#>  [979,] 0.618381618 0.768231768 0.823176823
#>  [980,] 0.571428571 0.428571429 0.601398601
#>  [981,] 0.719280719 0.891108891 0.774225774
#>  [982,] 0.252747253 0.116883117 0.220779221
#>  [983,] 0.655344655 0.458541459 0.695304695
#>  [984,] 0.712287712 0.745254745 0.592407592
#>  [985,] 0.148851149 0.030969031 0.452547453
#>  [986,] 0.128871129 0.109890110 0.088911089
#>  [987,] 0.819180819 0.686313686 0.876123876
#>  [988,] 0.136863137 0.029970030 0.358641359
#>  [989,] 0.881118881 0.862137862 0.853146853
#>  [990,] 0.701298701 0.565434565 0.801198801
#>  [991,] 0.427572428 0.502497502 0.287712288
#>  [992,] 0.641358641 0.513486513 0.320679321
#>  [993,] 0.499500500 0.478521479 0.728271728
#>  [994,] 0.088911089 0.061938062 0.025974026
#>  [995,] 0.870129870 0.925074925 0.860139860
#>  [996,] 0.243756244 0.242757243 0.401598402
#>  [997,] 0.964153937 0.988429274 0.980902196
#>  [998,] 0.000999001 0.107892108 0.181818182
#>  [999,] 0.599400599 0.406593407 0.040959041
#> [1000,] 0.336663337 0.639360639 0.177822178