Ramos and Ledford test of independence

The Ramos and Ledford (2005) score test of independence is a modification of tests by Tawn (1988) and Ledford and Tawn (1996) for a logistic model parameter \(\alpha=1\); the latter two have scores with zero expectation, but the variance of the score are infinite, which produces non-regularity and yield test, once suitably normalized, that converge slowly to their asymptotic null distribution. The test, designed for bivariate samples, transforms observations to have unit Frechet margins and considers a bivariate censored likelihood approach for the logistic distribution.

Usage

test.scoreindep(xdat, p, test = c("ledford", "tawn"))

Arguments

xdat

a n by 2 matrix of observations

p

probability level for the marginal threshold

test

string; if tawn, only censor observations in the upper quadrant when both variables are large as in Tawn (1988), otherwise censor marginally for ledford as in Ledford and Tawn (1996).

Value

a list with elements

stat

value of the score test statistic

pval

asymptotic p-value

test

test argument

Examples

samp <- rmev(n = 1000, d = 2,
    param = 0.99, model = "log")
(test.scoreindep(samp, p = 0.9))
#> Score test of independence of Ramos and Ledford (2005)
#> Censoring scheme of Ledford and Tawn (1996) 
#> Test statistic: 2.16 
#> P-value: 0.031