Generalized quantile threshold selection

The methodology proposed by Beirlant, Vynckier and Teugels (1996) uses an asymptotic expansion of the mean squared error for Hill's estimator given a random sample with Pareto tails and positive shape, using an exponential regression. The value of k is selected to minimize the mean squared error given optimal weighting scheme. This depends on the order of regular variation \(\rho\), which is obtained based on the slope of the difference in Hill estimators, suitably reweighted. The iterative procedure of Beirlant et al. alternates between parameter estimation until convergence. It returns the Hill shape estimate, the number of higher order statistic, the parameter rho and estimates of the standard error of the shape and the mean squared error, based on the ultimate parameter values. Since the weights can become negative, there is no guarantee that the mean squared error estimate is positive, nor that the estimated value of \(\rho\) is nonpositive.

Usage

thselect.expgqt(
  xdat,
  maxiter = 10L,
  tol = 2,
  kmin = max(10, floor(length(xdat)/100)),
  kmax = floor(0.8 * length(xdat)),
  ...
)

Arguments

xdat

[vector] sample of exceedances

maxiter

[int] maximum number of iteration

tol

[double] tolerance for difference in value of \(k\) for the fixed point

kmin

[int] minimum number of exceedances for the estimator

kmax

[int] maximum number of exceedances for the estimator

...

additional arguments, currently ignored

Value

a list with components

shape the Hill estimator of the shape, based on the k largest order statistics

k0 number of high order statistics for estimation of the shape using Hill's estimator

rho estimate of the second order regular variation parameter

mse mean squared error estimate of the shape parameter

se standard error of the shape parameter

convergence logical; if TRUE, indicates that the method converged to a fixed point within tol before reaching the maximum number of iterations maxiter

References

Beirlant, J., Vynckier, P., & Teugels, J. L. (1996). Excess Functions and Estimation of the Extreme-Value Index. Bernoulli, 2(4), 293–318. doi:10.2307/3318416

Examples

# Simulate Pareto data - log(xdat) is exponential with rate 2
xdat <- rgp(n = 200, loc = 1, scale = 0.5, shape = 0.5)
(thselect.expgqt(xdat))
#> Threshold selection method: Beirlant, Vynckier and Teugels (1996)
#> Generalized quantile threshold selection
#> 
#> Selected threshold: 3.829 
#> Number of exceedances: 22 
#> Shape estimate: 0.001