Metric-based threshold selection

Adaptation of Varty et al.'s metric-based threshold automated diagnostic for the independent and identically distributed case with no rounding.

Usage

vmetric.diag(
  xdat,
  thresh,
  B = 199L,
  type = c("qq", "pp"),
  dist = c("l1", "l2"),
  neval = 1000L,
  ci = 0.95
)

Arguments

xdat

vector of observations

thresh

vector of thresholds

B

number of bootstrap replications

type

string indicating scale, either qq for exponential quantile-quantile plot or pp for probability-probability plot (uniform)

dist

string indicating norm, either l1 for absolute error or l2 for quadratic error

neval

number of points at which to estimate the metric. Default to 1000L

ci

level of symmetric confidence interval. Default to 0.95

Value

an invisible list with components

threshscalar threshold minimizing criterion

cthreshvector of candidate thresholds

metricvalue of the metric criterion evaluated at each threshold

typeargument type

distargument dist

Details

The algorithm proceeds by first computing the maximum likelihood algorithm and then simulating datasets from replication with parameters drawn from a bivariate normal approximation to the maximum likelihood estimator distribution.

For each bootstrap sample, we refit the model and convert the quantiles to exponential or uniform variates. The mean absolute or mean squared distance is calculated on these. The threshold returned is the one with the lowest value of the metric.

References

Varty, Z. and J.A. Tawn and P.M. Atkinson and S. Bierman (2021+), Inference for extreme earthquake magnitudes accounting for a time-varying measurement process