Coefficient of extremal asymmetry

This function implements estimators of the bivariate coefficient of extremal asymmetry proposed in Semadeni's (2021) PhD thesis. Two estimators are implemented: one based on empirical distributions, the second using empirical likelihood.

Usage

xasym(
  data,
  u = NULL,
  nq = 40,
  qlim = c(0.8, 0.99),
  method = c("empirical", "emplik"),
  confint = c("none", "wald", "bootstrap"),
  level = 0.95,
  B = 999L,
  ties.method = "random",
  plot = TRUE,
  ...
)

Arguments

data

an n by 2 matrix of observations

u

vector of probability levels at which to evaluate extremal asymmetry

nq

integer; number of quantiles at which to evaluate the coefficient if u is NULL

qlim

a vector of length 2 with the probability limits for the quantiles

method

string indicating the estimation method, one of empirical or empirical likelihood (emplik)

confint

string for the method used to derive confidence intervals, either none (default) or a nonparametric bootstrap

level

probability level for confidence intervals, default to 0.95 or bounds for the interval

B

integer; number of bootstrap replicates (if applicable)

ties.method

string; method for handling ties. See the documentation of rank for available options.

plot

logical; if TRUE, return a plot.

...

additional parameters for plots

Value

an invisible data frame with columns

threshold

vector of thresholds on the probability scale

coef

extremal asymmetry coefficient estimates

confint

either NULL or a matrix with two columns containing the lower and upper bounds for each threshold

Details

Let U, V be uniform random variables and define the partial extremal dependence coefficients \(\varphi_{+}(u) = \Pr(V > U | U > u, V > u)\), \(\varphi_{-}(u) = \Pr(V < U | U > u, V > u)\) and \(\varphi_0(u) = \Pr(V = U | U > u, V > u)\) Define $$ \varphi(u) = \frac{\varphi_{+} - \varphi_{-}}{\varphi_{+} + \varphi_{-}}$$

The empirical likelihood estimator, derived for max-stable vectors with unit Frechet margins, is $$\frac{\sum_i p_i I(w_i \leq 0.5) - 0.5}{0.5 - 2\sum_i p_i(0.5-w_i) I(w_i \leq 0.5)}$$ where \(p_i\) is the empirical likelihood weight for observation \(i\) and \(w_i\) is the pseudo-angle associated to the first coordinate.

References

Semadeni, C. (2020). Inference on the Angular Distribution of Extremes, PhD thesis, EPFL, no. 8168.

Examples

if (FALSE) { # \dontrun{
samp <- rmev(n = 1000,
             d = 2,
             param = 0.2,
             model = "log")
xasym(samp, confint = "wald")
xasym(samp, method = "emplik")
} # }