Bivariate angular function for extrapolation based on rays

The scale parameter \(g(w)\) in the Ledford and Tawn approach is estimated empirically for \(x\) large as $$\frac{\Pr(X_P>xw, Y_P>x(1-w))}{\Pr(X_P>x, Y_P>x)}$$ where the sample (\(X_P, Y_P\)) are observations on a common unit Pareto scale. The coefficient \(\eta\) is estimated using maximum likelihood as the shape parameter of a generalized Pareto distribution on \(\min(X_P, Y_P)\).

Usage

angextrapo(dat, qu = 0.95, w = seq(0.05, 0.95, length = 20))

Arguments

dat

an \(n\) by \(2\) matrix of multivariate observations

qu

quantile level on uniform scale at which to threshold data. Default to 0.95

w

vector of unique angles between 0 and 1 at which to evaluate scale empirically.

Value

a list with elements

w: angles between zero and one

g: scale function at a given value of w

eta: Ledford and Tawn tail dependence coefficient

References

Ledford, A.W. and J. A. Tawn (1996), Statistics for near independence in multivariate extreme values. Biometrika, 83(1), 169--187.

Examples

angextrapo(rmev(n = 1000, model = 'log', d = 2, param = 0.5))
#> $w
#>  [1] 0.05000000 0.09736842 0.14473684 0.19210526 0.23947368 0.28684211
#>  [7] 0.33421053 0.38157895 0.42894737 0.47631579 0.52368421 0.57105263
#> [13] 0.61842105 0.66578947 0.71315789 0.76052632 0.80789474 0.85526316
#> [19] 0.90263158 0.95000000
#> 
#> $g
#>  [1] 30.941176 16.735294 11.852941  9.441176  8.058824  7.176471  6.588235
#>  [8]  6.205882  5.970588  5.882353  5.882353  5.970588  6.205882  6.588235
#> [15]  7.176471  8.058824  9.441176 11.852941 16.735294 30.941176
#> 
#> $eta
#>    shape 
#> 1.259229 
#>