Simulation from R-Pareto processes

Simulation from R-Pareto processes

Usage

rparp(
  n,
  shape = 1,
  riskf = c("sum", "site", "max", "min", "l2"),
  siteindex = NULL,
  d,
  param,
  sigma,
  model = c("log", "neglog", "bilog", "negbilog", "hr", "br", "xstud", "smith",
    "schlather", "ct", "sdir", "dirmix"),
  weights,
  vario,
  coord = NULL,
  ...
)

Arguments

n

number of observations

shape

shape tail index of Pareto variable

riskf

string indicating risk functional.

siteindex

integer between 1 and d specifying the index of the site or variable

d

dimension of sample

param

parameter vector for the logistic, bilogistic, negative bilogistic and extremal Dirichlet (Coles and Tawn) model. Parameter matrix for the Dirichlet mixture. Degree of freedoms for extremal student model. See Details.

sigma

covariance matrix for Brown-Resnick and extremal Student-t distributions. Symmetric matrix of squared coefficients \(\lambda^2\) for the Husler-Reiss model, with zero diagonal elements.

model

for multivariate extreme value distributions, users can choose between 1-parameter logistic and negative logistic, asymmetric logistic and negative logistic, bilogistic, Husler-Reiss, extremal Dirichlet model (Coles and Tawn) or the Dirichlet mixture. Spatial models include the Brown-Resnick, Smith, Schlather and extremal Student max-stable processes.

weights

vector of length m for the m mixture components. Must sum to one

vario

semivariogram function whose first argument must be distance. Used only if provided in conjunction with coord and if sigma is missing

coord

d by k matrix of coordinates, used as input in the variogram vario or as parameter for the Smith model. If grid is TRUE, unique entries should be supplied.

...

additional arguments for the vario function

Value

an n by d sample from the R-Pareto process, with attributesaccept.rate if the procedure uses rejection sampling.

Details

For riskf=max and riskf=min, the procedure uses rejection sampling based on Pareto variates sampled from sum and may be slow if d is large.

Examples

rparp(n=10, riskf='site', siteindex=2, d=3, param=2.5, model='log')
#>            [,1]     [,2]      [,3]
#>  [1,] 3.0705820 6.898690 1.7811997
#>  [2,] 0.4187338 1.050321 0.7206781
#>  [3,] 3.1142322 1.941442 2.1663294
#>  [4,] 0.3416151 1.108762 0.2811433
#>  [5,] 1.7256173 1.389237 1.9690732
#>  [6,] 0.5507356 2.915300 0.3831337
#>  [7,] 1.2361560 1.380551 1.7378134
#>  [8,] 0.6217867 1.160798 0.4732747
#>  [9,] 1.5450479 2.761058 0.9203676
#> [10,] 5.0421507 3.319403 3.7378095
rparp(n=10, riskf='min', d=3, param=2.5, model='neglog')
#>            [,1]      [,2]      [,3]
#>  [1,] 10.011630 29.487136 19.577961
#>  [2,]  3.172307  2.602747  4.975430
#>  [3,]  2.918126  1.868746  2.571869
#>  [4,]  4.640023  3.856793  3.211399
#>  [5,]  5.145384  3.907979  2.332019
#>  [6,]  3.247898  1.971408  4.345620
#>  [7,]  2.159573  1.981640  1.754810
#>  [8,]  2.534616  2.494687  2.483782
#>  [9,]  1.051708  3.615180  1.749199
#> [10,]  5.819990  5.819990  5.819990
#> attr(,"accept.rate")
#> [1] 0.07092199
rparp(n=10, riskf='max', d=4, param=c(0.2,0.1,0.9,0.5), model='bilog')
#>               [,1]         [,2]      [,3]         [,4]
#>  [1,] 1.474630e+00 1.364460e+00 0.5003000 5.692104e-01
#>  [2,] 5.444600e-04 5.128759e-04 4.0133464 3.074541e-04
#>  [3,] 8.559278e-09 7.479521e-09 1.3082068 2.968178e-09
#>  [4,] 4.466558e+01 3.359784e+01 9.3818442 1.635296e+01
#>  [5,] 7.841005e+00 9.020112e+00 3.8677721 6.959005e+00
#>  [6,] 1.131577e+00 1.708033e+00 0.3125240 7.168801e-01
#>  [7,] 2.266500e+00 2.259968e+00 0.2729027 1.805056e+00
#>  [8,] 1.532215e+00 2.085249e+00 1.4463610 7.123242e-01
#>  [9,] 2.991308e+00 2.243467e+00 0.4347594 3.362642e+00
#> [10,] 2.182127e-01 2.705780e-01 1.5377191 3.728419e-01
#> attr(,"accept.rate")
#> [1] 0.4347826
rparp(n=10, riskf='sum', d=3, param=c(0.8,1.2,0.6, -0.5), model='sdir')
#>               [,1]         [,2]       [,3]
#>  [1,] 1.615349e-01 9.598978e-01 0.17653771
#>  [2,] 1.367543e+00 1.568355e-01 0.05495440
#>  [3,] 1.914588e-01 6.045678e-01 0.32212026
#>  [4,] 5.297949e-02 1.039894e-01 3.21446807
#>  [5,] 2.279316e-01 8.914601e-01 0.12642634
#>  [6,] 4.114223e-01 6.751378e-01 6.00480531
#>  [7,] 5.472881e-02 1.316949e+00 0.07305779
#>  [8,] 2.281327e-01 4.518326e-01 0.50335468
#>  [9,] 1.902402e-05 3.030021e-05 1.47433006
#> [10,] 1.860653e-01 9.508976e-01 0.27203075
vario <- function(x, scale=0.5, alpha=0.8){ scale*x^alpha }
grid.coord <- as.matrix(expand.grid(runif(4), runif(4)))
rparp(n=10, riskf='max', vario=vario, coord=grid.coord, model='br')
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 1.77230385 1.49751735 1.41951154 1.64514770 1.79349915 2.39428054
#>  [2,] 0.52828634 0.54054095 1.05806836 0.70036983 0.51241013 0.52207832
#>  [3,] 1.38852338 0.39903126 0.90267882 0.78604846 1.01764167 0.98236962
#>  [4,] 1.27083799 0.78722533 1.14226771 1.26175933 1.65377885 2.02962617
#>  [5,] 0.58396778 0.43977603 0.40465543 0.49024169 1.54683668 0.68540117
#>  [6,] 0.61913646 0.89596021 1.14781125 0.89659282 0.84728682 0.71323732
#>  [7,] 0.04500564 0.06335646 0.03857993 0.05496761 0.06422292 0.05711904
#>  [8,] 0.68647600 0.35656228 0.57992953 0.52298821 1.04981929 0.31413788
#>  [9,] 0.71175395 0.54673021 0.65321060 0.50673024 0.88633740 0.89280182
#> [10,] 0.98545148 1.59081985 0.68931456 1.02403967 0.72858825 1.59425497
#>             [,7]       [,8]      [,9]     [,10]     [,11]     [,12]      [,13]
#>  [1,] 2.05830865 2.22974872 1.2345546 0.5320206 1.1104743 1.0462039 0.83199246
#>  [2,] 0.54432981 0.63312681 0.5386513 1.0587488 0.9880912 0.6763056 0.94227447
#>  [3,] 0.55713003 1.19106452 0.4870495 0.8621921 0.8745162 0.6706556 0.50128161
#>  [4,] 2.45916013 1.41319159 0.5264309 0.6228048 0.7238994 0.5108373 0.49192502
#>  [5,] 0.96260110 1.11983397 0.1258928 0.2014684 0.1746814 0.1161757 0.08968249
#>  [6,] 0.62654785 0.61049395 0.2124748 0.1610319 0.1867617 0.2579053 0.24976367
#>  [7,] 0.08241944 0.06198499 0.6450561 0.3125140 1.2609168 0.5117937 0.51085936
#>  [8,] 0.40161361 0.66822973 0.3513354 0.2345212 0.2838641 0.2984742 0.28882467
#>  [9,] 0.48419681 1.01503864 0.5043154 0.3193356 0.4598692 0.3921489 0.51707195
#> [10,] 0.71231653 1.10111619 0.8176904 1.1027317 0.6362818 0.6502851 0.63547273
#>           [,14]     [,15]     [,16]
#>  [1,] 0.7200598 0.8483509 0.7801234
#>  [2,] 0.5757242 1.0949353 0.9438266
#>  [3,] 0.5824337 0.5801588 0.5844089
#>  [4,] 0.4799863 0.3484021 0.4939109
#>  [5,] 0.1659791 0.1323298 0.1260149
#>  [6,] 0.2749009 0.3574052 0.2603804
#>  [7,] 0.1332660 0.4355390 0.2835384
#>  [8,] 0.3192293 0.2649084 0.3814494
#>  [9,] 0.3741975 0.3283437 0.6176970
#> [10,] 0.7748748 0.8127487 0.5749981
#> attr(,"accept.rate")
#> [1] 0.1941176