Simulation from Pareto processes (max) using composition sampling

The algorithm performs forward sampling by simulating first from a mixture, then sample angles conditional on them being less than one. The resulting sample from the angular distribution is then multiplied by Pareto variates with tail index shape.

Usage

rparpcs(
  n,
  Lambda = NULL,
  Sigma = NULL,
  df = NULL,
  model = c("br", "xstud"),
  riskf = c("max", "min"),
  shape = 1
)

Arguments

n

sample size.

Lambda

parameter matrix for the Brown--Resnick model. See Details.

Sigma

correlation matrix if model = 'xstud', otherwise the covariance matrix formed from the stationary Brown-Resnick process.

df

degrees of freedom for extremal Student process.

model

string indicating the model family.

riskf

string indicating the risk functional. Only max and min are currently supported.

shape

tail index of the Pareto variates (reciprocal shape parameter). Must be strictly positive.

Value

an n by d matrix of samples, where d = ncol(Sigma), with attributes

mixt.weights.

Details

Only extreme value models based on elliptical processes are handled. The Lambda matrix is formed by evaluating the semivariogram \(\gamma\) at sites \(s_i, s_j\), meaning that \(\Lambda_{i,j} = \gamma(s_i, s_j)/2\).

The argument Sigma is ignored for the Brown-Resnick model if Lambda is provided by the user.

See also

Author

Leo Belzile

Examples

if (FALSE) {
#Brown-Resnick, Wadsworth and Tawn (2014) parametrization
D <- 20L
coord <- cbind(runif(D), runif(D))
semivario <- function(d, alpha = 1.5, lambda = 1){0.5 * (d/lambda)^alpha}
Lambda <- semivario(as.matrix(dist(coord))) / 2
rparpcs(n = 10, Lambda = Lambda, model = 'br', shape = 0.1)
#Extremal Student
Sigma <- stats::rWishart(n = 1, df = 20, Sigma = diag(10))[,,1]
rparpcs(n = 10, Sigma = cov2cor(Sigma), df = 3, model = 'xstud')
}