Semi-parametric marginal transformation to uniform

The function spunif transforms a matrix or vector of data x to the pseudo-uniform scale using a semiparametric transform. Data below the threshold are transformed to pseudo-uniforms using a rank transform, while data above the threshold are assumed to follow a generalized Pareto distribution. The parameters of the latter are estimated using maximum likelihood if either scale = NULL or shape = NULL.

Usage

spunif(x, thresh, scale = NULL, shape = NULL)

Arguments

x

matrix or vector of data

thresh

vector of marginal thresholds

scale

vector of marginal scale parameters for the generalized Pareto

shape

vector of marginal shape parameters for the generalized Pareto

Value

a matrix or vector of the same dimension as x, with pseudo-uniform observations

Author

Leo Belzile

Examples

x <- rmev(1000, d = 3, param = 2, model = 'log')
thresh <- apply(x, 2, quantile, 0.95)
spunif(x, thresh)
#>                [,1]        [,2]        [,3]
#>    [1,] 0.144855145 0.182817183 0.743256743
#>    [2,] 0.533466533 0.962188034 0.646353646
#>    [3,] 0.137862138 0.534465534 0.445554446
#>    [4,] 0.670329670 0.463536464 0.481518482
#>    [5,] 0.939060939 0.956664882 0.984379518
#>    [6,] 0.693306693 0.599400599 0.686313686
#>    [7,] 0.182817183 0.274725275 0.324675325
#>    [8,] 0.614385614 0.664335664 0.619380619
#>    [9,] 0.259740260 0.137862138 0.803196803
#>   [10,] 0.132867133 0.163836164 0.192807193
#>   [11,] 0.937062937 0.951322529 0.954489408
#>   [12,] 0.513486513 0.592407592 0.478521479
#>   [13,] 0.780219780 0.931068931 0.784215784
#>   [14,] 0.232767233 0.310689311 0.453546454
#>   [15,] 0.569430569 0.567432567 0.347652348
#>   [16,] 0.195804196 0.011988012 0.342657343
#>   [17,] 0.357642358 0.236763237 0.123876124
#>   [18,] 0.738261738 0.349650350 0.395604396
#>   [19,] 0.757242757 0.395604396 0.277722278
#>   [20,] 0.745254745 0.017982018 0.355644356
#>   [21,] 0.873126873 0.372627373 0.792207792
#>   [22,] 0.029970030 0.256743257 0.368631369
#>   [23,] 0.501498501 0.128871129 0.311688312
#>   [24,] 0.211788212 0.016983017 0.240759241
#>   [25,] 0.727272727 0.501498501 0.889110889
#>   [26,] 0.636363636 0.252747253 0.065934066
#>   [27,] 0.579420579 0.761238761 0.543456543
#>   [28,] 0.053946054 0.422577423 0.162837163
#>   [29,] 0.466533467 0.311688312 0.440559441
#>   [30,] 0.475524476 0.538461538 0.731268731
#>   [31,] 0.265734266 0.355644356 0.730269730
#>   [32,] 0.840159840 0.958998575 0.804195804
#>   [33,] 0.800199800 0.509490509 0.593406593
#>   [34,] 0.851148851 0.348651349 0.666333666
#>   [35,] 0.005994006 0.143856144 0.096903097
#>   [36,] 0.839160839 0.881118881 0.675324675
#>   [37,] 0.336663337 0.101898102 0.336663337
#>   [38,] 0.981053820 0.960550788 0.854145854
#>   [39,] 0.958417712 0.786213786 0.790209790
#>   [40,] 0.678321678 0.815184815 0.531468531
#>   [41,] 0.301698302 0.139860140 0.017982018
#>   [42,] 0.272727273 0.429570430 0.338661339
#>   [43,] 0.543456543 0.805194805 0.272727273
#>   [44,] 0.977651532 0.980618045 0.982111848
#>   [45,] 0.155844156 0.405594406 0.172827173
#>   [46,] 0.556443556 0.949050949 0.714285714
#>   [47,] 0.789210789 0.795204795 0.729270729
#>   [48,] 0.665334665 0.663336663 0.628371628
#>   [49,] 0.063936064 0.276723277 0.249750250
#>   [50,] 0.930069930 0.665334665 0.602397602
#>   [51,] 0.368631369 0.061938062 0.038961039
#>   [52,] 0.994676139 0.997828189 0.993688593
#>   [53,] 0.905094905 0.820179820 0.841158841
#>   [54,] 0.392607393 0.542457542 0.047952048
#>   [55,] 0.059940060 0.315684316 0.310689311
#>   [56,] 0.497502498 0.199800200 0.573426573
#>   [57,] 0.299700300 0.507492507 0.455544456
#>   [58,] 0.520479520 0.394605395 0.450549451
#>   [59,] 0.566433566 0.213786214 0.221778222
#>   [60,] 0.337662338 0.451548452 0.305694306
#>   [61,] 0.885114885 0.595404595 0.732267732
#>   [62,] 0.793206793 0.774225774 0.738261738
#>   [63,] 0.253746254 0.577422577 0.390609391
#>   [64,] 0.706293706 0.694305694 0.693306693
#>   [65,] 0.033966034 0.005994006 0.006993007
#>   [66,] 0.412587413 0.745254745 0.926073926
#>   [67,] 0.159840160 0.157842158 0.582417582
#>   [68,] 0.292707293 0.283716284 0.760239760
#>   [69,] 0.549450549 0.590409590 0.343656344
#>   [70,] 0.904095904 0.695304695 0.705294705
#>   [71,] 0.961966322 0.927072927 0.938061938
#>   [72,] 0.340659341 0.368631369 0.553446553
#>   [73,] 0.707292707 0.928071928 0.812187812
#>   [74,] 0.936063936 0.970485931 0.950980833
#>   [75,] 0.919080919 0.973984051 0.776223776
#>   [76,] 0.697302697 0.693306693 0.772227772
#>   [77,] 0.813186813 0.718281718 0.925074925
#>   [78,] 0.870129870 0.871128871 0.656343656
#>   [79,] 0.317682318 0.364635365 0.500499500
#>   [80,] 0.822177822 0.612387612 0.903096903
#>   [81,] 0.498501499 0.271728272 0.330669331
#>   [82,] 0.385614386 0.218781219 0.541458541
#>   [83,] 0.454545455 0.151848152 0.517482517
#>   [84,] 0.845154845 0.807192807 0.868131868
#>   [85,] 0.529470529 0.676323676 0.300699301
#>   [86,] 0.224775225 0.396603397 0.546453546
#>   [87,] 0.899100899 0.742257742 0.690309690
#>   [88,] 0.875124875 0.970282691 0.933066933
#>   [89,] 0.703296703 0.833166833 0.941058941
#>   [90,] 0.591408591 0.142857143 0.035964036
#>   [91,] 0.796203796 0.752247752 0.762237762
#>   [92,] 0.576423576 0.872127872 0.393606394
#>   [93,] 0.967823409 0.988950212 0.962192283
#>   [94,] 0.457542458 0.697302697 0.717282717
#>   [95,] 0.965601185 0.955335478 0.961057881
#>   [96,] 0.026973027 0.387612388 0.217782218
#>   [97,] 0.469530470 0.836163836 0.502497502
#>   [98,] 0.560439560 0.465534466 0.479520480
#>   [99,] 0.609390609 0.874125874 0.847152847
#>  [100,] 0.999821946 0.999858021 0.999976206
#>  [101,] 0.020979021 0.829170829 0.072927073
#>  [102,] 0.995246420 0.994175604 0.997116882
#>  [103,] 0.215784216 0.180819181 0.260739261
#>  [104,] 0.424575425 0.084915085 0.007992008
#>  [105,] 0.239760240 0.114885115 0.215784216
#>  [106,] 0.771228771 0.625374625 0.535464535
#>  [107,] 0.838161838 0.817182817 0.678321678
#>  [108,] 0.734265734 0.233766234 0.665334665
#>  [109,] 0.109890110 0.177822178 0.234765235
#>  [110,] 0.929070929 0.915084915 0.936063936
#>  [111,] 0.837162837 0.723276723 0.403596404
#>  [112,] 0.900099900 0.495504496 0.526473526
#>  [113,] 0.723276723 0.402597403 0.194805195
#>  [114,] 0.112887113 0.806193806 0.190809191
#>  [115,] 0.634365634 0.536463536 0.405594406
#>  [116,] 0.778221778 0.261738262 0.117882118
#>  [117,] 0.233766234 0.284715285 0.176823177
#>  [118,] 0.953106553 0.916083916 0.931068931
#>  [119,] 0.740259740 0.982358911 0.836163836
#>  [120,] 0.671328671 0.235764236 0.406593407
#>  [121,] 0.371628372 0.223776224 0.663336663
#>  [122,] 0.967688070 0.964159708 0.979930531
#>  [123,] 0.010989011 0.569430569 0.013986014
#>  [124,] 0.545454545 0.924075924 0.658341658
#>  [125,] 0.240759241 0.339660340 0.057942058
#>  [126,] 0.714285714 0.700299700 0.778221778
#>  [127,] 0.338661339 0.229770230 0.224775225
#>  [128,] 0.230769231 0.149850150 0.289710290
#>  [129,] 0.648351648 0.207792208 0.575424575
#>  [130,] 0.238761239 0.026973027 0.266733267
#>  [131,] 0.749250749 0.667332667 0.699300699
#>  [132,] 0.057942058 0.080919081 0.028971029
#>  [133,] 0.403596404 0.404595405 0.171828172
#>  [134,] 0.797202797 0.899100899 0.720279720
#>  [135,] 0.348651349 0.631368631 0.090909091
#>  [136,] 0.829170829 0.647352647 0.671328671
#>  [137,] 0.628371628 0.598401598 0.763236763
#>  [138,] 0.603396603 0.397602398 0.700299700
#>  [139,] 0.902097902 0.914085914 0.939060939
#>  [140,] 0.692307692 0.574425574 0.339660340
#>  [141,] 0.176823177 0.297702298 0.367632368
#>  [142,] 0.842157842 0.730269730 0.914085914
#>  [143,] 0.923076923 0.882117882 0.910089910
#>  [144,] 0.168831169 0.136863137 0.518481518
#>  [145,] 0.654345654 0.419580420 0.086913087
#>  [146,] 0.593406593 0.657342657 0.789210789
#>  [147,] 0.303696304 0.427572428 0.950706344
#>  [148,] 0.998949691 0.996043679 0.995546923
#>  [149,] 0.014985015 0.335664336 0.014985015
#>  [150,] 0.279720280 0.155844156 0.043956044
#>  [151,] 0.525474525 0.215784216 0.461538462
#>  [152,] 0.820179820 0.644355644 0.837162837
#>  [153,] 0.122877123 0.278721279 0.104895105
#>  [154,] 0.119880120 0.145854146 0.796203796
#>  [155,] 0.655344655 0.421578422 0.475524476
#>  [156,] 0.019980020 0.269730270 0.052947053
#>  [157,] 0.992479103 0.974834354 0.962101683
#>  [158,] 0.439560440 0.856143856 0.634365634
#>  [159,] 0.592407592 0.645354645 0.733266733
#>  [160,] 0.620379620 0.704295704 0.590409590
#>  [161,] 0.535464535 0.453546454 0.434565435
#>  [162,] 0.653346653 0.448551449 0.474525475
#>  [163,] 0.895104895 0.886113886 0.951933247
#>  [164,] 0.684315684 0.573426573 0.680319680
#>  [165,] 0.254745255 0.185814186 0.263736264
#>  [166,] 0.537462537 0.540459540 0.456543457
#>  [167,] 0.883116883 0.855144855 0.846153846
#>  [168,] 0.977671938 0.979554580 0.973227197
#>  [169,] 0.151848152 0.327672328 0.354645355
#>  [170,] 0.426573427 0.386613387 0.054945055
#>  [171,] 0.860139860 0.865134865 0.832167832
#>  [172,] 0.981062750 0.941058941 0.929070929
#>  [173,] 0.251748252 0.526473526 0.749250749
#>  [174,] 0.572427572 0.031968032 0.019980020
#>  [175,] 0.924075924 0.947052947 0.885114885
#>  [176,] 0.382617383 0.407592408 0.346653347
#>  [177,] 0.622377622 0.669330669 0.715284715
#>  [178,] 0.585414585 0.147852148 0.092907093
#>  [179,] 0.619380619 0.650349650 0.629370629
#>  [180,] 0.324675325 0.570429570 0.276723277
#>  [181,] 0.418581419 0.049950050 0.056943057
#>  [182,] 0.850149850 0.768231768 0.890109890
#>  [183,] 0.916083916 0.737262737 0.900099900
#>  [184,] 0.512487512 0.575424575 0.341658342
#>  [185,] 0.804195804 0.601398601 0.623376623
#>  [186,] 0.493506494 0.662337662 0.549450549
#>  [187,] 0.314685315 0.306693307 0.830169830
#>  [188,] 0.142857143 0.313686314 0.448551449
#>  [189,] 0.106893107 0.488511489 0.138861139
#>  [190,] 0.530469530 0.454545455 0.499500500
#>  [191,] 0.831168831 0.639360639 0.563436563
#>  [192,] 0.521478521 0.388611389 0.207792208
#>  [193,] 0.289710290 0.230769231 0.420579421
#>  [194,] 0.979592305 0.959712069 0.972940769
#>  [195,] 0.311688312 0.332667333 0.422577423
#>  [196,] 0.460539461 0.521478521 0.608391608
#>  [197,] 0.554445554 0.678321678 0.745254745
#>  [198,] 0.018981019 0.006993007 0.408591409
#>  [199,] 0.570429570 0.138861139 0.168831169
#>  [200,] 0.166833167 0.053946054 0.174825175
#>  [201,] 0.617382617 0.692307692 0.711288711
#>  [202,] 0.522477522 0.272727273 0.627372627
#>  [203,] 0.218781219 0.196803197 0.183816184
#>  [204,] 0.252747253 0.403596404 0.158841159
#>  [205,] 0.390609391 0.637362637 0.492507493
#>  [206,] 0.940059940 0.953792576 0.946053946
#>  [207,] 0.816183816 0.925074925 0.835164835
#>  [208,] 0.746253746 0.554445554 0.794205794
#>  [209,] 0.115884116 0.108891109 0.464535465
#>  [210,] 0.763236763 0.633366633 0.777222777
#>  [211,] 0.985308695 0.993165223 0.983971103
#>  [212,] 0.517482517 0.492507493 0.713286713
#>  [213,] 0.172827173 0.328671329 0.159840160
#>  [214,] 0.676323676 0.246753247 0.869130869
#>  [215,] 0.305694306 0.291708292 0.873126873
#>  [216,] 0.442557443 0.515484515 0.626373626
#>  [217,] 0.015984016 0.122877123 0.004995005
#>  [218,] 0.971879587 0.893106893 0.827172827
#>  [219,] 0.630369630 0.738261738 0.558441558
#>  [220,] 0.140859141 0.389610390 0.225774226
#>  [221,] 0.786213786 0.798201798 0.943056943
#>  [222,] 0.514485514 0.362637363 0.496503497
#>  [223,] 0.228771229 0.474525475 0.078921079
#>  [224,] 0.389610390 0.201798202 0.533466533
#>  [225,] 0.283716284 0.331668332 0.244755245
#>  [226,] 0.828171828 0.803196803 0.636363636
#>  [227,] 0.808191808 0.844155844 0.739260739
#>  [228,] 0.846153846 0.854145854 0.419580420
#>  [229,] 0.448551449 0.890109890 0.611388611
#>  [230,] 0.384615385 0.268731269 0.166833167
#>  [231,] 0.143856144 0.455544456 0.136863137
#>  [232,] 0.541458541 0.324675325 0.120879121
#>  [233,] 0.864135864 0.901098901 0.871128871
#>  [234,] 0.751248751 0.161838162 0.515484515
#>  [235,] 0.765234765 0.565434565 0.884115884
#>  [236,] 0.024975025 0.098901099 0.397602398
#>  [237,] 0.657342657 0.452547453 0.652347652
#>  [238,] 0.483516484 0.767232767 0.470529471
#>  [239,] 0.073926074 0.123876124 0.201798202
#>  [240,] 0.482517483 0.596403596 0.471528472
#>  [241,] 0.164835165 0.609390609 0.327672328
#>  [242,] 0.188811189 0.206793207 0.063936064
#>  [243,] 0.527472527 0.027972028 0.051948052
#>  [244,] 0.571428571 0.743256743 0.144855145
#>  [245,] 0.788211788 0.783216783 0.809190809
#>  [246,] 0.319680320 0.382617383 0.388611389
#>  [247,] 0.805194805 0.876123876 0.907092907
#>  [248,] 0.365634366 0.493506494 0.452547453
#>  [249,] 0.711288711 0.766233766 0.707292707
#>  [250,] 0.637362637 0.727272727 0.520479520
#>  [251,] 0.446553447 0.249750250 0.124875125
#>  [252,] 0.526473526 0.165834166 0.507492507
#>  [253,] 0.962127009 0.932067932 0.983727852
#>  [254,] 0.235764236 0.096903097 0.512487512
#>  [255,] 0.726273726 0.717282717 0.567432567
#>  [256,] 0.040959041 0.514485514 0.288711289
#>  [257,] 0.730269730 0.732267732 0.566433566
#>  [258,] 0.490509491 0.618381618 0.761238761
#>  [259,] 0.045954046 0.305694306 0.108891109
#>  [260,] 0.708291708 0.779220779 0.696303696
#>  [261,] 0.590409590 0.533466533 0.614385614
#>  [262,] 0.931068931 0.936063936 0.995323359
#>  [263,] 0.089910090 0.345654346 0.145854146
#>  [264,] 0.500499500 0.648351648 0.689310689
#>  [265,] 0.095904096 0.247752248 0.134865135
#>  [266,] 0.088911089 0.015984016 0.216783217
#>  [267,] 0.080919081 0.312687313 0.209790210
#>  [268,] 0.565434565 0.434565435 0.429570430
#>  [269,] 0.467532468 0.399600400 0.462537463
#>  [270,] 0.683316683 0.546453546 0.710289710
#>  [271,] 0.241758242 0.763236763 0.431568432
#>  [272,] 0.733266733 0.423576424 0.701298701
#>  [273,] 0.503496503 0.317682318 0.519480519
#>  [274,] 0.286713287 0.523476523 0.685314685
#>  [275,] 0.051948052 0.058941059 0.755244755
#>  [276,] 0.016983017 0.019980020 0.079920080
#>  [277,] 0.364635365 0.475524476 0.764235764
#>  [278,] 0.226773227 0.469530470 0.175824176
#>  [279,] 0.356643357 0.401598402 0.538461538
#>  [280,] 0.687312687 0.793206793 0.661338661
#>  [281,] 0.612387612 0.724275724 0.891108891
#>  [282,] 0.369630370 0.771228771 0.321678322
#>  [283,] 0.004995005 0.318681319 0.089910090
#>  [284,] 0.422577423 0.040959041 0.080919081
#>  [285,] 0.455544456 0.330669331 0.650349650
#>  [286,] 0.370629371 0.290709291 0.261738262
#>  [287,] 0.573426573 0.960752326 0.908091908
#>  [288,] 0.779220779 0.250749251 0.361638362
#>  [289,] 0.473526474 0.154845155 0.398601399
#>  [290,] 0.083916084 0.104895105 0.112887113
#>  [291,] 0.345654346 0.735264735 0.779220779
#>  [292,] 0.461538462 0.376623377 0.302697303
#>  [293,] 0.173826174 0.392607393 0.432567433
#>  [294,] 0.237762238 0.857142857 0.252747253
#>  [295,] 0.633366633 0.659340659 0.340659341
#>  [296,] 0.309690310 0.273726274 0.264735265
#>  [297,] 0.209790210 0.363636364 0.387612388
#>  [298,] 0.663336663 0.668331668 0.621378621
#>  [299,] 0.432567433 0.497502498 0.436563437
#>  [300,] 0.300699301 0.467532468 0.599400599
#>  [301,] 0.621378621 0.628371628 0.274725275
#>  [302,] 0.183816184 0.172827173 0.053946054
#>  [303,] 0.402597403 0.461538462 0.863136863
#>  [304,] 0.205794206 0.257742258 0.716283716
#>  [305,] 0.748251748 0.578421578 0.660339660
#>  [306,] 0.908091908 0.892107892 0.901098901
#>  [307,] 0.597402597 0.308691309 0.428571429
#>  [308,] 0.802197802 0.720279720 0.601398601
#>  [309,] 0.320679321 0.468531469 0.516483516
#>  [310,] 0.509490509 0.777222777 0.380619381
#>  [311,] 0.463536464 0.326673327 0.961411965
#>  [312,] 0.998101300 0.998410682 0.981770731
#>  [313,] 0.315684316 0.561438561 0.647352647
#>  [314,] 0.739260739 0.688311688 0.851148851
#>  [315,] 0.339660340 0.646353646 0.377622378
#>  [316,] 0.688311688 0.435564436 0.783216783
#>  [317,] 0.474525475 0.321678322 0.742257742
#>  [318,] 0.640359640 0.255744256 0.235764236
#>  [319,] 0.330669331 0.033966034 0.259740260
#>  [320,] 0.333666334 0.503496503 0.113886114
#>  [321,] 0.423576424 0.733266733 0.577422577
#>  [322,] 0.580419580 0.828171828 0.655344655
#>  [323,] 0.244755245 0.231768232 0.505494505
#>  [324,] 0.192807193 0.440559441 0.182817183
#>  [325,] 0.568431568 0.449550450 0.416583417
#>  [326,] 0.504495504 0.470529471 0.808191808
#>  [327,] 0.710289710 0.685314685 0.399600400
#>  [328,] 0.287712288 0.490509491 0.254745255
#>  [329,] 0.157842158 0.556443556 0.110889111
#>  [330,] 0.790209790 0.466533467 0.476523477
#>  [331,] 0.812187812 0.641358641 0.703296703
#>  [332,] 0.574425574 0.756243756 0.204795205
#>  [333,] 0.404595405 0.568431568 0.683316683
#>  [334,] 0.181818182 0.531468531 0.325674326
#>  [335,] 0.669330669 0.731268731 0.770229770
#>  [336,] 0.125874126 0.412587413 0.147852148
#>  [337,] 0.843156843 0.840159840 0.691308691
#>  [338,] 0.429570430 0.189810190 0.132867133
#>  [339,] 0.008991009 0.228771229 0.002997003
#>  [340,] 0.811188811 0.887112887 0.904095904
#>  [341,] 0.067932068 0.107892108 0.299700300
#>  [342,] 0.973073805 0.965607897 0.983740428
#>  [343,] 0.124875125 0.093906094 0.255744256
#>  [344,] 0.261738262 0.294705295 0.642357642
#>  [345,] 0.437562438 0.651348651 0.491508492
#>  [346,] 0.243756244 0.277722278 0.314685315
#>  [347,] 0.552447552 0.632367632 0.489510490
#>  [348,] 0.725274725 0.582417582 0.544455544
#>  [349,] 0.048951049 0.179820180 0.510489510
#>  [350,] 0.264735265 0.544455544 0.788211788
#>  [351,] 0.393606394 0.253746254 0.164835165
#>  [352,] 0.470529471 0.826173826 0.659340659
#>  [353,] 0.017982018 0.007992008 0.208791209
#>  [354,] 0.818181818 0.566433566 0.529470529
#>  [355,] 0.863136863 0.816183816 0.839160839
#>  [356,] 0.077922078 0.085914086 0.177822178
#>  [357,] 0.594405594 0.758241758 0.463536464
#>  [358,] 0.041958042 0.036963037 0.220779221
#>  [359,] 0.127872128 0.116883117 0.329670330
#>  [360,] 0.741258741 0.682317682 0.524475524
#>  [361,] 0.595404595 0.174825175 0.135864136
#>  [362,] 0.013986014 0.361638362 0.376623377
#>  [363,] 0.484515485 0.060939061 0.148851149
#>  [364,] 0.894105894 0.918081918 0.976317880
#>  [365,] 0.312687313 0.338661339 0.539460539
#>  [366,] 0.938061938 0.923076923 0.957598546
#>  [367,] 0.799200799 0.671328671 0.551448551
#>  [368,] 0.212787213 0.132867133 0.303696304
#>  [369,] 0.351648352 0.262737263 0.414585415
#>  [370,] 0.062937063 0.056943057 0.358641359
#>  [371,] 0.130869131 0.258741259 0.160839161
#>  [372,] 0.297702298 0.282717283 0.087912088
#>  [373,] 0.551448551 0.222777223 0.262737263
#>  [374,] 0.219780220 0.616383616 0.618381618
#>  [375,] 0.274725275 0.543456543 0.246753247
#>  [376,] 0.335664336 0.159840160 0.256743257
#>  [377,] 0.072927073 0.028971029 0.024975025
#>  [378,] 0.774225774 0.845154845 0.981609154
#>  [379,] 0.635364635 0.691308691 0.766233766
#>  [380,] 0.752247752 0.549450549 0.694305694
#>  [381,] 0.329670330 0.712287712 0.292707293
#>  [382,] 0.084915085 0.062937063 0.058941059
#>  [383,] 0.321678322 0.547452547 0.554445554
#>  [384,] 0.131868132 0.379620380 0.105894106
#>  [385,] 0.866133866 0.702297702 0.426573427
#>  [386,] 0.462537463 0.801198801 0.253746254
#>  [387,] 0.553446553 0.725274725 0.219780220
#>  [388,] 0.869130869 0.610389610 0.807192807
#>  [389,] 0.690309690 0.483516484 0.508491508
#>  [390,] 0.807192807 0.714285714 0.802197802
#>  [391,] 0.002997003 0.626373626 0.050949051
#>  [392,] 0.433566434 0.512487512 0.241758242
#>  [393,] 0.395604396 0.749250749 0.588411588
#>  [394,] 0.225774226 0.097902098 0.248751249
#>  [395,] 0.781218781 0.838161838 0.684315684
#>  [396,] 0.550449550 0.152847153 0.612387612
#>  [397,] 0.135864136 0.156843157 0.085914086
#>  [398,] 0.296703297 0.615384615 0.681318681
#>  [399,] 0.113886114 0.232767233 0.074925075
#>  [400,] 0.108891109 0.112887113 0.412587413
#>  [401,] 0.087912088 0.285714286 0.409590410
#>  [402,] 0.890109890 0.909090909 0.897102897
#>  [403,] 0.524475524 0.500499500 0.473526474
#>  [404,] 0.631368631 0.613386613 0.404595405
#>  [405,] 0.193806194 0.391608392 0.363636364
#>  [406,] 0.604395604 0.522477522 0.486513487
#>  [407,] 0.856143856 0.948051948 0.899100899
#>  [408,] 0.416583417 0.740259740 0.356643357
#>  [409,] 0.834165834 0.775224775 0.532467532
#>  [410,] 0.922077922 0.862137862 0.864135864
#>  [411,] 0.359640360 0.204795205 0.771228771
#>  [412,] 0.280719281 0.484515485 0.488511489
#>  [413,] 0.906093906 0.853146853 0.859140859
#>  [414,] 0.821178821 0.830169830 0.937062937
#>  [415,] 0.184815185 0.410589411 0.211788212
#>  [416,] 0.203796204 0.287712288 0.271728272
#>  [417,] 0.736263736 0.658341658 0.791208791
#>  [418,] 0.587412587 0.661338661 0.725274725
#>  [419,] 0.737262737 0.551448551 0.805194805
#>  [420,] 0.061938062 0.088911089 0.091908092
#>  [421,] 0.508491508 0.148851149 0.071928072
#>  [422,] 0.686313686 0.823176823 0.874125874
#>  [423,] 0.979190523 0.920079920 0.896103896
#>  [424,] 0.791208791 0.772227772 0.895104895
#>  [425,] 0.949050949 0.864135864 0.905094905
#>  [426,] 0.753246753 0.510489510 0.257742258
#>  [427,] 0.643356643 0.545454545 0.581418581
#>  [428,] 0.852147852 0.895104895 0.853146853
#>  [429,] 0.094905095 0.103896104 0.094905095
#>  [430,] 0.962866582 0.987791613 0.982960737
#>  [431,] 0.064935065 0.298701299 0.161838162
#>  [432,] 0.954780149 0.974139520 0.989546946
#>  [433,] 0.682317682 0.548451548 0.909090909
#>  [434,] 0.229770230 0.585414585 0.754245754
#>  [435,] 0.175824176 0.358641359 0.490509491
#>  [436,] 0.758241758 0.499500500 0.852147852
#>  [437,] 0.246753247 0.896103896 0.196803197
#>  [438,] 0.770229770 0.728271728 0.372627373
#>  [439,] 0.208791209 0.105894106 0.503496503
#>  [440,] 0.231768232 0.458541459 0.576423576
#>  [441,] 0.090909091 0.496503497 0.150849151
#>  [442,] 0.518481518 0.635364635 0.203796204
#>  [443,] 0.673326673 0.605394605 0.534465534
#>  [444,] 0.139860140 0.476523477 0.270729271
#>  [445,] 0.282717283 0.390609391 0.513486513
#>  [446,] 0.717282717 0.597402597 0.328671329
#>  [447,] 0.105894106 0.341658342 0.607392607
#>  [448,] 0.055944056 0.046953047 0.468531469
#>  [449,] 0.187812188 0.304695305 0.443556444
#>  [450,] 0.510489510 0.295704296 0.163836164
#>  [451,] 0.564435564 0.776223776 0.878121878
#>  [452,] 0.854145854 0.710289710 0.644355644
#>  [453,] 0.242757243 0.083916084 0.229770230
#>  [454,] 0.859140859 0.627372627 0.632367632
#>  [455,] 0.295704296 0.400599401 0.413586414
#>  [456,] 0.742257742 0.812187812 0.736263736
#>  [457,] 0.861138861 0.789210789 0.781218781
#>  [458,] 0.200799201 0.580419580 0.616383616
#>  [459,] 0.420579421 0.367632368 0.320679321
#>  [460,] 0.659340659 0.219780220 0.757242757
#>  [461,] 0.672327672 0.787212787 0.746253746
#>  [462,] 0.349650350 0.374625375 0.594405594
#>  [463,] 0.750249750 0.794205794 0.932067932
#>  [464,] 0.776223776 0.144855145 0.214785215
#>  [465,] 0.103896104 0.227772228 0.649350649
#>  [466,] 0.847152847 0.861138861 0.840159840
#>  [467,] 0.465534466 0.491508492 0.635364635
#>  [468,] 0.889110889 0.821178821 0.843156843
#>  [469,] 0.468531469 0.438561439 0.173826174
#>  [470,] 0.978017199 0.939060939 0.927072927
#>  [471,] 0.028971029 0.373626374 0.210789211
#>  [472,] 0.935064935 0.926073926 0.921078921
#>  [473,] 0.606393606 0.205794206 0.344655345
#>  [474,] 0.972105433 0.883116883 0.888111888
#>  [475,] 0.234765235 0.200799201 0.181818182
#>  [476,] 0.897102897 0.780219780 0.822177822
#>  [477,] 0.223776224 0.217782218 0.579420579
#>  [478,] 0.806193806 0.902097902 0.824175824
#>  [479,] 0.798201798 0.672327672 0.595404595
#>  [480,] 0.792207792 0.560439560 0.981123106
#>  [481,] 0.171828172 0.624375624 0.550449550
#>  [482,] 0.030969031 0.041958042 0.023976024
#>  [483,] 0.694305694 0.837162837 0.800199800
#>  [484,] 0.450549451 0.100899101 0.569430569
#>  [485,] 0.074925075 0.188811189 0.316683317
#>  [486,] 0.398601399 0.418581419 0.427572428
#>  [487,] 0.728271728 0.494505495 0.494505495
#>  [488,] 0.946053946 0.937062937 0.978863279
#>  [489,] 0.988735242 0.966695238 0.985933253
#>  [490,] 0.037962038 0.340659341 0.251748252
#>  [491,] 0.649350649 0.486513487 0.458541459
#>  [492,] 0.101898102 0.673326673 0.435564436
#>  [493,] 0.825174825 0.869130869 0.920079920
#>  [494,] 0.698301698 0.425574426 0.654345654
#>  [495,] 0.898101898 0.703296703 0.838161838
#>  [496,] 0.759240759 0.716283716 0.407592408
#>  [497,] 0.360639361 0.365634366 0.218781219
#>  [498,] 0.093906094 0.110889111 0.029970030
#>  [499,] 0.031968032 0.111888112 0.049950050
#>  [500,] 0.519480519 0.381618382 0.268731269
#>  [501,] 0.601398601 0.847152847 0.740259740
#>  [502,] 0.928071928 0.903096903 0.947052947
#>  [503,] 0.189810190 0.351648352 0.385614386
#>  [504,] 0.118881119 0.118881119 0.119880120
#>  [505,] 0.833166833 0.891108891 0.673326673
#>  [506,] 0.268731269 0.587412587 0.417582418
#>  [507,] 0.375624376 0.687312687 0.331668332
#>  [508,] 0.983230784 0.979530484 0.984557123
#>  [509,] 0.354645355 0.796203796 0.379620380
#>  [510,] 0.544455544 0.370629371 0.782217782
#>  [511,] 0.865134865 0.721278721 0.953940939
#>  [512,] 0.310689311 0.146853147 0.811188811
#>  [513,] 0.705294705 0.634365634 0.580419580
#>  [514,] 0.702297702 0.751248751 0.667332667
#>  [515,] 0.414585415 0.190809191 0.151848152
#>  [516,] 0.394605395 0.120879121 0.332667333
#>  [517,] 0.163836164 0.109890110 0.012987013
#>  [518,] 0.624375624 0.825174825 0.598401598
#>  [519,] 0.216783217 0.557442557 0.592407592
#>  [520,] 0.760239760 0.655344655 0.848151848
#>  [521,] 0.876123876 0.827172827 0.915084915
#>  [522,] 0.323676324 0.334665335 0.501498501
#>  [523,] 0.197802198 0.537462537 0.088911089
#>  [524,] 0.022977023 0.069930070 0.081918082
#>  [525,] 0.447552448 0.413586414 0.286713287
#>  [526,] 0.411588412 0.552447552 0.561438561
#>  [527,] 0.713286713 0.653346653 0.695304695
#>  [528,] 0.104895105 0.094905095 0.222777223
#>  [529,] 0.400599401 0.171828172 0.787212787
#>  [530,] 0.262737263 0.325674326 0.146853147
#>  [531,] 0.731268731 0.707292707 0.817182817
#>  [532,] 0.388611389 0.457542458 0.401598402
#>  [533,] 0.547452547 0.301698302 0.565434565
#>  [534,] 0.538461538 0.302697303 0.482517483
#>  [535,] 0.376623377 0.353646354 0.349650350
#>  [536,] 0.650349650 0.359640360 0.312687313
#>  [537,] 0.256743257 0.708291708 0.301698302
#>  [538,] 0.772227772 0.908091908 0.906093906
#>  [539,] 0.386613387 0.384615385 0.265734266
#>  [540,] 0.318681319 0.057942058 0.039960040
#>  [541,] 0.199800200 0.875124875 0.106893107
#>  [542,] 0.615384615 0.859140859 0.466533467
#>  [543,] 0.328671329 0.013986014 0.709290709
#>  [544,] 0.042957043 0.037962038 0.157842158
#>  [545,] 0.489510490 0.516483516 0.651348651
#>  [546,] 0.114885115 0.079920080 0.287712288
#>  [547,] 0.881118881 0.729270729 0.735264735
#>  [548,] 0.116883117 0.072927073 0.521478521
#>  [549,] 0.639360639 0.654345654 0.326673327
#>  [550,] 0.666333666 0.378621379 0.297702298
#>  [551,] 0.679320679 0.168831169 0.315684316
#>  [552,] 0.602397602 0.764235764 0.819180819
#>  [553,] 0.557442557 0.459540460 0.597402597
#>  [554,] 0.704295704 0.873126873 0.775224775
#>  [555,] 0.257742258 0.135864136 0.236763237
#>  [556,] 0.641358641 0.485514486 0.702297702
#>  [557,] 0.695304695 0.741258741 0.589410589
#>  [558,] 0.165834166 0.039960040 0.460539461
#>  [559,] 0.141858142 0.119880120 0.101898102
#>  [560,] 0.803196803 0.689310689 0.865134865
#>  [561,] 0.153846154 0.038961039 0.360639361
#>  [562,] 0.855144855 0.848151848 0.845154845
#>  [563,] 0.689310689 0.593406593 0.793206793
#>  [564,] 0.417582418 0.532467532 0.290709291
#>  [565,] 0.629370629 0.541458541 0.319680320
#>  [566,] 0.495504496 0.506493506 0.727272727
#>  [567,] 0.102897103 0.076923077 0.027972028
#>  [568,] 0.848151848 0.906093906 0.849150849
#>  [569,] 0.250749251 0.265734266 0.415584416
#>  [570,] 0.955413799 0.986898267 0.977083566
#>  [571,] 0.001998002 0.117882118 0.197802198
#>  [572,] 0.583416583 0.441558442 0.045954046
#>  [573,] 0.331668332 0.649350649 0.195804196
#>  [574,] 0.886113886 0.674325674 0.548451548
#>  [575,] 0.611388611 0.690309690 0.774225774
#>  [576,] 0.910089910 0.950874747 0.974797729
#>  [577,] 0.085914086 0.066933067 0.189810190
#>  [578,] 0.814185814 0.558441558 0.975237776
#>  [579,] 0.341658342 0.134865135 0.433566434
#>  [580,] 0.486513487 0.187812188 0.223776224
#>  [581,] 0.428571429 0.481518482 0.228771229
#>  [582,] 0.715284715 0.800199800 0.662337662
#>  [583,] 0.277722278 0.162837163 0.381618382
#>  [584,] 0.661338661 0.048951049 0.530469530
#>  [585,] 0.893106893 0.765234765 0.844155844
#>  [586,] 0.071928072 0.043956044 0.142857143
#>  [587,] 0.128871129 0.289710290 0.304695305
#>  [588,] 0.491508492 0.489510490 0.269730270
#>  [589,] 0.260739261 0.075924076 0.250749251
#>  [590,] 0.419580420 0.524475524 0.613386613
#>  [591,] 0.438561439 0.129870130 0.239760240
#>  [592,] 0.651348651 0.393606394 0.583416583
#>  [593,] 0.202797203 0.307692308 0.318681319
#>  [594,] 0.290709291 0.437562438 0.128871129
#>  [595,] 0.387612388 0.035964036 0.279720280
#>  [596,] 0.596403596 0.170829171 0.365634366
#>  [597,] 0.298701299 0.337662338 0.557442557
#>  [598,] 0.427572428 0.314685315 0.459540460
#>  [599,] 0.332667333 0.832167832 0.317682318
#>  [600,] 0.783216783 0.270729271 0.242757243
#>  [601,] 0.925074925 0.584415584 0.604395604
#>  [602,] 0.677322677 0.620379620 0.333666334
#>  [603,] 0.563436563 0.052947053 0.070929071
#>  [604,] 0.954560437 0.982614912 0.913086913
#>  [605,] 0.849150849 0.125874126 0.273726274
#>  [606,] 0.932067932 0.917082917 0.975245288
#>  [607,] 0.248751249 0.426573427 0.114885115
#>  [608,] 0.729270729 0.350649351 0.373626374
#>  [609,] 0.269730270 0.195804196 0.795204795
#>  [610,] 0.477522478 0.430569431 0.375624376
#>  [611,] 0.258741259 0.021978022 0.193806194
#>  [612,] 0.307692308 0.460539461 0.062937063
#>  [613,] 0.361638362 0.739260739 0.797202797
#>  [614,] 0.785214785 0.747252747 0.861138861
#>  [615,] 0.969762910 0.967790354 0.955593728
#>  [616,] 0.358641359 0.319680320 0.394605395
#>  [617,] 0.194805195 0.344655345 0.374625375
#>  [618,] 0.021978022 0.081918082 0.498501499
#>  [619,] 0.561438561 0.244755245 0.617382617
#>  [620,] 0.971072578 0.921078921 0.964708336
#>  [621,] 0.782217782 0.699300699 0.773226773
#>  [622,] 0.326673327 0.067932068 0.858141858
#>  [623,] 0.506493506 0.588411588 0.669330669
#>  [624,] 0.747252747 0.930069930 0.664335664
#>  [625,] 0.943056943 0.884115884 0.949050949
#>  [626,] 0.844155844 0.852147852 0.643356643
#>  [627,] 0.917082917 0.953633912 0.940059940
#>  [628,] 0.801198801 0.904095904 0.641358641
#>  [629,] 0.732267732 0.782217782 0.798201798
#>  [630,] 0.107892108 0.089910090 0.179820180
#>  [631,] 0.972269941 0.938061938 0.882117882
#>  [632,] 0.882117882 0.638361638 0.728271728
#>  [633,] 0.841158841 0.527472527 0.640359640
#>  [634,] 0.350649351 0.239760240 0.389610390
#>  [635,] 0.896103896 0.670329670 0.857142857
#>  [636,] 0.100899101 0.003996004 0.031968032
#>  [637,] 0.098901099 0.251748252 0.747252747
#>  [638,] 0.302697303 0.173826174 0.184815185
#>  [639,] 0.236763237 0.018981019 0.237762238
#>  [640,] 0.285714286 0.444555445 0.483516484
#>  [641,] 0.011988012 0.224775225 0.011988012
#>  [642,] 0.133866134 0.214785215 0.719280719
#>  [643,] 0.914085914 0.940059940 0.930069930
#>  [644,] 0.888111888 0.824175824 0.917082917
#>  [645,] 0.434565435 0.530469530 0.386613387
#>  [646,] 0.129870130 0.050949051 0.188811189
#>  [647,] 0.270729271 0.300699301 0.308691309
#>  [648,] 0.281718282 0.681318681 0.322677323
#>  [649,] 0.204795205 0.074925075 0.073926074
#>  [650,] 0.065934066 0.000999001 0.115884116
#>  [651,] 0.627372627 0.225774226 0.421578422
#>  [652,] 0.177822178 0.004995005 0.083916084
#>  [653,] 0.511488511 0.130869131 0.296703297
#>  [654,] 0.363636364 0.091908092 0.055944056
#>  [655,] 0.492507493 0.608391608 0.411588412
#>  [656,] 0.110889111 0.357642358 0.137862138
#>  [657,] 0.956966044 0.977394599 0.973336174
#>  [658,] 0.381618382 0.456543457 0.258741259
#>  [659,] 0.586413586 0.343656344 0.337662338
#>  [660,] 0.578421578 0.804195804 0.480519481
#>  [661,] 0.451548452 0.095904096 0.032967033
#>  [662,] 0.496503497 0.553446553 0.506493506
#>  [663,] 0.082917083 0.010989011 0.111888112
#>  [664,] 0.373626374 0.480519481 0.606393606
#>  [665,] 0.178821179 0.380619381 0.362637363
#>  [666,] 0.485514486 0.698301698 0.442557443
#>  [667,] 0.920079920 0.867132867 0.860139860
#>  [668,] 0.827172827 0.818181818 0.780219780
#>  [669,] 0.862137862 0.734265734 0.831168831
#>  [670,] 0.795204795 0.508491508 0.504495504
#>  [671,] 0.222777223 0.086913087 0.309690310
#>  [672,] 0.744255744 0.762237762 0.829170829
#>  [673,] 0.480519481 0.809190809 0.281718282
#>  [674,] 0.505494505 0.433566434 0.130869131
#>  [675,] 0.441558442 0.242757243 0.430569431
#>  [676,] 0.991685082 0.992412144 0.953087532
#>  [677,] 0.810189810 0.834165834 0.814185814
#>  [678,] 0.680319680 0.660339660 0.633366633
#>  [679,] 0.488511489 0.431568432 0.191808192
#>  [680,] 0.070929071 0.858141858 0.143856144
#>  [681,] 0.096903097 0.077922078 0.366633367
#>  [682,] 0.901098901 0.889110889 0.872127872
#>  [683,] 0.058941059 0.360639361 0.769230769
#>  [684,] 0.342657343 0.127872128 0.477522478
#>  [685,] 0.027972028 0.788211788 0.449550450
#>  [686,] 0.227772228 0.498501499 0.067932068
#>  [687,] 0.980591332 0.976997802 0.976102843
#>  [688,] 0.775224775 0.811188811 0.752247752
#>  [689,] 0.600399600 0.518481518 0.556443556
#>  [690,] 0.179820180 0.176823177 0.283716284
#>  [691,] 0.787212787 0.808191808 0.886113886
#>  [692,] 0.266733267 0.191808192 0.198801199
#>  [693,] 0.660339660 0.750249750 0.879120879
#>  [694,] 0.068931069 0.087912088 0.033966034
#>  [695,] 0.449550450 0.583416583 0.638361638
#>  [696,] 0.915084915 0.850149850 0.893106893
#>  [697,] 0.185814186 0.237762238 0.180819181
#>  [698,] 0.076923077 0.082917083 0.100899101
#>  [699,] 0.699300699 0.913086913 0.523476523
#>  [700,] 0.056943057 0.071928072 0.202797203
#>  [701,] 0.644355644 0.683316683 0.722277722
#>  [702,] 0.039960040 0.106893107 0.059940060
#>  [703,] 0.941058941 0.946053946 0.919080919
#>  [704,] 0.406593407 0.736263736 0.584415584
#>  [705,] 0.158841159 0.166833167 0.093906094
#>  [706,] 0.069930070 0.234765235 0.034965035
#>  [707,] 0.536463536 0.606393606 0.509490509
#>  [708,] 0.667332667 0.623376623 0.622377622
#>  [709,] 0.567432567 0.630369630 0.815184815
#>  [710,] 0.134865135 0.102897103 0.021978022
#>  [711,] 0.066933067 0.073926074 0.226773227
#>  [712,] 0.913086913 0.813186813 0.877122877
#>  [713,] 0.291708292 0.259740260 0.353646354
#>  [714,] 0.756243756 0.719280719 0.587412587
#>  [715,] 0.766233766 0.696303696 0.697302697
#>  [716,] 0.613386613 0.070929071 0.037962038
#>  [717,] 0.150849151 0.164835165 0.169830170
#>  [718,] 0.399600400 0.680319680 0.810189810
#>  [719,] 0.006993007 0.034965035 0.041958042
#>  [720,] 0.494505495 0.346653347 0.610389610
#>  [721,] 0.255744256 0.436563437 0.823176823
#>  [722,] 0.367632368 0.266733267 0.400599401
#>  [723,] 0.562437562 0.760239760 0.767232767
#>  [724,] 0.145854146 0.264735265 0.125874126
#>  [725,] 0.877122877 0.713286713 0.826173826
#>  [726,] 0.823176823 0.935064935 0.883116883
#>  [727,] 0.425574426 0.428571429 0.418581419
#>  [728,] 0.271728272 0.555444555 0.423576424
#>  [729,] 0.111888112 0.115884116 0.596403596
#>  [730,] 0.458541459 0.141858142 0.485514486
#>  [731,] 0.664335664 0.629370629 0.559440559
#>  [732,] 0.584415584 0.839160839 0.624375624
#>  [733,] 0.769230769 0.126873127 0.167832168
#>  [734,] 0.036963037 0.030969031 0.046953047
#>  [735,] 0.206793207 0.513486513 0.154845155
#>  [736,] 0.456543457 0.773226773 0.015984016
#>  [737,] 0.623376623 0.831168831 0.677322677
#>  [738,] 0.146853147 0.099900100 0.129870130
#>  [739,] 0.582417582 0.482517483 0.371628372
#>  [740,] 0.186813187 0.012987013 0.010989011
#>  [741,] 0.618381618 0.894105894 0.444555445
#>  [742,] 0.148851149 0.140859141 0.064935065
#>  [743,] 0.487512488 0.877122877 0.369630370
#>  [744,] 0.534465534 0.621378621 0.751248751
#>  [745,] 0.768231768 0.754245754 0.585414585
#>  [746,] 0.156843157 0.409590410 0.155844156
#>  [747,] 0.170829171 0.193806194 0.278721279
#>  [748,] 0.871128871 0.579420579 0.721278721
#>  [749,] 0.325674326 0.450549451 0.555444555
#>  [750,] 0.944055944 0.860139860 0.934065934
#>  [751,] 0.716283716 0.770229770 0.887112887
#>  [752,] 0.344655345 0.211788212 0.357642358
#>  [753,] 0.346653347 0.600399600 0.586413586
#>  [754,] 0.610389610 0.898101898 0.525474525
#>  [755,] 0.445554446 0.447552448 0.668331668
#>  [756,] 0.951970406 0.900099900 0.942057942
#>  [757,] 0.136863137 0.178821179 0.178821179
#>  [758,] 0.719280719 0.759240759 0.637362637
#>  [759,] 0.308691309 0.519480519 0.284715285
#>  [760,] 0.985782445 0.933066933 0.856143856
#>  [761,] 0.379620380 0.184815185 0.060939061
#>  [762,] 0.220779221 0.245754246 0.133866134
#>  [763,] 0.273726274 0.354645355 0.457542458
#>  [764,] 0.012987013 0.216783217 0.048951049
#>  [765,] 0.123876124 0.275724276 0.116883117
#>  [766,] 0.978878484 0.972734280 0.945054945
#>  [767,] 0.872127872 0.472527473 0.801198801
#>  [768,] 0.043956044 0.617382617 0.107892108
#>  [769,] 0.625374625 0.511488511 0.653346653
#>  [770,] 0.892107892 0.911088911 0.866133866
#>  [771,] 0.391608392 0.347652348 0.495504496
#>  [772,] 0.502497502 0.505494505 0.095904096
#>  [773,] 0.956439502 0.748251748 0.679320679
#>  [774,] 0.245754246 0.029970030 0.396603397
#>  [775,] 0.086913087 0.022977023 0.077922078
#>  [776,] 0.867132867 0.640359640 0.821178821
#>  [777,] 0.767232767 0.866133866 0.562437562
#>  [778,] 0.009990010 0.009990010 0.149850150
#>  [779,] 0.453546454 0.169830170 0.648351648
#>  [780,] 0.410589411 0.293706294 0.126873127
#>  [781,] 0.853146853 0.819180819 0.748251748
#>  [782,] 0.656343656 0.411588412 0.084915085
#>  [783,] 0.214785215 0.292707293 0.768231768
#>  [784,] 0.581418581 0.502497502 0.484515485
#>  [785,] 0.347652348 0.594405594 0.850149850
#>  [786,] 0.091908092 0.385614386 0.025974026
#>  [787,] 0.991885633 0.984300722 0.993365228
#>  [788,] 0.267732268 0.336663337 0.726273726
#>  [789,] 0.472527473 0.329670330 0.020979021
#>  [790,] 0.988007814 0.988243308 0.984164948
#>  [791,] 0.284715285 0.550449550 0.075924076
#>  [792,] 0.832167832 0.943056943 0.806193806
#>  [793,] 0.478521479 0.064935065 0.156843157
#>  [794,] 0.794205794 0.679320679 0.674325674
#>  [795,] 0.986507588 0.945054945 0.924075924
#>  [796,] 0.343656344 0.619380619 0.187812188
#>  [797,] 0.078921079 0.078921079 0.487512488
#>  [798,] 0.836163836 0.802197802 0.818181818
#>  [799,] 0.947052947 0.929070929 0.922077922
#>  [800,] 0.959634128 0.942057942 0.916083916
#>  [801,] 0.712287712 0.535464535 0.828171828
#>  [802,] 0.167832168 0.045954046 0.102897103
#>  [803,] 0.945054945 0.846153846 0.935064935
#>  [804,] 0.263736264 0.243756244 0.186813187
#>  [805,] 0.948051948 0.987767981 0.948051948
#>  [806,] 0.918081918 0.656343656 0.605394605
#>  [807,] 0.658341658 0.248751249 0.069930070
#>  [808,] 0.817182817 0.160839161 0.493506494
#>  [809,] 0.025974026 0.279720280 0.068931069
#>  [810,] 0.887112887 0.880119880 0.750249750
#>  [811,] 0.396603397 0.352647353 0.205794206
#>  [812,] 0.372627373 0.316683317 0.040959041
#>  [813,] 0.577422577 0.529470529 0.335664336
#>  [814,] 0.555444555 0.320679321 0.923076923
#>  [815,] 0.987568785 0.983771225 0.994768123
#>  [816,] 0.675324675 0.464535465 0.472527473
#>  [817,] 0.316683317 0.238761239 0.382617383
#>  [818,] 0.507492507 0.288711289 0.118881119
#>  [819,] 0.452547453 0.822177822 0.446553447
#>  [820,] 0.956593960 0.961651647 0.979142386
#>  [821,] 0.049950050 0.051948052 0.231768232
#>  [822,] 0.764235764 0.675324675 0.928071928
#>  [823,] 0.046953047 0.150849151 0.066933067
#>  [824,] 0.126873127 0.210789211 0.042957043
#>  [825,] 0.217782218 0.371628372 0.497502498
#>  [826,] 0.709290709 0.576423576 0.542457542
#>  [827,] 0.880119880 0.905094905 0.833166833
#>  [828,] 0.409590410 0.303696304 0.765234765
#>  [829,] 0.539460539 0.572427572 0.537462537
#>  [830,] 0.476523477 0.194805195 0.131868132
#>  [831,] 0.762237762 0.934065934 0.855144855
#>  [832,] 0.722277722 0.591408591 0.572427572
#>  [833,] 0.835164835 0.797202797 0.687312687
#>  [834,] 0.276723277 0.322677323 0.465534466
#>  [835,] 0.652347652 0.042957043 0.044955045
#>  [836,] 0.121878122 0.208791209 0.127872128
#>  [837,] 0.761238761 0.814185814 0.867132867
#>  [838,] 0.032967033 0.092907093 0.247752248
#>  [839,] 0.353646354 0.333666334 0.631368631
#>  [840,] 0.306693307 0.652347652 0.439560440
#>  [841,] 0.047952048 0.055944056 0.294705295
#>  [842,] 0.034965035 0.309690310 0.141858142
#>  [843,] 0.826173826 0.799200799 0.825174825
#>  [844,] 0.598401598 0.296703297 0.359640360
#>  [845,] 0.647352647 0.705294705 0.391608392
#>  [846,] 0.542457542 0.868131868 0.799200799
#>  [847,] 0.809190809 0.559440559 0.672327672
#>  [848,] 0.436563437 0.375624376 0.813186813
#>  [849,] 0.149850150 0.383616384 0.097902098
#>  [850,] 0.558441558 0.473526474 0.682317682
#>  [851,] 0.378621379 0.753246753 0.348651349
#>  [852,] 0.000999001 0.001998002 0.000999001
#>  [853,] 0.884115884 0.835164835 0.842157842
#>  [854,] 0.626373626 0.377622378 0.206793207
#>  [855,] 0.191808192 0.014985015 0.370629371
#>  [856,] 0.942057942 0.919080919 0.862137862
#>  [857,] 0.499500500 0.192807193 0.918081918
#>  [858,] 0.079920080 0.065934066 0.645354645
#>  [859,] 0.196803197 0.478521479 0.527472527
#>  [860,] 0.247752248 0.203796204 0.243756244
#>  [861,] 0.481518482 0.614385614 0.536463536
#>  [862,] 0.377622378 0.406593407 0.364635365
#>  [863,] 0.054945055 0.175824176 0.200799201
#>  [864,] 0.531468531 0.849150849 0.438561439
#>  [865,] 0.926073926 0.586413586 0.282717283
#>  [866,] 0.874125874 0.790209790 0.820179820
#>  [867,] 0.977848410 0.989039648 0.966587083
#>  [868,] 0.516483516 0.445554446 0.615384615
#>  [869,] 0.934065934 0.952024386 0.911088911
#>  [870,] 0.857142857 0.960043610 0.957560494
#>  [871,] 0.830169830 0.863136863 0.786213786
#>  [872,] 0.912087912 0.603396603 0.528471528
#>  [873,] 0.668331668 0.722277722 0.834165834
#>  [874,] 0.632367632 0.589410589 0.892107892
#>  [875,] 0.548451548 0.564435564 0.712287712
#>  [876,] 0.540459540 0.424575425 0.657342657
#>  [877,] 0.784215784 0.791208791 0.350649351
#>  [878,] 0.313686314 0.398601399 0.213786214
#>  [879,] 0.755244755 0.841158841 0.957407336
#>  [880,] 0.162837163 0.369630370 0.676323676
#>  [881,] 0.075924076 0.113886114 0.410589411
#>  [882,] 0.607392607 0.181818182 0.454545455
#>  [883,] 0.691308691 0.183816184 0.139860140
#>  [884,] 0.721278721 0.842157842 0.708291708
#>  [885,] 0.599400599 0.280719281 0.467532468
#>  [886,] 0.743256743 0.666333666 0.578421578
#>  [887,] 0.221778222 0.415584416 0.076923077
#>  [888,] 0.515484515 0.442557443 0.212787213
#>  [889,] 0.674325674 0.746253746 0.723276723
#>  [890,] 0.198801199 0.885114885 0.238761239
#>  [891,] 0.380619381 0.520479520 0.437562438
#>  [892,] 0.322677323 0.197802198 0.511488511
#>  [893,] 0.435564436 0.604395604 0.514485514
#>  [894,] 0.201798202 0.286713287 0.692307692
#>  [895,] 0.685314685 0.677322677 0.424575425
#>  [896,] 0.430569431 0.769230769 0.352647353
#>  [897,] 0.754245754 0.897102897 0.718281718
#>  [898,] 0.421578422 0.439560440 0.383616384
#>  [899,] 0.366633367 0.744255744 0.670329670
#>  [900,] 0.815184815 0.622377622 0.571428571
#>  [901,] 0.408591409 0.504495504 0.630369630
#>  [902,] 0.099900100 0.158841159 0.298701299
#>  [903,] 0.249750250 0.443556444 0.552447552
#>  [904,] 0.700299700 0.432567433 0.704295704
#>  [905,] 0.401598402 0.121878122 0.870129870
#>  [906,] 0.777222777 0.323676324 0.759240759
#>  [907,] 0.431568432 0.417582418 0.082917083
#>  [908,] 0.003996004 0.024975025 0.293706294
#>  [909,] 0.294705295 0.562437562 0.447552448
#>  [910,] 0.334665335 0.020979021 0.103896104
#>  [911,] 0.956873411 0.912087912 0.816183816
#>  [912,] 0.007992008 0.044955045 0.005994006
#>  [913,] 0.720279720 0.709290709 0.547452547
#>  [914,] 0.662337662 0.870129870 0.688311688
#>  [915,] 0.052947053 0.068931069 0.030969031
#>  [916,] 0.646353646 0.267732268 0.152847153
#>  [917,] 0.444555445 0.167832168 0.140859141
#>  [918,] 0.397602398 0.025974026 0.122877123
#>  [919,] 0.589410589 0.701298701 0.912087912
#>  [920,] 0.304695305 0.416583417 0.334665335
#>  [921,] 0.383616384 0.922077922 0.291708292
#>  [922,] 0.605394605 0.342657343 0.378621379
#>  [923,] 0.081918082 0.784215784 0.285714286
#>  [924,] 0.608391608 0.643356643 0.880119880
#>  [925,] 0.735264735 0.755244755 0.744255744
#>  [926,] 0.724275724 0.944055944 0.785214785
#>  [927,] 0.407592408 0.241758242 0.402597403
#>  [928,] 0.405594406 0.487512488 0.392607393
#>  [929,] 0.327672328 0.642357642 0.758241758
#>  [930,] 0.773226773 0.843156843 0.734265734
#>  [931,] 0.459540460 0.124875125 0.345654346
#>  [932,] 0.718281718 0.781218781 0.280719281
#>  [933,] 0.278721279 0.002997003 0.109890110
#>  [934,] 0.992723282 0.989787062 0.984298476
#>  [935,] 0.120879121 0.047952048 0.003996004
#>  [936,] 0.038961039 0.090909091 0.099900100
#>  [937,] 0.210789211 0.479520480 0.232767233
#>  [938,] 0.044955045 0.198801199 0.568431568
#>  [939,] 0.190809191 0.023976024 0.008991009
#>  [940,] 0.479520480 0.706293706 0.603396603
#>  [941,] 0.097902098 0.054945055 0.001998002
#>  [942,] 0.909090909 0.910089910 0.944055944
#>  [943,] 0.374625375 0.686313686 0.441558442
#>  [944,] 0.891108891 0.989804156 0.902097902
#>  [945,] 0.858141858 0.684315684 0.625374625
#>  [946,] 0.275724276 0.220779221 0.121878122
#>  [947,] 0.169830170 0.356643357 0.600399600
#>  [948,] 0.413586414 0.063936064 0.307692308
#>  [949,] 0.824175824 0.973421452 0.894105894
#>  [950,] 0.921078921 0.611388611 0.522477522
#>  [951,] 0.161838162 0.715284715 0.313686314
#>  [952,] 0.575424575 0.414585415 0.451548452
#>  [953,] 0.138861139 0.366633367 0.165834166
#>  [954,] 0.060939061 0.032967033 0.267732268
#>  [955,] 0.933066933 0.878121878 0.881118881
#>  [956,] 0.117882118 0.711288711 0.275724276
#>  [957,] 0.174825175 0.263736264 0.227772228
#>  [958,] 0.443556444 0.299700300 0.230769231
#>  [959,] 0.546453546 0.212787213 0.026973027
#>  [960,] 0.616383616 0.528471528 0.706293706
#>  [961,] 0.352647353 0.059940060 0.018981019
#>  [962,] 0.213786214 0.281718282 0.153846154
#>  [963,] 0.907092907 0.446553447 0.756243756
#>  [964,] 0.701298701 0.607392607 0.609390609
#>  [965,] 0.819180819 0.260739261 0.741258741
#>  [966,] 0.588411588 0.525474525 0.540459540
#>  [967,] 0.903096903 0.907092907 0.875124875
#>  [968,] 0.355644356 0.008991009 0.009990010
#>  [969,] 0.964014242 0.969499389 0.977552093
#>  [970,] 0.879120879 0.726273726 0.898101898
#>  [971,] 0.638361638 0.563436563 0.560439560
#>  [972,] 0.035964036 0.420579421 0.036963037
#>  [973,] 0.362637363 0.462537463 0.591408591
#>  [974,] 0.023976024 0.240759241 0.016983017
#>  [975,] 0.868131868 0.810189810 0.698301698
#>  [976,] 0.878121878 0.792207792 0.876123876
#>  [977,] 0.293706294 0.757242757 0.620379620
#>  [978,] 0.645354645 0.539460539 0.469530470
#>  [979,] 0.092907093 0.226773227 0.185814186
#>  [980,] 0.642357642 0.221778222 0.323676324
#>  [981,] 0.471528472 0.888111888 0.639360639
#>  [982,] 0.559440559 0.571428571 0.022977023
#>  [983,] 0.464535465 0.785214785 0.425574426
#>  [984,] 0.927072927 0.778221778 0.753246753
#>  [985,] 0.440559441 0.581418581 0.384615385
#>  [986,] 0.288711289 0.602397602 0.245754246
#>  [987,] 0.050949051 0.202797203 0.233766234
#>  [988,] 0.415584416 0.153846154 0.545454545
#>  [989,] 0.532467532 0.851148851 0.306693307
#>  [990,] 0.152847153 0.209790210 0.737262737
#>  [991,] 0.523476523 0.471528472 0.570429570
#>  [992,] 0.147852148 0.133866134 0.061938062
#>  [993,] 0.696303696 0.186813187 0.199800200
#>  [994,] 0.160839161 0.131868132 0.295704296
#>  [995,] 0.911088911 0.879120879 0.724275724
#>  [996,] 0.528471528 0.477522478 0.574425574
#>  [997,] 0.681318681 0.636363636 0.564435564
#>  [998,] 0.154845155 0.517482517 0.098901099
#>  [999,] 0.180819181 0.254745255 0.170829171
#> [1000,] 0.207792208 0.408591409 0.351648352