Coefficient of tail correlation and tail dependence

For data with unit Pareto margins, the coefficient of tail dependence \(\eta\) is defined via $$\Pr(\min(X) > x) = L(x)x^{-1/\eta},$$ where \(L(x)\) is a slowly varying function; \(0 < \eta\). Ignoring the latter, several estimators of \(\eta\) can be defined. In unit Pareto margins, \(\eta\) is a shape parameter that can be estimated by fitting a generalized Pareto distribution above a high threshold. In exponential margins, \(\eta\) is a scale parameter and the maximum likelihood estimator of the latter is the Hill estimator. Both methods are based on peaks-over-threshold and the user can choose between pointwise confidence confint obtained through a likelihood ratio test statistic ("lrt") or the Wald statistic ("wald").

Usage

taildep(
  data,
  u = NULL,
  nq = 40,
  qlim = c(0.8, 0.99),
  depmeas = c("eta", "chi"),
  method = list(eta = c("emp", "betacop", "gpd", "hill"), chi = c("emp", "betacop",
    "emplik")),
  confint = c("wald", "lrt"),
  level = 0.95,
  trunc = TRUE,
  ties.method = "random",
  plot = TRUE,
  ...
)

Arguments

data

an \(n\) by \(d\) matrix of multivariate observations

u

vector of percentiles between 0 and 1 at which to evaluate the plot

nq

number of quantiles at which to form a grid; only used if u = NULL.

qlim

limits for the sequence u

depmeas

dependence measure, either of "eta" or "chi"

method

named list giving the estimation method for eta and chi. Default to "emp" for both.

confint

string indicating the type of confidence interval for \(\eta\), one of "wald" or "lrt"

level

the confidence level required (default to 0.95).

trunc

logical indicating whether the estimates and confidence intervals should be truncated in \([0,1]\)

ties.method

string indicating the type of method for rank; see rank for a list of options. Default to "random"

plot

logical; should graphs be plotted?

...

additional arguments passed to plot; current support for main, xlab, ylab, add and further pch, lty, type, col for points; additional arguments for confidence intervals are handled via cipch, cilty, citype, cicol.

Value

a named list with elements

u: a K vector of percentile levels

eta: a K by 3 matrix with point estimates, lower and upper confidence intervals

chi: a K by 3 matrix with point estimates, lower and upper confidence intervals

Details

The most common approach for estimation is the empirical survival copula, by evaluating the proportion of sample minima with uniform margins that exceed a given \(x\). An alternative estimator uses a smoothed estimator of the survival copula using Bernstein polynomial, resulting in the so-called betacop estimator. Approximate pointwise confidence confint for the latter are obtained by assuming the proportion of points is binomial.

The coefficient of tail correlation \(\chi\) is $$\chi = \lim_{u \to 1} \frac{\Pr(F_1(X_1)>u, \ldots, F_D(X_D)>u)}{1-u}.$$ Asymptotically independent vectors have \(\chi = 0\). The estimator uses an estimator of the survival copula

See also

Examples

if (FALSE) {
set.seed(765)
# Max-stable model
dat <- rmev(n = 1000, d = 4, param = 0.7, model = "log")
taildep(dat, confint = 'wald')
}